### 2026-04-02 #10363 @editorialbot
[PRE REVIEW]: PyNumDiff 2.0: Practical Numerical Differentiation for Noisy Data
**Submitting author:** @pavelkomarov (<a href="http://orcid.org/0009-0007-7482-2807">Pavel Komarov</a>)
**Repository:** https://github.com/florisvb/PyNumDiff
**Branch with paper.md** (empty if default branch): joss2.0
**Version:** v0.2.3
