### 2026-04-02 #10363 @danielskatz
@pavelkomarov - thanks for this submission. It appears to me to be in scope for JOSS review, so I'll next look for an editor.

For what it's worth, you probably should remove the first two lines from your LICENSE.txt file so that it is seen as the MIT license by automated tools.
