### 2026-04-02 #10363 @danielskatz
@pavelkomarov - I also see that there's a DOI suggested above that you could add to your bib file, if it's correct.
