### 2026-04-03 825a54b
removed lines of MIT license that are confusing the editorbot, and added suggested DOI, which is indeed correct
