+++ paper/paper.bib
@@ -111,7 +111,8 @@
   year    = {2013},
   volume  = {14},
   pages   = {2513--2558},
-  url     = {https://jmlr.org/papers/volume14/aravkin13a/aravkin13a.pdf}
+  url     = {https://jmlr.org/papers/volume14/aravkin13a/aravkin13a.pdf},
+  doi     = {10.1007/978-3-642-38398-4_8}
 }
 
 @software{derivative_pkg,
