### 2026-04-03 65fef9f
paper: trim two phrases to prevent single-line page 5 overflow
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
