### 2026-04-03 florisvb/PyNumDiff#201 @pavelkomarov
After investigation, we've decided not to implement this. `finufft` requires OpenMP and a full C++ build toolchain, making it a heavy optional dependency that is difficult to install on macOS (Apple's Clang doesn't ship OpenMP; users need to manually install `libomp` via Homebrew and pass several CMake flags by hand). This friction is incompatible with PyNumDiff's otherwise pure Python/SciPy install experience.

The approach is sound — the three-step pipeline (Type 1 NUFFT → filter → Type 2 NUFFT) with average-`dt` padding for edge handling works correctly — but the dependency cost isn't worth it. `spectraldiff` will remain uniform-grid only. Users needing irregular-grid spectral differentiation should look at `finufft` directly or use one of the variable-step methods already in the package (`splinediff`, `rbfdiff`, `rtsdiff`, `robustdiff`).

— Claude Sonnet 4.6
