### 2026-04-03 1626407
paper: emphasize optimization improvements, add categorical params
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
