+++ paper/paper.md
@@ -37,14 +37,14 @@ PyNumDiff is an open-source Python package consolidating a broad suite of numeri
 
 Derivative estimation from noisy measurements arises throughout experimental science, data-driven modeling, system identification, and control. The field offers a diverse ecosystem of specialized algorithms, each with different strengths in outlier robustness, computational cost, irregular sampling, or missing observations, but no consolidated home. Without one, practitioners are left assembling solutions piecemeal from disparate packages [@vanBreugel2022].
 
-PyNumDiff addresses this gap. Its unified interface lets users compare methods on the same data, exploit specialized capabilities, and select hyperparameters without ground-truth derivatives. Derivative estimation trades data fidelity against smoothness; PyNumDiff frames this as a multi-objective optimization, finding hyperparameter settings that minimize a weighted combination of the two. A single scalar weight `tvgamma` steers between objectives; when ground-truth derivatives are available, they replace it as the optimization target. This framework spans all methods, enabling principled comparison and selection [@vanBreugel2020numerical]. A natural and growing application is SINDy [@brunton2016discovering], which discovers governing equations by regressing measured derivatives, making the quality of those estimates a direct determinant of model accuracy.
+PyNumDiff addresses this gap. Its unified interface lets users compare methods on the same data, exploit specialized capabilities, and select hyperparameters without ground-truth derivatives. Derivative estimation trades data fidelity against smoothness; PyNumDiff frames this as a multi-objective optimization, finding hyperparameter settings that minimize a weighted combination of the two. A single scalar weight `tvgamma` steers between objectives in the absence of ground-truth, or, when a ground-truth derivative is available, it can serve it as the optimization target. This framework spans all methods, enabling principled comparison and selection [@vanBreugel2020numerical]. A natural and growing application is SINDy [@brunton2016discovering], which discovers governing equations by regressing measured derivatives, making the quality of those estimates a direct determinant of model accuracy.
 
 
 # State of the Field
 
 Relevant Python tools exist, but none covers PyNumDiff's breadth. `numpy.gradient` and `scipy.signal.savgol_filter` [@virtanen2020scipy] handle only a sliver of the method space; `findiff` offers high-order finite difference stencils suited to clean simulation data, not noisy measurements. Historically, practitioners have had to stitch together PyKalman, PyDMD, and standalone TVR scripts [@chartrand2011numerical] with no shared API or principled way to compare results. The `derivative` package [@derivative_pkg] overlaps substantially but lacks multidimensional support, NaN handling, and hyperparameter optimization. No existing package spans PyNumDiff's seven method families with a consistent interface.
 
-The original PyNumDiff publication [@vanBreugel2022] established the core method set and optimization framework. This version substantially revises it: the taxonomy was reorganized and corrected; the interface unified around explicit keyword arguments; test coverage expanded; and four capabilities were added throughout: multidimensional data, irregular sample spacing, missing observations, and circular domains.
+The original PyNumDiff publication [@vanBreugel2022] established the core method set and optimization framework. This version substantially revises it: the taxonomy was reorganized and corrected; the interface unified around explicit keyword arguments; the hyperparameter optimizer was substantially improved; test coverage expanded; and four capabilities were added throughout: multidimensional data, irregular sample spacing, missing observations, and circular domains.
 
 
 # Software Design
@@ -87,7 +87,7 @@ Table: Specialized capabilities by method.
 
 **Hyperparameter optimization.** `pynumdiff.optimize` minimizes the weighted combination described above [@vanBreugel2020numerical]. The smoothness weight `tvgamma` can be initialized from the signal's estimated cutoff frequency $f_c$ via
 $$\texttt{tvgamma} = \exp(-1.6\ln f_c - 0.71\ln \Delta t - 5.1).$$
[…]
