### 2026-04-17 0086c1d
Add waveletdiff: wavelet-based denoising differentiator
- Implements Donoho & Johnstone (1994) universal threshold estimator
- Vectorised thresholding over multi-column inputs
- Supports variable step size via np.gradient
- Supports arbitrary axis via ascontiguousarray + moveaxis
- Adds pywt dependency
