+++ pynumdiff/basis_fit.py
@@ -2,6 +2,7 @@
 from warnings import warn
 import numpy as np
 from scipy import sparse
+import pywt
 
 from pynumdiff.utils import utility
 
@@ -137,3 +138,125 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01):
     alpha = sparse.linalg.spsolve(rbf_regularized, x) # solve sparse system targeting the noisy data, O(N sigma^2)
 
     return rbf @ alpha, drbfdt @ alpha # find samples of reconstructions using the smooth bases
+
+
+def waveletdiff(x, dt_or_t, wavelet='db4', level=None, threshold=1.0, axis=0, mode='periodization'):
+    """Smooth and differentiate noisy data via discrete wavelet denoising.
+
+    Decomposes x into wavelet detail and approximation coefficients, soft-thresholds
+    the detail coefficients to remove noise using the Donoho & Johnstone (1994)
+    universal threshold estimator, reconstructs a smoothed signal, then
+    differentiates with finite differences via np.gradient.
+
+    :param np.array x: data to differentiate
+    :param float or array dt_or_t: scalar dt or array of sample times. If an
+        array is provided it is passed directly to np.gradient, giving correct
+        results for non-uniformly sampled data.
+    :param str wavelet: PyWavelets wavelet name, e.g. 'db4', 'sym4', 'coif2'.
+        'db4' is a solid general-purpose default. Biorthogonal wavelets such as
+        'bior2.2' or 'bior4.4' are symmetric and designed for smooth reconstruction
+        but may need a lower threshold value.
+    :param int level: decomposition depth. None (default) resolves to
+        min(pywt.dwt_max_level(N, wavelet), 5) to avoid over-decomposing short
+        signals. Increase for heavily oversampled data.
+    :param float threshold: soft-thresholding scale factor in [0, inf).
+        Multiplies the universal threshold sigma * sqrt(2 * log(N)).
+        threshold=1.0 is the classical Donoho & Johnstone universal threshold
+        and is the recommended starting point. Values < 1.0 give less smoothing;
+        values > 1.0 give more aggressive smoothing. This parameter maps onto
+        tvgamma in the pynumdiff.optimize framework.
+    :param int axis: axis along which to differentiate (default 0).
+    :param str mode: PyWavelets signal extension mode passed to wavedec/waverec.
+        'periodization' (default) keeps coefficient arrays exactly length N and
+        is the most numerically stable choice for differentiation. 'reflect' is
+        a good alternative for clearly non-periodic signals.
+        See pywt.Modes.modes for all options.
+    :return: - **x_hat** (np.array) -- estimated (smoothed) x
+             - **dxdt_hat** (np.array) -- estimated derivative of x
+    """
+    N = x.shape[axis]
+
+    # Axis normalisation — bring target axis to front.
+    # Skip moveaxis when axis is already 0 to avoid an unnecessary allocation.
+    # When we do move, call ascontiguousarray immediately so the subsequent
+    # reshape is guaranteed zero-copy.
+    if axis == 0:
+        x_work = x if x.flags['C_CONTIGUOUS'] else np.ascontiguousarray(x)
+    else:
+        x_work = np.ascontiguousarray(np.moveaxis(x, axis, 0))
+
+    shape = x_work.shape
+    x_flat = x_work.reshape(N, -1)  # (N, M) contiguous, no hidden copy
+    M = x_flat.shape[1]
+
+    if np.isscalar(dt_or_t):
+        grad_arg = dt_or_t
[…]
