### 2026-04-20 #10363 @prashjha
Hi @pavelkomarov, sorry for the delay...semester end is crazy. Anyways, could you please suggest a few potential reviewers for your submission? You can get some names from the reviewer database [here](https://reviewers.joss.theoj.org/).
