### 2026-04-23 #10363 @pavelkomarov
I feel like this is a lazy answer, but I'm not sure what to do aside from use the search and pick a few names. I have no one in particular in mind, but whoever it is should know Python and care about numerical methods and differentiation, with expertise in dynamical systems and signal processing being plusses.
