+++ pynumdiff/basis_fit.py
@@ -136,18 +136,21 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-def waveletdiff(x, dt_or_t, wavelet='db4', level=None, threshold=1.0, axis=0, mode='periodization'):
+def waveletdiff(x, dt, wavelet='db4', level=None, threshold=1.0, axis=0, mode='periodization'):
     """Smooth and differentiate noisy data via discrete wavelet denoising.
 
     Decomposes x into wavelet detail and approximation coefficients, soft-thresholds
     the detail coefficients to remove noise using the Donoho & Johnstone (1994)
     universal threshold estimator, reconstructs a smoothed signal, then
-    differentiates with finite differences via np.gradient.
+    differentiates analytically by applying derivative reconstruction filters to
+    the denoised wavelet coefficients.
 
-    :param np.array x: data to differentiate
-    :param float or array dt_or_t: scalar dt or array of sample times. If an
-        array is provided it is passed directly to np.gradient, giving correct
-        results for non-uniformly sampled data.
+    Because the DWT requires uniform spacing, this method only accepts a scalar
+    time step dt (not a vector of sample times). For non-uniformly sampled data,
+    use :func:`rbfdiff` or :func:`splinediff` instead.
+
+    :param np.array x: data to differentiate. May be multidimensional; see :code:`axis`.
+    :param float dt: uniform time step between samples.
     :param str wavelet: PyWavelets wavelet name, e.g. 'db4', 'sym4', 'coif2'.
         'db4' is a solid general-purpose default. Biorthogonal wavelets such as
         'bior2.2' or 'bior4.4' are symmetric and designed for smooth reconstruction
@@ -170,39 +173,32 @@ def waveletdiff(x, dt_or_t, wavelet='db4', level=None, threshold=1.0, axis=0, mo
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
+    if not np.isscalar(dt):
+        raise ValueError(
+            "`dt` must be a scalar. The DWT requires uniformly sampled data. "
+            "For variable step sizes, use rbfdiff or splinediff instead."
+        )
+
     N = x.shape[axis]
 
-    # Axis normalisation — bring target axis to front.
-    # Skip moveaxis when axis is already 0 to avoid an unnecessary allocation.
-    # When we do move, call ascontiguousarray immediately so the subsequent
-    # reshape is guaranteed zero-copy.
-    if axis == 0:
-        x_work = x if x.flags['C_CONTIGUOUS'] else np.ascontiguousarray(x)
-    else:
-        x_work = np.ascontiguousarray(np.moveaxis(x, axis, 0))
-
+    # Bring axis of differentiation to front so each column of x_flat is one
+    # signal to differentiate. moveaxis returns a view with updated strides,
+    # so ascontiguousarray ensures the subsequent reshape is zero-copy.
+    x_work = np.ascontiguousarray(np.moveaxis(x, axis, 0))
     shape = x_work.shape
-    x_flat = x_work.reshape(N, -1)  # (N, M) contiguous, no hidden copy
+    x_flat = x_work.reshape(N, -1)  # (N, M)
     M = x_flat.shape[1]
 
-    if np.isscalar(dt_or_t):
-        grad_arg = dt_or_t
-    else:
-        if len(dt_or_t) != N:
-            raise ValueError(
[…]
