+++ pynumdiff/tests/test_diff_methods.py
@@ -332,19 +332,15 @@ multidim_methods_and_params = [
     (kerneldiff, {'kernel': 'gaussian', 'window_size': 5}),
     (butterdiff, {'filter_order': 3, 'cutoff_freq': 1 - 1e-6}),
     (finitediff, {}),
-<<<<<<< HEAD
-    (savgoldiff, {'degree': 3, 'window_size': 11, 'smoothing_win': 3}),
-    (waveletdiff, {'wavelet': 'db4', 'threshold': 1.0}),
-=======
     (polydiff, {'degree': 2, 'window_size': 5}),
     (savgoldiff, {'degree': 3, 'window_size': 11, 'smoothing_win': 3}),
+    (waveletdiff, {'wavelet': 'db4', 'threshold': 1.0}),
     (rtsdiff, {'order':2, 'log_qr_ratio':7, 'forwardbackward':True}),
     (spectraldiff, {'high_freq_cutoff': 0.25, 'pad_to_zero_dxdt': False}),
     (rbfdiff, {'sigma': 0.5, 'lmbd': 1e-6}),
     (splinediff, {'degree': 9, 's': 1e-6}),
     (robustdiff, {'order':2, 'log_q':7, 'log_r':2}),
     (tvrdiff, {'order': 3, 'gamma': 1e-4})
->>>>>>> b38199f982cb4036065f599b3fe00f6076671a6a
 ]
 
 # Similar to the error_bounds table, index by method first. But then we test against only one 2D function,
@@ -355,10 +351,7 @@ multidim_error_bounds = {
     kerneldiff: [(2, 1), (3, 2)],
     butterdiff: [(0, -1), (1, -1)],
     finitediff: [(0, -1), (1, -1)],
-<<<<<<< HEAD
-    savgoldiff: [(0, -1), (1, 1)],
     waveletdiff: [(1, 0), (2, 1)],
-=======
     polydiff: [(1, -1), (1, 0)],
     savgoldiff: [(0, -1), (1, 1)],
     rtsdiff: [(1, -1), (1, 0)],
@@ -367,7 +360,6 @@ multidim_error_bounds = {
     splinediff: [(0, -1), (1, 0)],
     robustdiff: [(-2, -3), (0, -1)],
     tvrdiff: [(0, -1), (1, 0)]
->>>>>>> b38199f982cb4036065f599b3fe00f6076671a6a
 }
 
 @mark.parametrize("multidim_method_and_params", multidim_methods_and_params)
@@ -420,9 +412,6 @@ def test_multidimensionality(multidim_method_and_params, request):
         ax2.plot_wireframe(T1, T2, computed_d2)
         ax3.plot_wireframe(T1, T2, computed_laplacian, label='computed')
         legend = ax3.legend(bbox_to_anchor=(0.7, 0.8)); legend.legend_handles[0].set_facecolor(pyplot.cm.viridis(0.6))
-<<<<<<< HEAD
-        fig.suptitle(f'{diff_method.__name__}', fontsize=16)
-=======
         fig.suptitle(f'{diff_method.__name__}', fontsize=16)
 
 
@@ -475,4 +464,3 @@ def test_missing_data(diff_method_and_params):
     
     assert np.all(np.isfinite(x_hat))
     assert np.all(np.isfinite(dxdt_hat))
->>>>>>> b38199f982cb4036065f599b3fe00f6076671a6a
