### 2026-05-01 68f9255
Fix waveletdiff: use np.gradient on denoised signal for derivative
