### 2026-05-08 5bcbe29
Fix waveletdiff: use spectral differentiation of denoised signal
