### 2026-06-12 dbf2f3f
Replace FFT-based wavelet differentiation with wavelet basis derivative operator
