+++ pynumdiff/basis_fit.py
@@ -1,7 +1,9 @@
 """Methods based on fitting basis functions to data"""
+from functools import lru_cache
 from warnings import warn
 import numpy as np
 from scipy import sparse
+from scipy.interpolate import CubicSpline
 import pywt
 
 from pynumdiff.utils import utility
@@ -136,14 +138,94 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-def waveletdiff(x, dt, wavelet='db4', level=None, threshold=1.0, axis=0, mode='periodization'):
-    """Smooth and differentiate noisy data via discrete wavelet denoising.
+@lru_cache(maxsize=32)
+def _wavelet_derivative_synthesis_matrix(N, dt, wavelet, level, mode):
+    """Build sparse samples of d/dt of the inverse-DWT synthesis basis.
 
-    Decomposes x into wavelet detail and approximation coefficients, soft-thresholds
-    the detail coefficients to remove noise using the Donoho & Johnstone (1994)
-    universal threshold estimator, reconstructs a smoothed signal, then
-    differentiates analytically by applying derivative reconstruction filters to
-    the denoised wavelet coefficients.
+    For a fixed wavelet/level/mode/length, wavedec/waverec define a linear
+    synthesis map
+
+        x(t_n) = sum_k c_k phi_k(t_n).
+
+    This routine samples phi'_k(t_n) once, stores those samples sparsely, and
+    lets waveletdiff compute
+
+        x'(t_n) = sum_k c_k phi'_k(t_n)
+
+    without differentiating the reconstructed signal.  The derivative samples
+    are obtained from a local cubic interpolant of each compactly supported
+    synthesis basis vector; this is bookkeeping on the basis functions, not a
+    finite-difference derivative of the data.
+    """
+    zero = np.zeros(N)
+    template = pywt.wavedec(zero, wavelet, level=level, mode=mode)
+    coeff_lengths = tuple(len(c) for c in template)
+    coeff_offsets = np.cumsum((0,) + coeff_lengths[:-1])
+    n_coeffs = sum(coeff_lengths)
+    t = np.arange(N, dtype=float) * dt
+
+    rows, cols, vals = [], [], []
+    eps = 1e-12
+
+    for band, (offset, length) in enumerate(zip(coeff_offsets, coeff_lengths)):
+        for local_idx in range(length):
+            coeffs = [np.zeros_like(c, dtype=float) for c in template]
+            coeffs[band][local_idx] = 1.0
+            basis = pywt.waverec(coeffs, wavelet, mode=mode)[:N]
+
+            # Basis functions are compactly supported, but boundary extension can
+            # split support across the two ends.  Differentiating only the active
+            # samples keeps the matrix sparse and avoids global sinusoidal bases.
+            active = np.flatnonzero(np.abs(basis) > eps)
+            if active.size == 0:
+                continue
+
+            # Include one-sample padding around active support so the cubic has
+            # enough context near the edges of the support.  If support wraps or
+            # covers most of the signal, fall back to all samples.
+            support = np.zeros(N, dtype=bool)
+            support[active] = True
+            support[np.maximum(active - 1, 0)] = True
+            support[np.minimum(active + 1, N - 1)] = True
+            idx = np.flatnonzero(support)
+            if idx.size < 4 or (idx[-1] - idx[0] + 1) > 2 * idx.size:
+                idx = np.arange(N)
+
[…]
