### 2026-06-12 baf9946
Compute wavelet derivative analytically via connection coefficients
Replace the cubic-spline-of-sampled-basis derivative with the exact
analytic derivative of the wavelet basis. The denoised reconstruction is
the wavelet interpolant x(t) = sum_n a_n phi(t/dt - n), so its derivative
is x' = Phi' Phi^-1 x_hat, where Phi and Phi' are circulant samples of the
scaling function phi and its derivative phi' at integers. Those samples are
the eigenvalue-1 and eigenvalue-1/2 eigenvectors of the wavelet refinement
relation (connection coefficients), normalized to reproduce constants and
differentiate ram
[…]
