+++ pynumdiff/basis_fit.py
@@ -3,7 +3,6 @@ from functools import lru_cache
 from warnings import warn
 import numpy as np
 from scipy import sparse
-from scipy.interpolate import CubicSpline
 import pywt
 
 from pynumdiff.utils import utility
@@ -139,93 +138,70 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
 
 
 @lru_cache(maxsize=32)
-def _wavelet_derivative_synthesis_matrix(N, dt, wavelet, level, mode):
-    """Build sparse samples of d/dt of the inverse-DWT synthesis basis.
+def _wavelet_derivative_operator(N, dt, wavelet):
+    """Build the sparse operators that turn denoised samples into a derivative.
 
-    For a fixed wavelet/level/mode/length, wavedec/waverec define a linear
-    synthesis map
+    PyWavelets treats the input samples as the finest-level scaling coefficients,
+    so the denoised reconstruction x_hat represents the continuous interpolant
+    x(t) = sum_n a_n phi(t/dt - n), where phi is the wavelet's scaling function.
+    Sampling x and its analytic derivative on the grid t_m = m*dt gives two
+    convolutions against phi and phi' evaluated at *integers*:
 
-        x(t_n) = sum_k c_k phi_k(t_n).
+        x_hat[m] = sum_n a_n phi(m - n)            ->  x_hat = Phi @ a
+        x'(t_m)  = (1/dt) sum_n a_n phi'(m - n)    ->  x'    = Phi_prime @ a
 
-    This routine samples phi'_k(t_n) once, stores those samples sparsely, and
-    lets waveletdiff compute
+    so x' = Phi_prime @ Phi^-1 @ x_hat. This is the exact derivative of the
+    wavelet interpolant, with no spline or finite-difference approximation.
 
-        x'(t_n) = sum_k c_k phi'_k(t_n)
+    phi and phi' at the integers are the eigenvectors of the wavelet's refinement
+    (dilation) relation: differentiating phi(t) = sqrt2 * sum_k h_k phi(2t - k)
+    and evaluating at integers shows phi sampled at integers is the eigenvalue-1
+    eigenvector and phi' the eigenvalue-1/2 eigenvector of T[p,q] = sqrt2 * h_{2p-q}.
+    Normalizations come from reproduction of constants (sum phi(p) = 1) and of
+    linears (sum p*phi'(p) = -1, so the operator differentiates a ramp exactly).
 
-    without differentiating the reconstructed signal.  The derivative samples
-    are obtained from a local cubic interpolant of each compactly supported
-    synthesis basis vector; this is bookkeeping on the basis functions, not a
-    finite-difference derivative of the data.
+    :return: - **Phi** (csc_matrix) -- circulant samples of phi, to be inverted
+             - **Phi_prime** (csr_matrix) -- circulant samples of phi'/dt
     """
-    zero = np.zeros(N)
-    template = pywt.wavedec(zero, wavelet, level=level, mode=mode)
-    coeff_lengths = tuple(len(c) for c in template)
-    coeff_offsets = np.cumsum((0,) + coeff_lengths[:-1])
-    n_coeffs = sum(coeff_lengths)
-    t = np.arange(N, dtype=float) * dt
+    h = np.array(pywt.Wavelet(wavelet).rec_lo)   # reconstruction low-pass = refinement filter h_k
+    h = h / h.sum() * np.sqrt(2)                  # enforce sum(h) = sqrt2, i.e. integral of phi is 1
+    L = len(h)                                    # phi is supported on [0, L-1]; sample those integers
+    p = np.arange(L)
 
-    rows, cols, vals = [], [], []
-    eps = 1e-12
[…]
