### 2026-06-12 142ebe4
Loosen near-machine-precision wavelet derivative bound for CI portability
The clean constant-derivative Linf error lands at ~1.2e-14 on Linux vs
<=1e-14 on macOS; give that one bound a decade of headroom.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
