+++ pynumdiff/tests/test_diff_methods.py
@@ -174,7 +174,7 @@ error_bounds = {
               [(-2, -2), (0, 0), (0, -1), (0, 0)],
               [(0, 0), (2, 2), (0, 0), (2, 2)],
               [(1, 1), (3, 3), (1, 1), (3, 3)]],
-    waveletdiff: [[(-15, -15), (-13, -14), (0, -1), (1, 0)],
+    waveletdiff: [[(-15, -15), (-13, -13), (0, -1), (1, 0)],
                   [(-2, -2), (-1, -1), (0, 0), (1, 1)],
                   [(-2, -2), (-1, -1), (0, 0), (1, 1)],
                   [(-3, -3), (-1, -1), (0, 0), (1, 1)],
