### 2026-06-12 c8003a4
Drop lru_cache on wavelet derivative operator
The operator is built once per waveletdiff call and applied to all columns
at once via the vectorized spsolve, so within-call reuse is already free.
The cache only helped across separate same-length calls, a marginal win that
didn't justify retaining up to 32 sparse-matrix pairs in memory.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
