### 2026-06-12 8e8d46e
Inline the wavelet derivative operator into waveletdiff
Fold the private _wavelet_derivative_operator into waveletdiff as a
standalone block, mirroring how rbfdiff builds its basis and derivative
matrices in one loop. No other caller needs it, and the connection-coefficient
math now lives in waveletdiff's own docstring.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
