+++ pynumdiff/basis_fit.py
@@ -136,73 +136,27 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-def _wavelet_derivative_operator(N, dt, wavelet):
-    """Build the sparse operators that turn denoised samples into a derivative.
-
-    Depends only on the grid (N, dt) and the wavelet, not on the data, and is
-    built once per waveletdiff call then applied to every column at once.
-
-    PyWavelets treats the input samples as the finest-level scaling coefficients,
-    so the denoised reconstruction x_hat represents the continuous interpolant
-    x(t) = sum_n a_n phi(t/dt - n), where phi is the wavelet's scaling function.
-    Sampling x and its analytic derivative on the grid t_m = m*dt gives two
-    convolutions against phi and phi' evaluated at *integers*:
-
-        x_hat[m] = sum_n a_n phi(m - n)            ->  x_hat = Phi @ a
-        x'(t_m)  = (1/dt) sum_n a_n phi'(m - n)    ->  x'    = Phi_prime @ a
-
-    so x' = Phi_prime @ Phi^-1 @ x_hat. This is the exact derivative of the
-    wavelet interpolant, with no spline or finite-difference approximation.
-
-    phi and phi' at the integers are the eigenvectors of the wavelet's refinement
-    (dilation) relation: differentiating phi(t) = sqrt2 * sum_k h_k phi(2t - k)
-    and evaluating at integers shows phi sampled at integers is the eigenvalue-1
-    eigenvector and phi' the eigenvalue-1/2 eigenvector of T[p,q] = sqrt2 * h_{2p-q}.
-    Normalizations come from reproduction of constants (sum phi(p) = 1) and of
-    linears (sum p*phi'(p) = -1, so the operator differentiates a ramp exactly).
-
-    :return: - **Phi** (csc_matrix) -- circulant samples of phi, to be inverted
-             - **Phi_prime** (csr_matrix) -- circulant samples of phi'/dt
-    """
-    h = np.array(pywt.Wavelet(wavelet).rec_lo)   # reconstruction low-pass = refinement filter h_k
-    h = h / h.sum() * np.sqrt(2)                  # enforce sum(h) = sqrt2, i.e. integral of phi is 1
-    L = len(h)                                    # phi is supported on [0, L-1]; sample those integers
-    p = np.arange(L)
-
-    # Transition matrix T[p,q] = sqrt2 * h_{2p-q}; entries outside the filter are 0.
-    cols = 2 * p[:, None] - p[None, :]
-    T = np.where((cols >= 0) & (cols < L), np.sqrt(2) * h[np.clip(cols, 0, L - 1)], 0.0)
-
-    evals, evecs = np.linalg.eig(T)
-    phi = np.real(evecs[:, np.argmin(np.abs(evals - 1.0))])    # phi(p):  eigenvalue 1
-    dphi = np.real(evecs[:, np.argmin(np.abs(evals - 0.5))])   # phi'(p): eigenvalue 1/2
-    phi = phi / phi.sum()                                      # sum_p phi(p) = 1
-    dphi = dphi / np.dot(p, dphi) * (-1.0)                     # sum_p p*phi'(p) = -1
-
-    # Both kernels become circulant matrices under periodic boundaries; a common
-    # shift of both cancels in Phi_prime @ Phi^-1, so the offset choice is cosmetic.
-    def circulant(kernel):
-        rows, cols, vals = [], [], []
-        m = np.arange(N)
-        for offset, val in zip(p, kernel):
[…]
