### 2026-06-12 d02f186
paper: add Maria Protogerou and the wavelet differentiation method
Add Maria Protogerou (UW Applied Mathematics) as an author, list wavelet
fitting under method family (4), add the waveletdiff capability-table row,
and credit PyWavelets among the delegated dependencies.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
