+++ paper/paper.bib
@@ -154,3 +154,14 @@
   pages   = {2104},
   doi     = {10.21105/joss.02104}
 }
+
+@article{lee2019pywavelets,
+  author  = {Gregory R. Lee and Ralf Gommers and Filip Wasilewski and Kai Wohlfahrt and Aaron O'Leary},
+  title   = {{PyWavelets}: A {Python} package for wavelet analysis},
+  journal = {Journal of Open Source Software},
+  year    = {2019},
+  volume  = {4},
+  number  = {36},
+  pages   = {1237},
+  doi     = {10.21105/joss.01237}
+}
