### 2026-06-12 f61b008
paper: tighten Software Design prose to keep Table 1 on one page
Trim ~18 words of filler from the Package design and Software architecture
paragraphs so Table 1 no longer breaks across pages 2-3.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
