### 2026-06-12 f25be8b
ci: bump draft-pdf workflow to Node 24 actions
actions/checkout@v4 -> v6 and actions/upload-artifact@v4 -> v7 (Node 24
runtime; inputs unchanged).

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
