### 2026-06-12 8cff0e0
paper: note Claude Opus 4.8 in the AI usage disclosure
Recent code revisions and paper edits used Claude Opus 4.8 alongside the
earlier Sonnet 4.6 drafting.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
