### 2026-06-12 #10363 @pavelkomarov
minor updates to get Table 1 all on page 2.
