### 2026-07-28 d4ba7d7
tests: floor near-exact error bounds so machine-precision round-off can't fail CI
The bounds table asserted L2/Linf errors against hard powers of ten calibrated
right at each method's achieved round-off. Analytically exact methods land at
sqrt(N)*eps .. ~100*eps (1e-15..1e-14 here), and numpy 2.5 / scipy 1.18 /
Python 3.13 shifted those least-significant bits just over the thresholds,
failing CI with no code change.

Introduce flr = -12 and record every at/below-floor bound as `flr`, so exact
methods need only hit 1e-12, not their exact (non-reproducible) round-off value.
The --bounds calibrat
[…]
