+++ pynumdiff/tests/test_diff_methods.py
@@ -66,103 +66,104 @@ diff_methods_and_params = [
     (lineardiff, {'order':3, 'gamma':0.01, 'window_size':11, 'solver':'CLARABEL'}), (lineardiff, [3, 0.01, 11], {'solver':'CLARABEL'})
     ]
 
-# All the testing methodology follows the exact same pattern; the only thing that changes is the
-# closeness to the right answer various methods achieve with the given parameterizations and random seed.
-# So index a big ol' table by method, then by test function number, and finally by the truth-result pair
-# being compared. The tuples are order of magnitude of (L2,Linf) distances for pairs
-# (x,x_hat), (dxdt,dxdt_hat), (x,x_hat_noisy), (dxdt,dxdt_hat_noisy).
+# All the testing methodology follows the exact same pattern; the only thing that changes is the closeness to the
+# right answer various methods achieve with the given parameterizations and random seed. So index a big ol' table
+# by method, then by test function number, and finally by the truth-result pair being compared. The tuples are order
+# of magnitude of (L2,Linf) distances for pairs (x,x_hat), (dxdt,dxdt_hat), (x,x_hat_noisy), (dxdt,dxdt_hat_noisy).
+flr = -12 # machine-precision floor: methods this accurate are exact up to round-off, with least-significant bits
+          # drifting across BLAS/numpy/Python builds, so we don't assert tighter error.
 error_bounds = {
-    meandiff: [[(-25, -25), (-25, -25), (0, -1), (0, 0)],
+    meandiff: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
                [(0, 0), (1, 1), (0, 0), (1, 1)],
                [(0, 0), (1, 1), (0, 0), (1, 1)],
                [(0, 0), (1, 1), (0, 0), (1, 1)],
                [(1, 1), (2, 2), (1, 1), (2, 2)],
                [(1, 1), (3, 3), (1, 1), (3, 3)]],
-    mediandiff: [[(-25, -25), (-25, -25), (-1, -1), (0, 0)],
+    mediandiff: [[(flr, flr), (flr, flr), (-1, -1), (0, 0)],
                  [(0, 0), (1, 1), (0, 0), (1, 1)],
                  [(0, 0), (1, 1), (0, 0), (1, 1)],
                  [(-1, -1), (0, 0), (0, 0), (1, 1)],
                  [(0, 0), (2, 2), (0, 0), (2, 2)],
                  [(1, 1), (3, 3), (1, 1), (3, 3)]],
-    gaussiandiff: [[(-14, -15), (-14, -14), (0, -1), (1, 0)],
+    gaussiandiff: [[(flr, flr), (flr, flr), (0, -1), (1, 0)],
                    [(-1, -1), (1, 0), (0, 0), (1, 1)],
                    [(0, 0), (1, 1), (0, 0), (1, 1)],
                    [(0, -1), (1, 1), (0, 0), (1, 1)],
                    [(1, 1), (2, 2), (1, 1), (2, 2)],
                    [(1, 1), (3, 3), (1, 1), (3, 3)]],
-    friedrichsdiff: [[(-25, -25), (-25, -25), (0, -1), (1, 0)],
+    friedrichsdiff: [[(flr, flr), (flr, flr), (0, -1), (1, 0)],
                      [(-1, -1), (1, 0), (0, 0), (1, 1)],
                      [(0, 0), (1, 1), (0, 0), (1, 1)],
[…]
