### 2026-07-28 74b80ac
tests: simplify parallel/serial param comparison to a single order-safe allclose
Compare the two param dicts by indexing both with the same key sequence and
passing the value lists to one np.allclose, rather than an all() over per-key
allclose calls. Order-safe without depending on the dicts' internal insertion
order lining up.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
