### 2026-08-18 #10363 @prashjha
Hi @mscroggs, @atzberg, @DamynChipman, @jonaspleyer, @ygahounzo, @MarcelKoch, @sidxz, @meliao, I am looking for two reviewers for the above JOSS submission, and I am reaching out as I think this may be in your domain. Would you be interested in reviewing this?

You can find the PDF at this link [pdf link](https://github.com/openjournals/joss-papers/blob/joss.10363/joss.10363/10.21105.joss.10363.pdf), and the corresponding software is at this [link](https://github.com/florisvb/PyNumDiff).

The [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) cover almost everything about the review process; I will be happy to provide more information.

Please let me know if you are interested. Thank you!
