### 2026-08-18 #10363 @alexzerg
Hi @prashjha , I'm still interested if you need the second reviewer.
My strongest fit is the software side: installation, CI, tests, packaging, docs and operational reproducibility. I can commit to an initial checklist pass within two weeks.

Thanks,
Aleksei
