### 2026-08-18 #11172 @prashjha
Hi @alexzerg, @sidxz, please read the first couple of comments in this thread and create your review checklist. You can read the reviewer guidelines [here](https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html). Also, you can browse the [closed "REVIEW" issues](https://github.com/openjournals/joss-reviews/issues?q=label%3Areview+is%3Aclosed) on the "joss-reviews" repository to get some ideas on how to complete the reviews.

For a software or a Joss article draft, you can directly create a GitHub issue on the repository (see the link at the top). When you create an issue, include the text https://github.com/openjournals/joss-reviews/issues/10630 to refer to this review.
