### 2026-08-18 #11172 @prashjha
Hi @alexzerg, @sidxz, the authors have indicated that this is a second JOSS publication on the software pynumdiff and I asked for some clarification [here](https://github.com/openjournals/joss-reviews/issues/10363#issuecomment-4960041879) and the authors response can be found [here](https://github.com/openjournals/joss-reviews/issues/10363#issuecomment-5092565362).

I will suggest accounting for this conversation in your reviews.
