### 2026-08-19 florisvb/PyNumDiff#204 @alexzerg
Clarify what 2.0 adds beyond the 2022 JOSS paper
The new work looks substantial but it's still hard to separate it from what was covered by the first (2022 JOSS) paper.

- Could you add a short 1.x vs 2.0 comparison?
- Btw the paper says 2.0 while the reviewed release is 0.2.3. Is 2.0 the product generation rather than the package version?

Review reference: openjournals/joss-reviews#11172
