### 2026-08-19 florisvb/PyNumDiff#205 @alexzerg
Base install misses tqdm
I tried a local install from readme

```
pip install .
python -c "import pynumdiff”
python -c "from pynumdiff.optimize import optimize”
ModuleNotFoundError: No module named ‘tqdm’
```
tqdm is only in the advanced extra but optimize.py imports it directly.

Review reference: openjournals/joss-reviews#11172
