### 2026-08-19 florisvb/PyNumDiff#207 @alexzerg
optimize crashes when all values in search_space_updates are fixed
I was checking test_search_space_updates_applied and tried the same test with all butterdiff params fixed:
```
def test_all_params_fixed():
    optimize(
        butterdiff,
        x,
        dt,
        dxdt_truth=dxdt_truth,
        parallel=False,
        search_space_updates={
            "filter_order": 2,
            "cutoff_freq": 0.1,
            "num_iterations": 1,
        },
    )
```
It fails on current joss2.0 python 3.13:
`ValueError: not enough values to unpack (expected 2, got 0)`
I think there are no values left for scipy to optimize here. The comment in optimize.py says singleton params are just passed through but when all of them are singletons it still calls scipy.optimize.minimize() with an empty point.
Most probably this case should just evaluate the fixed params once and return the score. What do you think?

Review reference: https://github.com/openjournals/joss-reviews/issues/11172
