### 2026-08-20 florisvb/PyNumDiff#204 @pavelkomarov
The 0.2 vs 2.0 is a bit informal, where the actual releases lead with a 0 to denote it's use-at-your-own-risk research software that isn't supported by the full weight of a company with paid developers, but it's the second major release of this thing. I'm open to changing the title so there is less possibility for confusion there. It did cross my mind.

For new work, I cover some of that info here https://github.com/openjournals/joss-reviews/issues/10363#issuecomment-5092565362. I could see a dedicated digest of that being useful as a section in the JOSS paper. Happy to make it happen.

What are your preferences on these two action items?
