### 2026-08-20 #11172 @pavelkomarov
Nice, thanks for all the movement on this. I notice the License checkbox isn't ticked, but there certainly is one. https://github.com/florisvb/PyNumDiff/blob/master/LICENSE.txt
