### 2026-08-20 florisvb/PyNumDiff#205 @pavelkomarov
Good find. I'll amend the .toml right away. I think my rationale was that only advanced users would be optimizing hyperparameters, but it is kind of core functionality.
