### 2026-08-20 cbfd657
moved tqdm to base dependencies since the optimize module needs it
