### 2026-08-20 florisvb/PyNumDiff#205 @pavelkomarov
https://github.com/florisvb/PyNumDiff/commit/cbfd6578bb609783994fa98c69784845965ed765
