### 2026-08-20 florisvb/PyNumDiff#206 @pavelkomarov
Ah, again good eye. The issue here was that certain warnings might get triggered in subprocesses as you're trying to optimize something, and it can be overwhelming at the level of the optimizer to see the flood of them come in. For example, certain choices of hyperparameters might cause a failure internal to a method that makes it return a bunch of NaNs, but the optimizer just throws that result out and looks for a hyperparam combo that works and returns a good result. Silencing was a bit of a challenge, so I eventually got pretty aggressive, but I didn't realize that leaves warnings turned off for future, unrelated function calls. We want the scope to be local to `optimize()`, certainly.
