### 2026-08-20 e8a2f92
Scope optimize()'s UserWarning silencing to the call itself
filterwarnings() was called bare inside optimize(), so the "ignore"
entry it inserts into warnings.filters stayed there for the life of the
process and muted unrelated UserWarnings long after the optimizer
returned. Wrap the search in catch_warnings(action="ignore",
category=UserWarning), which installs the same filter on entry and
restores the caller's filters on exit, including if the call raises.

The Pool initializer keeps its filterwarnings() call, since that runs in
worker processes where there is nothing to restore.

catch_warn
[…]
