+++ pynumdiff/optimize.py
@@ -2,7 +2,7 @@
 from hashlib import sha1
 from itertools import product
 from functools import partial
-from warnings import filterwarnings, warn
+from warnings import catch_warnings, filterwarnings, warn
 from multiprocessing import Pool, Manager
 import scipy.optimize
 import numpy as np
@@ -218,26 +218,28 @@ def optimize(func, x, dt, dxdt_truth=None, tvgamma=1e-2, search_space_updates={}
             None for k,v in search_space_types.items()] # None means no bound on a dimension
 
     results = []
-    filterwarnings("ignore", '', UserWarning) # An extra filtering call, because some worker work can actually be done in the main process
-    if parallel:
-        with Manager() as manager:
-            cache = manager.dict() # cache answers to avoid expensive repeat queries
-            with Pool(initializer=filterwarnings, initargs=["ignore", '', UserWarning]) as pool: # The heavy lifting
-                for categorical_combo in categorical_combos:
-                    # wrap the objective and scipy.optimize.minimize to pass kwargs and a host of other things that remain the same
-                    _obj_fun = partial(_objective_function, func=func, x=x, dt=dt, singleton_params=singleton_params,
-                        categorical_params=categorical_combo, search_space_types=search_space_types, dxdt_truth=dxdt_truth,
-                        metric=metric, tvgamma=tvgamma, padding=padding, cache=cache, huberM=huberM)
-                    _minimize = partial(scipy.optimize.minimize, _obj_fun, method=opt_method, bounds=bounds, options={'maxiter':maxiter})
-                    results += pool.map(_minimize, starting_points) # returns a bunch of OptimizeResult objects
-    else: # For experiments, where I want to parallelize optimization calls and am not allowed to have each spawn further processes
-        cache = {} # dict
-        for categorical_combo in categorical_combos:
-            _obj_fun = partial(_objective_function, func=func, x=x, dt=dt, singleton_params=singleton_params,
-                categorical_params=categorical_combo, search_space_types=search_space_types, dxdt_truth=dxdt_truth,
-                metric=metric, tvgamma=tvgamma, padding=padding, cache=cache, huberM=huberM)
-            _minimize = partial(scipy.optimize.minimize, _obj_fun, method=opt_method, bounds=bounds, options={'maxiter':maxiter})
-            results += [_minimize(p) for p in starting_points]
+    # An extra filtering call, because some worker work can actually be done in the main process. Scoped to this
+    # call, so the caller's filters are restored on the way out.
+    with catch_warnings(action="ignore", category=UserWarning):
+        if parallel:
+            with Manager() as manager:
+                cache = manager.dict() # cache answers to avoid expensive repeat queries
+                with Pool(initializer=filterwarnings, initargs=["ignore", '', UserWarning]) as pool: # The heavy lifting
+                    for categorical_combo in categorical_combos:
+                        # wrap the objective and scipy.optimize.minimize to pass kwargs and a host of other things that remain the same
+                        _obj_fun = partial(_objective_function, func=func, x=x, dt=dt, singleton_params=singleton_params,
+                            categorical_params=categorical_combo, search_space_types=search_space_types, dxdt_truth=dxdt_truth,
+                            metric=metric, tvgamma=tvgamma, padding=padding, cache=cache, huberM=huberM)
+                        _minimize = partial(scipy.optimize.minimize, _obj_fun, method=opt_method, bounds=bounds, options={'maxiter':maxiter})
+                        results += pool.map(_minimize, starting_points) # returns a bunch of OptimizeResult objects
+        else: # For experiments, where I want to parallelize optimization calls and am not allowed to have each spawn further processes
+            cache = {} # dict
[…]
