### 2026-08-20 9a7bf61
Loosen a spectraldiff error bound
Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
