+++ pynumdiff/tests/test_diff_methods.py
@@ -165,7 +165,7 @@ error_bounds = {
                         [(1, 0), (3, 3), (1, 0), (3, 3)]],
     spectraldiff: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
                    [(0, 0), (1, 1), (0, 0), (1, 1)],
-                   [(1, 1), (1, 1), (1, 1), (1, 1)],
+                   [(1, 0), (1, 1), (1, 1), (1, 1)],
                    [(0, 0), (1, 1), (0, 0), (1, 1)],
                    [(1, 1), (2, 2), (1, 1), (2, 2)],
                    [(1, 1), (3, 3), (1, 1), (3, 3)]],
