### 2026-08-20 florisvb/PyNumDiff#206 @pavelkomarov
https://github.com/florisvb/PyNumDiff/commit/e8a2f92675f3a5888f9a97f8b4f1d8343b908ede
