### 2026-08-20 florisvb/PyNumDiff#207 @pavelkomarov
Hmm. A search space of only one point isn't much of a space, but it's an edge case worth addressing, if only to nip confusion in the bud. It could even surface a custom warning and then just fall through to a single run? I kind of don't want to create a whole different path in the code, though, especially since the way cartesian products and zipping tuples together is happening in here is already pretty confusing, necessarily so to handle categoricals, singletons, and numerical ranges. I'll figure out a way to make it not error in an unexpected way at least.
