### 2026-08-20 florisvb/PyNumDiff#207 @pavelkomarov
Claude figured out this error can also happen if no parameter is given as a list, like you only have categoricals and singletons. That's worse than expected.
