### 2026-08-20 620e2d2
Handle an empty numerical search space, and flatten the parallel sweep
optimize() built its Nelder-Mead starting points with a Cartesian
product over the parameters given as lists. With no such parameter that
product is [()] rather than empty, so minimize() got handed a
zero-length point and an empty bounds list and died in
old_bound_to_new with "not enough values to unpack (expected 2, got
0)". Passing bounds=None doesn't help; Nelder-Mead can't build a
simplex in 0-D either way. This hit more than the all-pinned case: a
categorical set alongside singletons leaves the numerical space empty
to
[…]
