+++ pynumdiff/optimize.py
@@ -134,14 +134,14 @@ def _objective_function(point, func, x, dt, singleton_params, categorical_params
     :return: float, cost of this objective at the point
     """
     # Short circuit if this hyperparam combo has already been queried, ~10% savings per #160
-    key = sha1((''.join(f"{v:.3e}" for v in point) + # This hash is stable across processes. Takes bytes
+    key = sha1((''.join(f"{v:.6e}" for v in point) + # This hash is stable across processes. Takes bytes
                ''.join(str(v) for k,v in sorted(categorical_params.items()))).encode()).digest()
     if key in cache: return cache[key]
 
     # Query the differentiation method at this choice of hyperparameters
     point_params = {k:(v if search_space_types[k] == float else int(np.round(v)))
                         for k,v in zip(search_space_types, point)} # point -> dict
-    # add back in singletons and categorical choices the Nelder-Mead isn't searching over
+    # add back in singletons and categorical choices the numerical minimzation isn't searching over
     try: x_hat, dxdt_hat = func(x, dt, **point_params, **singleton_params, **categorical_params) # estimate x and dxdt
     except np.linalg.LinAlgError: cache[key] = 1e10; return 1e10 # some methods can fail numerically
 
@@ -174,10 +174,13 @@ def optimize(func, x, dt, dxdt_truth=None, tvgamma=1e-2, search_space_updates={}
     :param np.array[float] dxdt_truth: actual time series of the derivative of x, if known
     :param float tvgamma: Only used if :code:`dxdt_truth` is *not* given. Regularization value used to select for parameters
                     that yield a smooth derivative. Larger value results in a smoother derivative.
-    :param dict search_space_updates: Each method has a default search space of parameter settings, structured as
-                    :code:`{param1:[numerical, values], param2:{categorical, values}, param3:value, ...}` (defined in
-                    :code:`_optimize.py`). The Cartesian product of values are used as initial starting points in optimization.
-                    If left None, the default search space is used, if :code:`{param1:[different,values]}`, these are applied.
+    :param dict search_space_updates: Each method has a default search space of parameter settings, encoded as
+                    :code:`{param1:[numerical, values], param2:{categorical, values}, param3:value, ...}` (defined at the top of
+                    :code:`pynumdiff/optimize.py`). The Cartesian product of dictionary values serves as initialization points,
+                    where values given as singletons or in sets are plugged into separate minimization runs (since some algos cannot
+                    handle discrete search), and values given in lists serve as seeds for search across continuous dimensions.
+                    This parameter optionally accepts a `dictionary update <https://docs.python.org/3/library/stdtypes.html#dict.update>`_
+                    to override particular or multiple parameter values.
     :param str metric: either :code:`'rmse'` or :code:`'error_correlation'`, only applies if :code:`dxdt_truth` is given
     :param int padding: number of steps to ignore at the beginning and end of the data series, or :code:`'auto'` to ignore
                     2.5% at each end. Larger value causes the optimization to emphasize the accuracy in the series middle.
@@ -193,8 +196,7 @@ def optimize(func, x, dt, dxdt_truth=None, tvgamma=1e-2, search_space_updates={}
     :return: - **opt_params** (dict) -- best parameter settings for the differentation method
              - **opt_value** (float) -- lowest value found for objective function
     """
-    if metric not in ['rmse','error_correlation']:
-        raise ValueError('`metric` should either be `rmse` or `error_correlation`.')
+    if metric not in ['rmse','error_correlation']: raise ValueError('`metric` should either be `rmse` or `error_correlation`.')
 
[…]
