### 2026-08-20 florisvb/PyNumDiff#207 @pavelkomarov
https://github.com/florisvb/PyNumDiff/commit/620e2d24fd4f239f65345b7f1e740349ebb69130
