### 2026-08-20 a55465c
improved robustdiff docstring some
