+++ pynumdiff/kalman_smooth.py
@@ -268,27 +268,25 @@ def constant_jerk(x, dt, params=None, options=None, r=None, q=None, forwardbackw
 def robustdiff(x, dt_or_t, order, log_q, log_r, proc_huberM=6, meas_huberM=0, axis=0):
     """Perform outlier-robust differentiation by solving the Maximum A Priori optimization problem:
     :math:`\\text{argmin}_{\\{x_n\\}} \\sum_{n=0}^{N-1} V(R^{-1/2}(y_n - C x_n)) + \\sum_{n=1}^{N-1} J(Q_{n-1}^{-1/2}(x_n - A_{n-1} x_{n-1}))`,
-    where :math:`A,Q,C,R` come from an assumed constant derivative model and :math:`V,J` are the :math:`\\ell_1` norm or Huber
-    loss rather than the :math:`\\ell_2` norm optimized by RTS smoothing. This problem is convex, so this method calls
-    :code:`convex_smooth`, which in turn forms a sparse CVXPY problem and invokes CLARABEL.
+    where :math:`A,Q,C,R` come from an assumed constant-derivative model (see `Section 7.7 of this paper <https://arxiv.org/pdf/2512.09090>`_)
+    and :math:`V,J` are the :math:`\\ell_1` norm or Huber loss rather than the :math:`\\ell_2` norm optimized by RTS smoothing. This problem is
+    convex, so this method calls :code:`convex_smooth`, which in turn forms a sparse CVXPY problem and invokes CLARABEL.
 
-    Note that for Huber losses, :code:`M` is the radius where the Huber loss function turns from quadratic to linear. Because
-    all loss function inputs are normalized by noise level, :math:`q^{1/2}` or :math:`r^{1/2}`, :code:`M` is in units of inlier
-    standard deviation. In other words, this choice affects which portion of inliers might be treated as outliers. For example,
-    assuming Gaussian inliers, the portion beyond :math:`M\\sigma` is :code:`outlier_portion = 2*(1 - scipy.stats.norm.cdf(M))`.
-    The inverse of this is :code:`M = scipy.stats.norm.ppf(1 - outlier_portion/2)`. As :math:`M \\to \\infty`, Huber becomes the
-    1/2-sum-of-squares case, :math:`\\frac{1}{2}\\|\\cdot\\|_2^2`, because the normalization constant of the Huber loss (See
-    :math:`c_2` in `section 6 of this paper <https://jmlr.org/papers/volume14/aravkin13a/aravkin13a.pdf>`_, missing a
-    :math:`\\sqrt{\\cdot}` term there, see p2700) approaches 1 as :math:`M` increases. Similarly, as :code:`M` approaches 0,
-    Huber reduces to the :math:`\\ell_1` norm case, because the normalization constant approaches :math:`\\frac{\\sqrt{2}}{M}`,
-    cancelling the :math:`M` multiplying :math:`|\\cdot|` in the Huber function, and leaving behind :math:`\\sqrt{2}`, the
-    proper :math:`\\ell_1` normalization.
+    Note that for Huber losses, :code:`M` is the radius where the Huber loss function turns from quadratic to linear. Because all loss function
+    inputs are normalized by noise level (scales :math:`q^{1/2}` and :math:`r^{1/2}` from the :math:`Q` and :math:`R` matrices in the
+    constant-derivative model), :code:`M` is in units of inlier standard deviation. In other words, this choice affects which portion of inliers
+    might be treated as outliers. For example, assuming Gaussian inliers, the portion beyond :math:`M\\sigma` is
+    :code:`outlier_portion = 2*(1 - scipy.stats.norm.cdf(M))`. The inverse of this is :code:`M = scipy.stats.norm.ppf(1 - outlier_portion/2)`.
[…]
