### 2026-08-20 florisvb/PyNumDiff#205 @alexzerg
Hello @pavelkomarov
I do confirm

```
pip_install_base=PASS
No broken requirements found.
import_pynumdiff=PASS
import_optimize=PASS True
tqdm_version=4.70.0
cvxpy_installed=False
```
