### 2026-08-20 florisvb/PyNumDiff#206 @alexzerg
Retested on current joss2.0, filters are restored after optimize() and warnings are visible again.
Thx!
