### 2026-08-20 florisvb/PyNumDiff#208 @alexzerg
rtsdiff gives wrong circular x_hat with forwardbackward=True
I was checking test_circular_rtsdiff and tried the same mode with forwardbackward=True. Derivative looks very good but returned x_hat becomes wrong around the wraps.
This is the small case I used on current joss2.0 python 3.13:
```
import numpy as np
from pynumdiff.kalman_smooth import rtsdiff

dt = 0.1
t = np.arange(0, 10 + dt, dt)
velocity = 5.0
th = np.angle(np.exp(1j * velocity * t))

th_hat, dthdt_hat = rtsdiff(
    th,
    dt,
    order=1,
    log_qr_ratio=1,
    circular=True,
    forwardbackward=True,
)

angle_error = np.angle(np.exp(1j * (th_hat - th)))

print("angle rmse:", np.sqrt(np.mean(angle_error**2)))
print("max angle error:", np.max(np.abs(angle_error)))
print("dxdt rmse:", np.sqrt(np.mean((dthdt_hat - velocity)**2)))
```

I get:
```
angle rmse:      1.803355683374138
max angle error: 3.0159292514473095
dxdt rmse:       6.98365209849463e-05
```

With forwardbackward=False angle rmse on the same signal is around 0.00052.
I checked the two smoother outputs and they can differ by several full rotations, raw difference was up to 50.27 radians here. They are averaged as normal numbers before wrapping back to [-pi, pi] so I think the average lands on a wrong branch.

Just in case I've mixed the two estimates as circular values instead. Angle rmse went down to 2.65e-05.
The current circular test uses default forwardbackward=False so this combination seems not covered. Please take a look.

Review reference: https://github.com/openjournals/joss-reviews/issues/11172
