### 2026-08-20 florisvb/PyNumDiff#204 @alexzerg
Thank you Pavel. This explains the 0.2 vs 2.0 part. I think the easiest way is to drop 2.0 from the title and mention in the introduction: this is the second major generation while the package is still using pre-1.0 versions.
