### 2026-08-20 #11172 @alexzerg
> Nice, thanks for all the movement on this. I notice the License checkbox isn't ticked, but there certainly is one. https://github.com/florisvb/PyNumDiff/blob/master/LICENSE.txt Do I need to drop the `.txt` for tools to pick it up automatically?

LICENSE.txt is fine, no need to rename it. That checkbox was on my side, I've ticked it now.
