### 2026-08-20 florisvb/PyNumDiff#209 @alexzerg
spectraldiff keeps full spectrum when cutoff rounds to zero FFT bins
I was checking the spectraldiff values used in optimize.py. The default search space includes 1e-5, 1e-4 and 1e-3 and for a normal sized signal they can all round to zero FFT bins. This is what I tried on current joss2.0 python 3.13:
```
import numpy as np
from pynumdiff.basis_fit import spectraldiff

N = 101
dt = 0.01
t = np.arange(N) * dt
x = np.sin(2*np.pi*3*t) + 0.2*np.sin(2*np.pi*30*t)

effective_n = 2 * (N + 200)

for cutoff in [0, 1e-5, 1e-4, 1e-3, 0.003, 0.004]:
    _, dxdt_hat = spectraldiff(x, dt, high_freq_cutoff=cutoff)
    fft_bin = int(cutoff * effective_n / 2)
    print(cutoff, fft_bin, np.linalg.norm(dxdt_hat))
```
I get cutoff, FFT bin and derivative norm:
```
0       0   298.4522855517395
1e-05   0   298.4522855517395
0.0001  0   298.4522855517395
0.001   0   298.4522855517395
0.003   0   298.4522855517395
0.004   1   0.3018792976642558
```
So all values which become bin 0 keep the full spectrum. Then there is a very large jump as soon as cutoff becomes one bin.
Most probably it's related to this:
`filt[discrete_cutoff:-discrete_cutoff] = 0`
When discrete_cutoff == 0 this becomes filt[0:0] and nothing is removed. Since several of these small values are in the default optimizer search space, this also gives optimizer a strange discontinuity. I think zero-bin cutoff needs to be handled before applying the slice.

Review reference: https://github.com/openjournals/joss-reviews/issues/11172
