### 2026-08-21 florisvb/PyNumDiff#208 @pavelkomarov
Again good eye. The issue seems to be that forward and backward can essentially alias by whole rotations, so I've added a line to correct.
