### 2026-08-21 3bdd3e0
Blend forward and backward circular estimates on a common branch
With circular=True, rtsdiff wraps only the innovation and the final
output, so the filter state tracks a continuous unwrapped angle. The
forward and backward passes are seeded from opposite ends of the series
and therefore settle on branches that can be many full turns apart. The
forwardbackward blend combined them as plain reals, sweeping the result
through every intermediate branch, so x_hat came out near-garbage while
dxdt_hat stayed correct, velocity being branch-independent. Shift the
backward estimate by the whole number of
[…]
