### 2026-08-21 florisvb/PyNumDiff#208 @pavelkomarov
3bdd3e0dd7c1119aa34bb316c552f9438f4f2b48
