### 2026-08-21 florisvb/PyNumDiff#209 @pavelkomarov
A sneaky indexing bug. Tragic. `vector[0:-0]` references nothing, when what was meant was `vector[0:N-0]`, which references the whole thing.
