### 2026-08-21 934fb92
Zero the whole spectrum when the cutoff rounds below one FFT bin
spectraldiff converts high_freq_cutoff into a whole number of FFT bins,
and any cutoff below 2/N rounds down to zero. The filter then erased
filt[0:-0], and since -0 == 0 that is an empty slice, so nothing got
zeroed and the entire unsmoothed spectrum passed through. Asking for
maximum smoothing returned minimum smoothing, with a derivative norm
roughly 1000x too large and identical to what high_freq_cutoff=1 gives.
Naming the end of the slice explicitly as N-discrete_cutoff fixes it, so
a sub-bin cutoff now keeps no modes at all
[…]
