### 2026-08-21 florisvb/PyNumDiff#209 @pavelkomarov
934fb92b6a07fa6be914fc6265be6366d28a7864
