### 2026-08-21 florisvb/PyNumDiff#211 @pavelkomarov
I was using `len(x)` instead of `x.shape[axis]`, a holdover from the pre-multidimensional era.
