### 2026-08-21 24cd55f
Measure savgoldiff's window limits along the axis being differentiated
savgoldiff clamps window_size and smoothing_win against the length of the
data, but it measured that with len(x), which is always axis 0's length no
matter what `axis` says. Differentiating a (2, 41) array along axis 1 was
therefore treated as 2 samples rather than 41, and np.clip(11, 4, 1) with
an inverted range returns the upper limit, so window_size silently became
1 and savgol_filter failed with "polyorder must be less than
window_length". Use x.shape[axis] on both clamps.

An audit of the other eleven axis-aware method
[…]
