### 2026-08-21 florisvb/PyNumDiff#211 @pavelkomarov
24cd55f704292eea6c582cd03d02ff94ff4c2edb
