### 2026-08-21 florisvb/PyNumDiff#210 @pavelkomarov
If the stride is wider than the window length, then you're leaving some holes as you go. Claude also tells me it's possible the last few elements in an array can go uncovered by any window. Fixing both.
