### 2026-08-21 26f0a53
Give the tail a window when the stride steps over it
slide_function places window midpoints at multiples of stride, so when
len(x)-1 is not one of them the final window can stop short of the end of
the array. Samples past it accumulate no weight at all, and the closing
x_hat/weight_sum divides 0 by 0, returning NaN. lineardiff runs the slider
forwards and backwards and then takes a finite difference, so a couple of
orphaned samples became four NaNs in x_hat and six in dxdt_hat.

Append the last index as an extra midpoint, but only when the leftover
after the final regular midpoint outruns that
[…]
