### 2026-08-21 florisvb/PyNumDiff#210 @pavelkomarov
26f0a5364eb631fe0f9773efe7fce4b391e1ea5a
