### 2026-08-21 #11172 @pavelkomarov
Thanks for finding so many bugs. I've been able to churn through and diagnose them with Claude's help and make some patches. Remarkably many fixes in fairly few lines changed.
