### 2026-08-21 a9bcbc1
Clarify what the second generation adds, and tighten the paper
Drops "2.0" from the title, since the reviewed release is 0.2.3 and the
mismatch was causing confusion, and notes the generation and version in the
Summary instead. Replaces the vague changelog paragraph in State of the Field
with concrete claims a reader can check against Table 1.

Trims roughly 110 words of duplication to make room: the Statement of Need no
longer restates the Summary's opening or the optimization details Software
Design covers, the variable-spacing and missing-data paragraphs merge into one
since Table 1 already
[…]
