### 2026-08-21 6747ce0
Note suggest_method and optimizer caching
Auditing the paper against the v0.2 through v0.2.3 release notes turned up two
omissions. Caching was dropped in the previous commit's trim, and
suggest_method was never mentioned at all, though it is the clearest
demonstration of what the unified interface buys a user.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
