### 2026-08-21 florisvb/PyNumDiff#204 @pavelkomarov
Both items are addressed on `joss2.0`.

**Title** now reads *PyNumDiff: Practical Numerical Differentiation for Noisy Data*, with the `2.0` dropped as you suggested. The Summary carries the clarification instead: this is the package's second generation, release numbers stay below 1.0 by convention, and the version described here is 0.2.3.

**1.x vs 2.0** is now a paragraph in State of the Field that makes checkable claims rather than reciting a changelog. `kerneldiff`, `rtsdiff`, and `tvrdiff` each cover by parameter what had been separate functions, so the package presents fewer entry points than before while spanning more methods. `rbfdiff`, `waveletdiff`, and `robustdiff` are new. A single optimizer serves all twelve methods of Table 1, and four capabilities span the package: multidimensional data, irregular sample spacing, missing observations, and circular domains. Each of those is verifiable against the table on the same page.

Auditing the paper against the v0.2 through v0.2.3 release notes separately turned up `suggest_method`, which was missing entirely and is now described in Software Design.

The paper holds at roughly its previous length (~1,200 words of prose). The room for the new material came from cutting duplication between sections rather than from compressing sentences.

— Claude (Opus 5), working with @pavelkomarov
