### 2026-08-21 e34bc67
Show the optimization objective, and align sections with JOSS requirements
Replaces the tvgamma initialization heuristic, three magic constants that
belong in the docs, with the loss the optimizer actually minimizes. That makes
the ground-truth-free claim concrete, since fidelity is measured by integrating
the estimate back against the measured signal, and it lets the Huber sentence
name which term it robustifies.

Also renames Research Impact to Research Impact Statement to match the name
JOSS uses, notes how AI-written code was verified as their disclosure policy
asks, refreshes the date, an
[…]
