### 2026-08-21 e78a8b0
Add ORCIDs for van Breugel and Kutz, restore submission date
The date field is the submission date, and the pre-review issue confirms that
was 1 April 2026, so the refresh to August was wrong. Also drops the
verification clause from the AI disclosure, since the existing sentence about
reviewing and editing by hand already describes how outputs were checked.

ORCIDs confirmed against the public registry by employment record. Komarov and
Protogerou remain unlisted.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
