### 2026-08-21 2571b44
Add Komarov ORCID
Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
