### 2026-08-21 370b3fa
Describe the notebooks as reproducible material, fix van Breugel citations
The Research Impact Statement had shrunk the notebooks to a passing clause,
which undersells them. JOSS accepts reproducible materials and benchmarks as
impact evidence, and the seven notebooks demonstrate nearly every capability
the paper claims, so they are now described by what they cover. Also corrects
the claim that they live on the docs site; they live in the repository, and
readthedocs carries the API documentation.

The bibliography mixed two name encodings for van Breugel, so the 2020 and
2022 entries rendered 
[…]
