+++ paper/paper.bib
@@ -6,14 +6,14 @@
   volume    = {7},
   number    = {71},
   pages     = {4078},
-  author    = {Floris van Breugel and Yuying Liu and Bingni W. Brunton and J. Nathan Kutz},
+  author    = {van Breugel, Floris and Liu, Yuying and Brunton, Bingni W. and Kutz, J. Nathan},
   title     = {{PyNumDiff}: A {Python} package for numerical differentiation of noisy time-series data},
   journal   = {Journal of Open Source Software}
 }
 
 @misc{komarov2025,
   title         = {A Taxonomy of Numerical Differentiation Methods},
-  author        = {Pavel Komarov and Floris van Breugel and J. Nathan Kutz},
+  author        = {Komarov, Pavel and van Breugel, Floris and Kutz, J. Nathan},
   year          = {2025},
   eprint        = {2512.09090},
   archivePrefix = {arXiv},
@@ -24,7 +24,7 @@
 @article{vanBreugel2020numerical,
   doi     = {10.1109/ACCESS.2020.3034077},
   year    = {2020},
-  author  = {Floris {van Breugel} and J. Nathan Kutz and Bingni W. Brunton},
+  author  = {van Breugel, Floris and Kutz, J. Nathan and Brunton, Bingni W.},
   journal = {IEEE Access},
   title   = {Numerical differentiation of noisy data: A unifying multi-objective optimization framework},
   volume  = {8},
