### 2026-08-21 8e10fa4
Keep Table 1 on one page, and fix the van Breugel surname
The table split across pages, stranding two rows under a repeated header. Six
of its twelve rows carried no checkmarks, so the table now lists only methods
with capabilities beyond the multidimensional support they all share, and the
lead-in states the full count.

Comma form alone left the CSL style treating "van" as a particle, so citations
rendered as "Breugel et al." Bracing the surname inside comma form makes it a
single unit.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
