### 2026-08-21 3fd0132
Group Table 1 by capability so every method appears
Restores the six methods dropped in the previous commit without bringing back
the page break. The matrix spent half its height on empty cells because it was
keyed by method, so it is now keyed by capability, listing the methods that
have each. Every cell carries information, all twelve methods are named, and
the table is shorter than either previous version.

It also answers the question users actually ask. Someone with gaps in their
data wants to know which methods tolerate that, rather than reading down a
column of blanks.

Co-Authored-By: 
[…]
