### 2026-08-21 5850b07
Return Table 1 to a method-keyed matrix, with the plain methods on one row
Keeps the matrix shape, which shows at a glance how many capabilities a method
carries, while still naming all twelve. The six that support only
multidimensional data share a single trailing row rather than six rows of
empty cells, so the table stays short enough to hold one page.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
