### 2026-08-21 3a81739
Revert Table 1 to the capability-keyed form, which fits on one page
The collapsed matrix did not work. Its first column is narrow, so the six
method names in the trailing row stacked one per line, making the table no
shorter than the version it replaced and splitting it across pages again.

The capability-keyed table gives its methods column half the page, so twelve
names wrap to two or three lines. Measured against the proof, it holds one
page while still naming every method.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
