### 2026-08-21 3ab9394
Give CITATION.cff a citation hierarchy, and add the companion paper
The file told humans to cite the article while telling tools to cite the
software, because the article sat under references, which the GitHub citation
widget ignores. It is now preferred-citation, so both agree.

The title was the paper's rather than the software's, which conflated the two
and would have gone stale every time the paper was retitled. It is now just
PyNumDiff.

The 2022 article carries version 0.1.2.4 and a note saying which generation it
describes, since CFF has no way to express that one work supersedes anothe
[…]
