### 2026-08-21 florisvb/PyNumDiff#209 @alexzerg
Thanks Pavel - the full-spectrum part is fixed. However...![meme](https://i.imgflip.com/28g5q4.jpg)

With mean around 5 and default padding:
```
cutoff 0.001, bin 0 -> mean(x_hat) = 0
cutoff 0.004, bin 1 -> mean(x_hat) = 5.0004
```
—> bin-0 case also removes DC. I think a positive cutoff below the first non-zero bin should keep DC and zero the rest.
