### 2026-08-21 florisvb/PyNumDiff#212 @alexzerg
default optimize(butterdiff, ...) crashes on its own search space
Hi Pavel. Finally I’ve found something critical!
I tested butterdiff with the default optimizer search, without any search space updates:
```
import numpy as np

from pynumdiff.optimize import optimize
from pynumdiff.smooth_finite_difference import butterdiff

dt = 0.01
t = np.arange(201) * dt
x = np.sin(2*np.pi*2*t)
truth = 2*np.pi*2*np.cos(2*np.pi*2*t)

optimize(
    butterdiff,
    x,
    dt,
    dxdt_truth=truth,
    parallel=False,
    maxiter=1,
)
```
—>

```
ValueError: Filter not stable due to sum(a) == 0,
i.e., having a pole at z = 1!
```
And the same with parallel=True.

It happens at filter_order=7 and cutoff_freq=0.0001. Both values come from the default search space so normal optimize(butterdiff, ...) call reaches this combination by itslef. butterdiff currently uses the (b, a) filter form. I tried the same order/cutoff with scipy output="sos" and sosfiltfilt and that stays finite.

Review reference: https://github.com/openjournals/joss-reviews/issues/11172
