### 2026-08-24 florisvb/PyNumDiff#209 @pavelkomarov
If you crank all the way down to zero, I think you should remove all modes, including DC. There's an argument to be made that a nonzero cutoff should keep DC, though, rather than round down and potentially exclude it. Perhaps I need a ciel() somewhere.
