### 2026-08-24 9035ee7
Keep the DC term for any positive spectraldiff cutoff
high_freq_cutoff is a fraction of Nyquist, converted to a whole number of
FFT bins. Truncating that conversion meant every cutoff below 2/N landed
on 0 bins and zeroed the entire spectrum, DC included, so a signal with
mean 5 came back with mean 0. Rounding up instead keeps wavenumber k
exactly when its frequency k/(N/2) is below the cutoff, so any positive
cutoff keeps at least DC and only high_freq_cutoff=0 asks for nothing at
all.

The filter's stop band was also off by one on the negative side: filt
zeroed [cutoff, N-cutoff), which keep
[…]
