### 2026-08-24 florisvb/PyNumDiff#209 @pavelkomarov
Claudefucius say:

Rounding the cutoff down to whole FFT bins was the leftover half of this: any cutoff below 2/N landed on 0 bins, which zeroed everything including DC. `ceil` instead of `int` means bin $k$ is kept exactly when its frequency $k/(N/2)$ is below the cutoff, so any positive cutoff keeps at least DC, and only `high_freq_cutoff=0` asks for nothing at all.

While in there I also fixed an off-by-one on the negative side of the stop band. `filt[cutoff:N-cutoff]` keeps $-cutoff$ while $+cutoff$ is already gone, so the kept spectrum wasn't conjugate-symmetric and `np.real()` was quietly returning half of that boundary mode. That's why a DC-only cutoff still showed mean 5.00039 and a nonzero derivative norm rather than a flat line.

Now, on your signal with mean 5:

```
cutoff 0       mean(x_hat) 0.00000   ||dxdt|| 0.00000
cutoff 1e-05   mean(x_hat) 5.00039   ||dxdt|| 0.00000
cutoff 0.001   mean(x_hat) 5.00039   ||dxdt|| 0.00000
cutoff 0.003   mean(x_hat) 5.00039   ||dxdt|| 0.00000
cutoff 0.004   mean(x_hat) 5.00039   ||dxdt|| 0.60376
cutoff 0.05    mean(x_hat) 4.99951   ||dxdt|| 35.54631
cutoff 1.0     mean(x_hat) 5.00000   ||dxdt|| 298.45229
```

Fixed in 9035ee76273ad7dd98d7be6a9813c0ef606964a6.
