### 2026-08-24 florisvb/PyNumDiff#209 @pavelkomarov
Done, and in the process Claude found another bug, where I wasn't keeping only conjugate pairs of terms. The `np.real` call a few lines down would cull the leftover unpaired term, so nbd from a results perspective, but better to get the indexing fully corrected. So in the end, not just `N-cutoff` rather than `-cutoff`, but also a `+1` in there.
