### 2026-08-24 336abe8
realized there is a much less bug-prone and shorter way to to the same wavenumber zeroing in spectraldiff, following the example from my https://github.com/pavelkomarov/spectral-derivatives package. Just do a where() call.
