### 2026-08-24 florisvb/PyNumDiff#209 @pavelkomarov
Actually, there's a simpler solution that gets the same result in a much more readable, less bug-prone way by using `where()`, as I did over in my other package https://github.com/pavelkomarov/spectral-derivatives/blob/main/specderiv/specderiv.py, https://github.com/pavelkomarov/spectral-derivatives/blob/main/notebooks/filtering_noise.ipynb

https://github.com/florisvb/PyNumDiff/commit/336abe806e773ecd73e2656ef1163a8915820eff
