### 2026-08-24 cff93c6
Build butterdiff's filter in second-order sections
The (b, a) transfer function form falls apart at high order and low cutoff,
where the poles bunch up near z = 1. At order 7 with cutoff 1e-4 the
denominator coefficients reach magnitude 35 but must sum to 3e-25, so the
computed sum is 3.5e-15 of pure cancellation noise and the computed poles
land outside the unit circle. Recent scipy checks for this and raises
"Filter not stable due to sum(a) == 0"; older scipy silently returns a
flattened signal instead. Both order and cutoff come from butterdiff's
default search space, so plain optimize(butt
[…]
