### 2026-08-24 florisvb/PyNumDiff#212 @pavelkomarov
From Claude:

Confirmed, and it turned out to be hiding a second bug next to it. Fixed in cff93c6cc062d79880d8ac27e7c7570754f1630e.

**The filter form.** You're right that `sos` is the answer. The `(b, a)` transfer function is hopeless at high order and low cutoff, where the poles bunch up near $z=1$. At order 7, cutoff 1e-4, the denominator coefficients reach magnitude 35 but have to sum to 3.0e-25 (a lowpass Butterworth has unit DC gain, so `sum(a) == sum(b)`). In double precision, summing numbers of size 35 gets you an answer good to about 1e-14, so the computed `sum(a)` of 3.5e-15 is entirely cancellation noise - the computed poles even land outside the unit circle. Whether it rounds to exactly 0.0 and trips the stability check is luck of the arithmetic, which is why I couldn't reproduce your traceback at first: the check you hit is in scipy's rewritten `lfilter_zi` on main, and on scipy 1.15.2 the same call doesn't raise at all - it just silently returns a flattened signal. Louder on your version, quieter on mine, wrong on both. Second-order sections never build that polynomial, so `sosfiltfilt` is accurate across the whole search space.

**The second bug.** Chasing it, the padding length in the same three lines was hardcoded for one filter order:

```python
padlen = x.shape[axis]-1 if x.shape[axis] < 9 else None
```

scipy pads by `3*(filter_order + 1)` samples, so 9 is right only for order 2 - and the search space goes to order 10, which wants 33. Any signal shorter than ~34 samples died with *"The length of the input vector x must be greater than padlen, which is 33."* It was off by one at the boundary as well: a length-9 signal at order 2 failed, since scipy wants strictly greater. So there were two independent ways for the default search space to break, and short data hit the other one. Both are fixed by computing the real default and clipping it to what the data can afford.

**Effect on results.** None that I can find. I reconstructed the old `butterdiff` and ran both through `optimize` on all six simulated examples with the `tvgamma` objective, and the selected hyperparameters and resulting RMSE are identical - the optimizer already rejected the degenerate corner because a filter that narrow flattens the signal and scores badly, and everywhere else the two filter forms agree to ~1e-10. So the crash was costing runs, not corrupting numbers.
