### 2026-08-24 #11172 @alexzerg
Done from my side. I checked the fixes again locally and all issues I opened are resolved. The paper and checklist look good to me, all boxes are checked and consider my review is complete.

AI assistance was used for language editing. All review judgments and decisions were my own.
