### 2026-08-25 florisvb/PyNumDiff#216 @sidxz
Paper/README updates to match code.
I am consolidating text-level things I encountered while checking the paper's claims against the code.
None of them are big, just needs updating.

1. Table 1 and the sentence "All handle multidimensional data via `axis`" (and the README's "All methods can accept blocks of multidimensional data") include lineardiff, but it doesn't have an `axis` argument:
   ```python
   lineardiff(np.stack([x, x]), dt, order=3, gamma=0.01, window_size=11, solver='CLARABEL', axis=1)
   # TypeError: lineardiff() got an unexpected keyword argument 'axis'
   ```
   Your #211 commit message already says "the other eleven axis-aware methods", so I guess the text just didn't get updated.

2. State of the Field says the `derivative` package "lacks multidimensional support", but `derivative.dxdt(x, t, kind, axis=0)` does have an `axis` argument (and takes a `t` array). The NaN-handling and optimizer parts of that sentence look right to me.

3. Affiliations: Please cross check that the authors match their respective affiliations, as mentioned in their ORCID.

Review reference: https://github.com/openjournals/joss-reviews/issues/11172
