### 2026-08-25 #11172 @sidxz
@prashjha @pavelkomarov My review is complete.

Short version: This is an amazing work and I believe would benefit the greater scientific community.

The software is in good shape and the paper's claims hold up under direct testing.

First! 2.0's new claim holds. rtsdiff bumps at every glitch whereas robustdiff doesn't notice!

I have a handful of other findings, none of them are serious. Details below.

I installed the version (`pynumdiff 0.2.3` from PyPI, Python 3.13, macOS arm64) and checked all rows of the paper's Table 1 plus the optimizer and the performance claim against a noisy test signal with a known derivative.

- The unified `(x_hat, dxdt_hat)` interface works across all twelve methods, and the legacy positional calls still work with deprecation warnings.
- Variable-step sampling (5 methods), missing data for the Kalman and spline methods (`rtsdiff` bridges a 15-sample dropout with no visible artefact), outlier robustness (`robustdiff` is unmoved by glitches that pull `rtsdiff` off by 7×; `tvrdiff`'s Huber option is 3 - 10× better than squared loss at equal γ).
- Blind hyperparameter optimization with the README's γ heuristic lands within ~5× of optimum and well ahead of hand-picked values.
- `robustdiff` scales linearly with signal length (log–log slope 1.04 over 1k–16k samples).
- Coverage is 92 % at both v0.2.3 and `master`.

**New issues:**
- florisvb/PyNumDiff#213 `lineardiff` uses a non-converged solver output with the default solver cvxpy; the solver status is not checked *(should fix; not blocking)*

- florisvb/PyNumDiff#214 `polydiff` crashes on a run of consecutive NaNs, the limit might need some documentation. *(should fix or document; the paper's "sensors that drop samples need no preprocessing" should name the methods it's true for)*

- florisvb/PyNumDiff#215 Fixed-step methods accept a time array and return `inf`; NaN-unaware methods return all-NaN output silently. Needs guarding. *(minor)*

- florisvb/PyNumDiff#216 Doc/Paper Updates. Please have a look at this issue. *(should fix;  blocking)*

**Checklist.** I've ticked everything except *Functionality* and *Reproducibility*, which can be checked off when the above issues are patched or acknowledged/addressed and *Software paper* items that depend on the Table 1 correction and the citations are corrected. Happy to re-check quickly when those land.

*AI-assistance disclosure: AI was used to gather public materials in spirit of acquiring more theoretical domain knowledge, understand and explain failures I observed. Every judgement in this review is mine.*
