### 2026-08-25 florisvb/PyNumDiff#216 @pavelkomarov
1. I actually just forgot to generalize `lineardiff`, because it's been like an unwanted stepchild the whole time, prone to cause convex solver troubles, yielding poor results, and complicated to look at in code. It's the one method I didn't get knee deep in and entirely rewrite. It's excluded from the analysis in Notebook 4 for these reasons. I'll generalize it and try to clean it up if I can, maybe try to lengthen some windows or something and see whether I can get it to behave more stably.
2. That's what I get for using Claude. I know the `derivative` package does multidimensional data, because this is critical for its use in the `pysindy` package. Another collaborator (@Jacob-Stevens-Haas) opened a very old issue on `derivative` about possible duplication with `pynumdiff`. Even before all the updates he thought `pynumdiff` was the more method-complete suite, and we've talked about making `pysindy` dependent on `pynumdiff` now that it's more feature-complete and has really pulled ahead. Multidimensional data handling was always meant to support that effort. https://github.com/dynamicslab/pysindy/issues/688
3. The affiliations are correct. Kutz took an industry job and moved to London, where I presently am as well for a year, and Maria was an undergrad in Kutz's primary department, so AI keeps suggesting I must have flipped them, but I didn't. It is odd that they're numbered out of order, though. I'll fix that by hand shortly.
