### 2026-08-26 florisvb/PyNumDiff#220 @pavelkomarov
estimate_integration_constant's sigma < 1e-3 guard is scale-dependent and discontinuous
`estimate_integration_constant` picks between an L2, L1, and Huber estimate of the integration constant using two `1e-3` thresholds:

```python
sigma = median_abs_deviation(x - x_hat, axis=axis, scale='normal').reshape(s)
if M == float('inf') or np.all(sigma < 1e-3):   # If no scatter, then no outliers, so use L2
    return np.mean(x - x_hat, axis=axis).reshape(s)
elif M < 1e-3:                                   # small M looks like l1 loss ...
    return np.median(x - x_hat, axis=axis).reshape(s)
else:
    return minimize(...)                         # Huber
```

The `M < 1e-3` test is fine — Huber genuinely limits to L1 as `M` shrinks, so the branches converge at the boundary.

The `sigma < 1e-3` test is not. `sigma` is the MAD of the residuals, an **absolute** quantity in the units of the data, so this reads "if the data is small" rather than "if there is no scatter." The comment's premise, *"If no scatter, then no outliers"*, does not follow.

### The branches don't converge at the boundary

Same signal, same relative noise (10%), same relative outliers (5x amplitude, every 50th sample), varying only the overall amplitude:

| amplitude | sigma | `sigma < 1e-3` | returned c | huber branch | L2 branch | huber/L2 |
|---|---|---|---|---|---|---|
| 1.20e-02 | 1.255e-03 | False | -1.0908e-04 | -1.0908e-04 | -1.1561e-03 | 0.094 |
| 1.05e-02 | 1.098e-03 | False | -9.5449e-05 | -9.5450e-05 | -1.0115e-03 | 0.094 |
| 1.00e-02 | 1.046e-03 | False | **-9.0904e-05** | -9.0905e-05 | -9.6338e-04 | 0.094 |
| 9.00e-03 | 9.411e-04 | True  | **-8.6704e-04** | -8.1815e-05 | -8.6704e-04 | 0.094 |
| 8.00e-03 | 8.365e-04 | True  | -7.7070e-04 | -7.2724e-05 | -7.7070e-04 | 0.094 |

A 10% change in amplitude moves the returned constant by **10.6x**. The `huber/L2` ratio is a constant 0.094 at every scale, so the two branches disagree by a fixed factor everywhere — the threshold is choosing between two persistently different answers using absolute data scale as the tiebreaker, not settling a distinction that stops mattering near the boundary.

Consequence: anyone differentiating small-amplitude data (millivolts, microns, normalized units) silently gets the non-robust L2 estimate no matter what `M` they pass.

### Reproducer

```python
import numpy as np
from scipy import stats
from pynumdiff.utils import utility

rng = np.random.default_rng(0); N = 400
base_noise = rng.standard_normal(N); spike = np.zeros(N); spike[::50] = 5.0

for amp in [1e-2, 9e-3]:
    x  = amp*np.sin(np.arange(N)*0.03)
    xh = x + amp*0.1*base_noise + amp*spike        # outliers at 5x amplitude
    sigma = stats.median_abs_deviation(x - xh, scale='normal')
    c = float(np.ravel(utility.estimate_integration_constant(x, xh, M=6))[0])
    print(f"amp={amp:.0e}  sigma={sigma:.3e}  guard={sigma < 1e-3}  c={c:.4e}")
```

### Fix direction

Make the test relative rather than absolute. "No scatter" should mean scatter is negligible *compared to something* — the spread of `x`, or the residual magnitude — so the decision is invariant to rescaling the data. Alternatively drop the branch: as `sigma` shrinks with residuals shrinking proportionally, every residual stays within `M*sigma` and Huber already reduces to L2 on its own, so the special case may not be earning its keep.

---
🤖 Written by [Claude Code](https://claude.com/claude-code) on behalf of @pavelkomarov
