### 2026-08-26 63e0cca
fixing another potential source of overflow (still warning in the big optimizations of notebook 4), as part of #217
