+++ pynumdiff/utils/evaluate.py
@@ -96,7 +96,10 @@ def robust_rme(u, v, padding=0, M=6):
     s = slice(padding, len(u)-padding) # slice out data we want to measure
 
     sigma = stats.median_abs_deviation(u[s] - v[s], scale='normal') # M is in units of this robust scatter metric
-    return np.sqrt(2*np.mean(huber(M*sigma, u[s] - v[s])))
+    if sigma == 0: return 0. # no scatter; shortcircuit to the limit value to avert divide-by-zero below
+    # Scale residuals rather than M: huber(M*sigma, r) == sigma**2 * huber(M, r/sigma), so pull sigma back out
+    # through the sqrt, and normalize residuals so their squared sum (from mean of huber) doesn't overflow. See #217
+    return sigma*np.sqrt(2*np.mean(huber(M, (u[s] - v[s])/sigma)))
 
 
 def rmse(u, v, padding=0):
