### 2026-08-26 8b7d3ae
in fixing #217 Claude flagged that the conditions branching within estimate_integration_constant were sensitive to data scale, which opened a bit of a can of worms. The conditions are now simplified, although one of them now allows zero sigmas to slip through, which have to be specifically handled, which isn't actually so complicated in code but is complicated to understand. Finally happy with the comment there. Then claude also demonstrated to me that the optimization itself could be sensitive to data scale, so we've instituted a fix for that as well by solving for constants in units of sigma
[…]
