### 2026-08-26 7ff3bf8
realized there could be an easy test extension to make sure we don't regress on #220, since it's pretty subtle and depends on some sophisticated optimization setup. Claude tells me the changes we've made toward #220 allow the code to pass the new test, which sweeps down to little amplitudes, but the old code fails, as expected.
