### 2026-08-26 cb95c53
Loosen three machine-precision bounds after the #217/#220 rewrite
estimate_integration_constant used to short-circuit to the closed-form L2
mean whenever np.all(sigma < 1e-3). Removing that scale-dependent guard was
the point of #220, but every internal caller takes the default M=6, so in
practice the guard was doing something else entirely: on a noiseless fit the
residuals are pure round-off, sigma lands around 1e-14, and the guard routed
those cases past the optimizer to an exact answer. Nothing else sent them to
L2, so it was invisible until now.

Without it, exactly-representable noiseless
[…]
