+++ pynumdiff/tests/test_diff_methods.py
@@ -145,8 +145,8 @@ error_bounds = {
                           [(-2, -2), (0, 0), (0, -1), (1, 1)],
                           [(0, 0), (1, 1), (0, 0), (1, 1)],
                           [(0, 0), (3, 3), (0, 0), (3, 3)]],
-    savgoldiff: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
-                 [(flr, flr), (flr, flr), (0, -1), (0, 0)],
+    savgoldiff: [[(-11, flr), (flr, flr), (0, -1), (0, 0)],
+                 [(-11, flr), (flr, flr), (0, -1), (0, 0)],
                  [(-2, -2), (-1, -1), (0, -1), (0, 0)],
                  [(0, -1), (0, 0), (0, 0), (1, 0)],
                  [(1, 1), (2, 2), (1, 1), (2, 2)],
@@ -206,7 +206,7 @@ error_bounds = {
                          [(2, 1), (2, 2), (2, 1), (2, 2)],
                          [(1, 1), (3, 3), (1, 1), (3, 3)]],
     smooth_acceleration: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
-                          [(-10, -11), (-10, -10), (-1, -1), (-1, -1)],
+                          [(-10, -10), (-10, -10), (-1, -1), (-1, -1)],
                           [(-2, -2), (-1, -1), (-1, -1), (0, -1)],
                           [(0, 0), (1, 0), (0, -1), (1, 0)],
                           [(1, 1), (2, 2), (1, 1), (2, 2)],
