### 2026-08-26 728e5c6
Default lineardiff to CLARABEL, fixing #213
CVXPY picks OSQP for these subproblems and it stalls, returning a half-converged iterate as the answer, because the repeated integrals in X leave entries around 1e-5 and OSQP's absolute termination criteria never get there. Dropping the solver pin from the test table so it covers the default, and swapping the deprecated second_order import for finitediff while in here.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
