### 2026-08-26 19f794b
today I learned string literals concatenate without a + in python
