+++ pynumdiff/total_variation_regularization.py
@@ -71,6 +71,10 @@ def tvrdiff(x, dt, order, gamma, huberM=float('inf'), solver=None, axis=0):
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
+    if np.any(np.isnan(x)): raise ValueError("`x` may not contain NaN. CVXPY cannot form a problem with missing data.")
+    if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. The convex problem setup integrates with a cumulative"
+        "sum and penalizes variation between consecutive samples, both of which assume uniform steps.")
+
     x_hat = np.empty_like(x); dxdt_hat = np.empty_like(x)
 
     for vec_idx in np.ndindex(x.shape[:axis] + x.shape[axis+1:]):
