+++ pynumdiff/basis_fit.py
@@ -136,54 +136,44 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-def waveletdiff(x, dt, wavelet='db8', level=None, threshold=1.0, axis=0, mode='periodization'):
+def waveletdiff(x, dt, wavelet='db8', level=None, threshold=1.0, axis=0, mode='symmetric'):
     """Smooth and differentiate noisy data in a wavelet basis.
 
-    Three steps: (1) decompose x with the DWT and soft-threshold the detail
-    coefficients to denoise (Donoho-Johnstone universal threshold), reconstructing
-    a smoothed x_hat; (2) extend x_hat antisymmetrically so the periodic derivative
-    operator stays accurate at the edges; (3) recover the wavelet scaling
-    coefficients of x_hat and apply the analytic derivative of the wavelet basis.
+    Three steps: (1) decompose x with the DWT and soft-threshold the detail coefficients to denoise (Donoho-Johnstone
+    universal threshold), reconstructing a smoothed x_hat; (2) extend x_hat antisymmetrically so the periodic derivative
+    operator stays accurate at the edges; (3) recover the wavelet scaling coefficients of x_hat and apply the analytic
+    derivative of the wavelet basis.
 
-    The derivative differentiates the basis functions themselves rather than
-    finite-differencing the signal. PyWavelets treats the samples as finest-level
-    scaling coefficients, so x_hat is the interpolant x(t) = sum_n a_n phi(t/dt - n)
-    for the scaling function phi. Sampling x and its analytic derivative on the grid
-    gives two convolutions against phi and phi' evaluated at *integers*,
+    The derivative differentiates the basis functions themselves rather than finite-differencing the signal. PyWavelets
+    treats the samples as finest-level scaling coefficients, so x_hat is the interpolant x(t) = sum_n a_n phi(t/dt - n)
+    for the scaling function phi. Sampling x and its analytic derivative on the grid gives two convolutions against phi
+    and phi' evaluated at *integers*,
 
         x_hat = Phi @ a     and     x' = Phi_prime @ a,
 
-    so x' = Phi_prime @ Phi^-1 @ x_hat, exact for signals the basis can represent.
-    The integer samples phi(p), phi'(p) are the eigenvalue-1 and eigenvalue-1/2
-    eigenvectors of the refinement relation phi(t) = sqrt2 sum_k h_k phi(2t - k)
-    (the "connection coefficients"), normalized to reproduce constants and ramps.
-    This is the wavelet-basis representation of the derivative operator from
-    Beylkin (1992); the connection coefficients follow Latto, Resnikoff &
-    Tenenbaum (1991), for Daubechies' compactly supported wavelets (1988).
-
-    Because the DWT requires uniform spacing, this method only accepts a scalar
-    time step dt (not a vector of sample times). For non-uniformly sampled data,
-    use :func:`rbfdiff` or :func:`splinediff` instead.
+    so x' = Phi_prime @ Phi^-1 @ x_hat, exact for signals the basis can represent. The integer samples phi(p), phi'(p)
+    are the eigenvalue-1 and eigenvalue-1/2 eigenvectors of the refinement relation phi(t) = sqrt2 sum_k h_k phi(2t - k)
+    (the "connection coefficients"), normalized to reproduce constants and ramps. This is the wavelet-basis representation
+    of the derivative operator from Beylkin (1992); the connection coefficients follow Latto, Resnikoff & Tenenbaum (1991),
[…]
