### 2026-08-27 florisvb/PyNumDiff#222 @pavelkomarov
Test that methods are scale equivariant: f(a*x) should equal a*f(x)
Differentiation is a linear operator, so every method in this package should satisfy

```
f(a*x, dt) == a*f(x, dt)
```

for any scalar `a`. Nothing in `test_diff_methods.py` checks this today. It's cheap to test — one extra call per method, no ground truth needed, since the property is self-referential — and it catches a specific bug class that has bitten us repeatedly: hyperparameters carrying absolute units, which silently mean something different when the data is rescaled.

## Current state

Measuring `max|f(a*x) - a*f(x)| / max|a*f(x)|` over `a` in {1e-3, 1e3}, on `pi_cruise_control` with the parameters already in `diff_methods_and_params`:

```
iterative_velocity   1.06e+01
robustdiff           1.03e+00   (robust_irreg_step the same)
lineardiff           1.00e+00
jerk                 4.81e-07
rbfdiff              5.52e-13
all others          <=1.2e-14
```

A relative deviation near 1.0 means the output barely scales at all — the estimate is essentially decoupled from the input's amplitude.

The causes are all the same shape. `robustdiff`'s `log_q`/`log_r` are absolute process and measurement variances, `lineardiff`'s `gamma` is an absolute regularization weight, so in each case rescaling `x` changes the balance between data fidelity and the prior without the user asking for it. `rtsdiff` passes precisely because it takes `log_qr_ratio`, a ratio, instead of the two absolute terms.

`jerk` at 4.8e-07 is far better than the broken group but well off the ~1e-14 the rest achieve, so it's worth a look rather than an assertion.

## Why it's worth adding

`splinediff` was in the failing group until the fix in #218, at 1.84 and 35.5 on the two amplitudes. Its `s` was an absolute residual budget, so the same `s` meant wildly different smoothing at different noise levels — the root cause of #218. Rescaling `s` to a multiple of the noise energy made it exactly equivariant (now 6.33e-15). A scale-equivariance test would have surfaced that directly, instead of it being found by way of the optimizer running to its bounds floor and crashing worker processes.

The same reasoning applies to the `sigma < 1e-3` guard in #217/#220, which produced a 10.6x discontinuity for exactly this reason: an absolute threshold applied to data of arbitrary scale.

## Suggested shape

A parametrized test alongside the existing ones, asserting the deviation is at machine-precision level for methods that should be exact, with a per-method tolerance table like `error_bounds` for the ones that legitimately can't be (if any turn out to be). The three broken methods would need either fixing or an explicit documented exemption.

---
🤖 Written by [Claude Code](https://claude.com/claude-code) on behalf of @pavelkomarov
