### 2026-08-27 0b27146
Scale splinediff's `s` by the data's noise energy, addressing #218
`s` was an absolute residual budget, so a value tuned at one signal length and noise level was off by up to 512x elsewhere on notebook 4's grid, and the optimizer ran to its bounds floor and tripped the scipy knot-search failure that was killing pool workers. It now means a multiple of N*sigma^2 with sigma estimated by MAD on second differences, which is what UnivariateSpline's old s=None default assumed before the make_splrep port dropped it and left splinediff(x, dt) raising TypeError; seeds and bounds shrink to match, num_it
[…]
