### 2026-08-27 0da832c
Document capabilities in a table and stop claiming lineardiff takes `axis`, part of #216
The prose list of which methods handle what had drifted out of date and understated several capabilities, so it becomes a table with a column per capability, which also makes lineardiff's missing `axis` visible rather than contradicted by a blanket claim that every method accepts multidimensional blocks. Also lists all seven notebooks instead of two, names pywavelets and tqdm among the dependencies, and fixes an "intergration" typo.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
