### 2026-08-27 218e5ea
human touch to readme, and had claude set minimum version for scipy because we actually do depend on people being up to date enough
