### 2026-08-27 florisvb/PyNumDiff#216 @pavelkomarov
Okay, the readme and the paper have been updated to reflect reality. I'm still wanting to generalize `lineardiff` too, but it's actually a big lift to make it stable and reasonably optimizable, so I've put that task in #223 and just updated things to exclude `lineardiff` where it presently is left out in reality.
