+++ pynumdiff/optimize.py
@@ -1,5 +1,4 @@
 """Optimization functionality"""
-from hashlib import sha1
 from itertools import product
 from functools import partial
 from warnings import catch_warnings, filterwarnings, warn
@@ -20,13 +19,13 @@ from .linear_model import lineardiff
 # Map from method -> (search_space, bounds_low_hi)
 method_params_and_bounds = {
     kerneldiff: ({'kernel': {'mean', 'median', 'gaussian', 'friedrichs'},
-             'window_size': [5, 15, 30, 50],
+             'window_size': [5, 15, 31, 51],
           'num_iterations': [1, 5, 10]},
-            {'window_size': (1, 1e6),
-          'num_iterations': (1, 100)}),
-    meandiff: ({'window_size': [5, 15, 30, 50], # Deprecated method
+            {'window_size': (1, 1e6, 'odd'), # an even-width kernel has no center, so convolving with it
+          'num_iterations': (1, 100)}),         # shifts the signal half a sample before differencing
+    meandiff: ({'window_size': [5, 15, 31, 51], # Deprecated method
              'num_iterations': [1, 5, 10]},
-               {'window_size': (1, 1e6),
+               {'window_size': (1, 1e6, 'odd'),
              'num_iterations': (1, 100)}),
     butterdiff: ({'filter_order': set(i for i in range(1,11)), # categorical to save us from doing double work by guessing between orders
                    'cutoff_freq': [0.0001, 0.001, 0.005, 0.01, 0.1, 0.5],
@@ -44,13 +43,13 @@ method_params_and_bounds = {
               'window_size': [11, 31, 51, 91, 131]},
                {'step_size': (1, 100),
                    'degree': (1, 8),
-              'window_size': (10, 1000)}),
+              'window_size': (10, 1000, 'odd')}),
     savgoldiff: ({'degree': [2, 3, 5, 7, 10],
-             'window_size': [3, 10, 30, 50, 90, 130, 200, 300],
-           'smoothing_win': [3, 10, 30, 50, 90, 130, 200, 300]},
+             'window_size': [3, 11, 31, 51, 91, 131, 201, 301],
+           'smoothing_win': [3, 11, 31, 51, 91, 131, 201, 301]},
                  {'degree': (1, 12),
-             'window_size': (3, 1000),
-           'smoothing_win': (3, 1000)}),
+             'window_size': (3, 1000, 'odd'), # savgol_filter accepts even windows but silently returns a
+           'smoothing_win': (3, 1000, 'odd')}), # derivative evaluated half a sample late
     splinediff: ({'degree': {3, 4, 5}, # categorical, because degree is whole number, and there aren't many choices
                        's': [0.5, 1, 1.5], # multiples of the noise energy, so these hold at any N and noise level
           'num_iterations': [1, 3]}, # 5 and 10 never won across 24 sim/step/noise combinations
@@ -81,9 +80,9 @@ method_params_and_bounds = {
                 {'num_iterations': (1, 100), # gets expensive with more iterations
                           'gamma': (1e-4, 1e7)}),
     smooth_acceleration: ({'gamma': [1e-2, 1e-1, 1, 10, 100, 1000],
-                     'window_size': [3, 10, 30, 50, 90, 130]},
+                     'window_size': [3, 11, 31, 51, 91, 131]},
                           {'gamma': (1e-4, 1e7),
-                     'window_size': (3, 1000)}),
+                     'window_size': (3, 1000, 'odd')}),
     rtsdiff: ({'forwardbackward': {True, False},
                          'order': {1, 2, 3}, # for this few options, the optimization works better if this is categorical
[…]
