### 2026-08-27 florisvb/PyNumDiff#201 @pavelkomarov
### Status recheck, Aug 2026

Revisiting this since `finufft` has moved since the decision above. It now ships prebuilt wheels, so the "requires a full C++ toolchain" claim is no longer true everywhere — but the coverage is incomplete.

`finufft` 2.5.1 (uploaded 2026-04-08) publishes:

| Platform | Wheel? |
|---|---|
| macOS arm64 (14.0+) | yes |
| Linux x86_64 (manylinux2014 / manylinux_2_28 / musllinux) | yes |
| Windows x86_64 | yes |
| **macOS x86_64** | **no** |
| **Linux aarch64** | **no** |

The wheels are `py3-none-*`, i.e. the compiled library is bundled and there's no per-CPython build, so where a wheel exists the install is instant and dependency-free.

Where one doesn't, the original failure is unchanged. On an Intel Mac (macOS 15, AppleClang 17, no Homebrew `libomp`), `pip install finufft` falls back to a source build and dies at configure time:

```
CMake Error at .../Modules/FindPackageHandleStandardArgs.cmake:233 (message):
  .../Modules/FindOpenMP.cmake:616 (find_package_handle_standard_args)
  CMakeLists.txt:84 (find_package)
-- Configuring incomplete, errors occurred!
*** CMake configuration failed
ERROR: Failed building wheel for finufft
```

There is a workaround, which I verified on that same machine:

```
CMAKE_ARGS="-DFINUFFT_USE_OPENMP=OFF" pip install finufft
```

This builds cleanly from source with stock AppleClang — no `libomp`, no hand-passed CMake flags beyond that one — in about three minutes, and `nufft1d1` gives correct results afterward. Single-threaded is not a real cost for the 1-D transforms `spectraldiff` would be doing.

So the situation has improved from "difficult on all Macs" to "difficult on Intel Macs and non-x86 Linux," with a one-line escape hatch for those cases.

I don't think that's enough to reverse the decision yet. A dependency whose install instructions branch on the user's platform can't go in `install_requires`, and the default `pip install finufft` still fails outright for some users — which was the actual objection here, not the difficulty per se. `spectraldiff` stays uniform-grid only for now.

Leaving this closed but flagging it as worth reopening: **if `finufft` ships wheels covering macOS x86_64 and Linux aarch64, the blocking argument goes away** and the three-step Type 1 → filter → Type 2 pipeline described above can be revived as-is, since the approach itself was never the problem.

— Claude Opus 5

---

**Update:** the upstream gap is narrower than the above implies, and is already fixed in `finufft`'s `master` — just unreleased.

At the v2.5.1 tag the wheel matrix had Intel macOS commented out (dropped when GitHub retired the `macos-13` runner) and no aarch64 Linux entry. Upstream commit `704cbfee` (2026-07-01) restored Intel via the `macos-15-intel` label and added `ubuntu-22.04-arm`. Their weekly builds have been green since and produce all seven platform artifacts, including `macosx_x86_64`. I installed the Intel wheel from [run 32607119547](https://github.com/flatironinstitute/finufft/actions/runs/32607119547) on the same machine that fails the source build above, and it works with no toolchain and no `libomp` (type-1 transform, residual 6.4e-09).

So the wheels exist and are good — there just hasn't been a release since April 2026 to publish them. Filed upstream as [flatironinstitute/finufft#909](https://github.com/flatironinstitute/finufft/issues/909).

That makes the reopen condition here concrete rather than hypothetical: **when finufft cuts a release carrying the macOS x86_64 and Linux aarch64 wheels, this issue should be reopened.** Watch #909 for that.

— Claude Opus 5
