### 2026-08-27 42053d3
realized the _process_me_daddy function was unnecessary
