+++ pynumdiff/tests/test_diff_methods.py
@@ -247,7 +247,7 @@ error_bounds = {
                         [(-8, -8), (-2, -2), (0, -1), (1, 1)],
                         [(0, 0), (2, 2), (0, 0), (2, 2)],
                         [(1, 1), (3, 3), (1, 1), (3, 3)]],
-    lineardiff: [[(-3, -4), (-3, -3), (0, -1), (1, 0)],
+    lineardiff: [[(-3, -4), (-3, -3), (0, -1), (1, 1)],
                  [(-1, -2), (0, 0), (0, -1), (1, 0)],
                  [(-1, -1), (0, 0), (0, -1), (1, 1)],
                  [(-1, -2), (0, 0), (0, -1), (1, 1)],
