### 2026-08-28 4ad25ee
Normalize robustdiff by its own noise scale, addressing #222
Differentiation is linear, so f(a*x) should be a*f(x), but a hyperparameter carrying absolute units breaks that quietly: robustdiff's log_q and log_r were process and measurement variances in whatever units the data arrived in, so rescaling reweighted the prior against the measurements and the deviation sat at 4.06e-01. Solving in units of the vector's own noise, estimated by MAD on second differences, brings that to 2.92e-13 and makes one setting carry across datasets, which is the same fix #218 applied to splinediff's `s`. Centerin
[…]
