+++ pynumdiff/tests/test_diff_methods.py
@@ -56,8 +56,8 @@ diff_methods_and_params = [
     (constant_acceleration, {'r':1e-3, 'q':1e4}), (constant_acceleration, [1e-3, 1e4]),
     (constant_jerk, {'r':1e-4, 'q':1e5}), (constant_jerk, [1e-4, 1e5]),
     (rtsdiff, {'order':2, 'log_qr_ratio':7, 'forwardbackward':True}),
-    (robustdiff, {'order':3, 'log_q':7, 'log_r':2}),
-    (robust_irreg_step, {'order':3, 'log_q':7, 'log_r':2}),
+    (robustdiff, {'order':3, 'log_q':9, 'log_r':0}),
+    (robust_irreg_step, {'order':3, 'log_q':9, 'log_r':0}),
     (velocity, {'gamma':0.5}), (velocity, [0.5]),
     (acceleration, {'gamma':1}), (acceleration, [1]),
     (jerk, {'gamma':10}), (jerk, [10]),
@@ -238,15 +238,15 @@ error_bounds = {
     robustdiff: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                  [(flr, flr), (flr, flr), (0, -1), (1, 1)],
                  [(flr, flr), (flr, flr), (0, -1), (1, 1)],
-                 [(-7, -7), (-2, -2), (0, -1), (1, 1)],
-                 [(0, 0), (2, 2), (0, 0), (2, 2)],
-                 [(1, 1), (3, 3), (1, 1), (3, 3)]],
+                 [(flr, flr), (-2, -2), (0, -1), (1, 1)],
+                 [(-11, -11), (1, 1), (0, 0), (1, 1)],
+                 [(0, 0), (3, 3), (0, 0), (3, 2)]],
     robust_irreg_step: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                         [(flr, flr), (flr, flr), (0, -1), (1, 1)],
                         [(flr, flr), (flr, flr), (0, -1), (1, 1)],
-                        [(-8, -8), (-2, -2), (0, -1), (1, 1)],
-                        [(0, 0), (2, 2), (0, 0), (2, 2)],
-                        [(1, 1), (3, 3), (1, 1), (3, 3)]],
+                        [(flr, flr), (-2, -2), (0, -1), (1, 1)],
+                        [(-11, -11), (1, 1), (0, 0), (1, 1)],
+                        [(1, 1), (3, 2), (0, 0), (2, 2)]],
     lineardiff: [[(-3, -4), (-3, -3), (0, -1), (1, 0)],
                  [(-1, -2), (0, 0), (0, -1), (1, 0)],
                  [(-1, -1), (0, 0), (0, -1), (1, 1)],
@@ -254,6 +254,7 @@ error_bounds = {
                  [(0, 0), (2, 1), (0, 0), (2, 1)],
                  [(0, -1), (3, 3), (0, 0), (3, 3)]]
 }
+
 
 # Essentially run the cartesian product of [diff methods] x [test functions] through this one test
 @mark.filterwarnings("ignore::DeprecationWarning") # I want to test the old and new functionality intentionally
@@ -273,12 +274,12 @@ def test_diff_method(diff_method_and_params, test_func_and_deriv, request): # re
     x_noisy = x + noise
 
     # differentiate without and with noise, accounting for new and old styles of calling functions
-    x_hat, dxdt_hat = diff_method(x, _t, **params) if isinstance(params, dict) \
-        else diff_method(x, _t, params) if (isinstance(params, list) and len(diff_method_and_params) < 3) \
-        else diff_method(x, _t, params, options)
[…]
