### 2026-08-28 florisvb/PyNumDiff#223 @pavelkomarov
Rerun the relevant notebooks when this is done, especially the 2nd because optimization is presently not attempted for `lineardiff` due to complications.
