### 2026-08-28 a85f03f
Fix waveletdiff NaN and equivariance test, guard cvxpy warnings, rerun notebooks
Repairs the scale-equivariance assertion that reddened master by dividing the scaled derivative by `a` so both directions are held to the same relative tolerance rather than one being judged a thousand times more strictly, stops `waveletdiff` returning NaN at `threshold=0` where pywt's soft threshold evaluates `data*max(1 - value/|data|, 0)` as 0/0 for an exactly-zero coefficient (PyWavelets/pywt#866), decorates the three cvxpy entry points with `np.errstate` to swallow the spurious reduce warnings CVXPY emits whi
[…]
