+++ pynumdiff/kalman_smooth.py
@@ -356,6 +356,10 @@ def robustdiff(x, dt_or_t, order, log_q, log_r, proc_huberM=6, meas_huberM=0, ax
     return x_hat, dxdt_hat
 
 
+
+
+@np.errstate(invalid='ignore', over='ignore') # cvxpy#3503: canonicalizing norm1/huber/tv builds sum atoms, which reduce over uninitialized
+# memory just to read off a shape, so they warn when it holds garbage. This wall can come down if they ever fix it upstream.
 def convex_smooth(y, A, Q, C, R, B=None, u=None, proc_huberM=6, meas_huberM=0):
     """Solve the optimization problem for robust smoothing using CVXPY.
 
