### 2026-08-28 florisvb/PyNumDiff#222 @pavelkomarov
See also https://github.com/florisvb/PyNumDiff/commit/a85f03f2bc095249965032fbbd4862cc92af51ed, which applies a fix and reruns notebooks based on this change.
