### 2026-08-28 eb523da
Silence optimizer-internal warnings, fix self-inflicted even windows, chunk notebook 4 finer
Broadens the optimizer's existing warning suppression from UserWarning to Warning in both the parent filter and the Pool initializer, since the search visits bad parameterizations on purpose and scores whatever comes back, so complaints from underneath are noise, and a parent-side filter provably cannot reach the workers where most of them are raised; fixes savgoldiff's and polydiff's clamps to preserve odd widths, because `degree+1` and `degree*3+1` both land even at odd degree, so the methods were ma
[…]
