+++ pynumdiff/polynomial_fit.py
@@ -97,7 +97,7 @@ def polydiff(x, dt_or_t, params=None, options=None, degree=None, window_size=Non
 
     if window_size:
         if window_size < degree*3:
-            window_size = degree*3+1
+            window_size = degree*3 + 1 + degree%2 # parity term to keep this odd
         if window_size % 2 == 0:
             window_size += 1
             warn("Kernel window size should be odd. Added 1 to length.")
@@ -158,7 +158,7 @@ def savgoldiff(x, dt, params=None, options=None, degree=None, window_size=None, 
     if np.any(np.isnan(x)): raise ValueError("`x` may not contain NaN. Missing values spread through the filter.")
     if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. Savitzky-Golay assumes fixed-width windows with uniform sampling.")
 
-    window_size = np.clip(window_size, degree + 1, x.shape[axis]-1)
+    window_size = np.clip(window_size, degree + 1 + degree%2, x.shape[axis] - 1 + x.shape[axis]%2) # returns odd numbers
     if window_size % 2 == 0:
         window_size += 1 # window_size needs to be odd
         warn("Kernel window size should be odd. Added 1 to length.")
