### 2026-08-28 #11172 @pavelkomarov
Thanks for caring enough to do a review. Noticing and surfacing so many bugs really has helped improve the code, and it got me to revisit and rethink a few things too. Improvements, as remembered by me, the human (hi, hello), not Claude this time:

- more reliable install via a complete list of dependencies, to which I added some min version requirements in cases like `cvxpy` where I really am depending on recently-added features
- no more swallowing warnings forever in `optimize()`, which led directly to discovering a whole bundle of subtle numerical issues, like the following three:
- overflow in sum(huber()) functions, solved by normalizing residuals instead of scaling cutoffs (https://github.com/florisvb/PyNumDiff/issues/217)
- scaled values and improved initialization in the optimization problem in the huber branch of `estimate_integration_constant` so it solves more stably and correctly regardless of data scale (https://github.com/florisvb/PyNumDiff/issues/220), and added a test to ratchet down on this better accuracy
- `splinediff`'s `s` parameter was sensitive to data scale (https://github.com/florisvb/PyNumDiff/issues/218), which is new since it started depending on `make_splrep` vs the old `UnivariateSpline`, which did some cleverness; it is now once again relative so it works at any data scale, which led to:
- a test that methods retain their accuracy as data scales, since `f(a*x) = a*f(x)` mathematically (https://github.com/florisvb/PyNumDiff/issues/222)
- guards in `optimize()` so it doesn't crash on the edge case of all-singletons-and-categoricals
- a test to ensure there are hash collisions across parallel calls given the same query point, and a slight simplification of the key used as identifier
- a recognition that many window widths really shouldn't be returned as even numbers by the optimizer if they're just corrected by +1 with a warning internally, and that some which were allowed to be even really should be locked to odd because otherwise they silently degrade results (in a sawtooth way that's difficult for the optimization to navigate), which led to optimizer improvements to constrain/express this (https://github.com/florisvb/PyNumDiff/issues/219)
- all notebooks rerun to capture latest improvements, which surfaced more warnings and led me to chase down cvxpy userwarnings and spline fit runtimewarnings, consider their importance or lack thereof, catch/silence the useless ones, and open issues on upstream packages (https://github.com/cvxpy/cvxpy/issues/3503, https://github.com/PyWavelets/pywt/issues/866)
- restructuring of the parallelization in the expensive-to-run 4th notebook in order to make better use of cpu cores
- `waveletdiff` added to notebooks and given a default search space entry in `optimize.py`
- `robustdiff`'s params also made invariant to data scale so it passes the new test, and in the process realized:
- `robustdiff`'s annoying, spiritually-redundant hyperparameter search space (due to coupling such that at fixed huber knees there are invariant characteristic curves in q,r space along which the cost function will have constant value) could in fact now be somewhat deduplicated because `log_r` has a natural set point at 0 (which essentially takes at face value the robust noise estimate by which we now normalize data prior to convex solve), and therefore it is sufficient to sweep primarily along `log_q` (which is more sharply transverse to the characteristic curves), thereby greatly reducing the number of initial seeds and therefore optimization runtime without hurting performance! (https://github.com/florisvb/PyNumDiff/issues/222, super proud of this one, because I spent ages thinking it through in the past but never cracked it, and now all the notebooks run way faster!)
- windows are now not allowed to skip completely past each other or leave the tail of an array uncovered in `slide_function`
- simplified mode killing in `spectraldiff` to avoid hard-to-see indexing issues that caused unexpected behavior at an edge and broke up a conjugate pair
- fixed aliasing bug in circular domains for `rtsdiff`
- `savgoldiff` now properly uses data length along the correct `axis`
- use second-order sections in `butterdiff` for stability
- default to CLARABEL in `lineardiff` so it doesn't immediately crash, because OSQP isn't right for it; opened https://github.com/florisvb/PyNumDiff/issues/223 and a branch to beat this method into shape further
- put tight bound on window size relative to degree in `polydiff` to avert underdetermined least squares solves
- methods now disallow missing values or variable timesteps with valueerrors explaining why they can't work, and the readme now has a table of who can do what

I also revisited https://github.com/florisvb/PyNumDiff/issues/201 about using the nonuniform FFT to allow `spectraldiff` to accept variable step sizes and realized it's not as hopeless as I'd thought, so opened an issue on the upstream Flatiron Institute repo (https://github.com/flatironinstitute/finufft/issues/909) and reopened my downstream issue.

I'm presently rerunning the big notebook 4 experiment, because reviewers on [the Taxonomy paper](https://arxiv.org/abs/2512.09090) where we make use of `PyNumDiff` to compare methods head-to-head, justifiably asked about version information, and it's possible this massive series of edits will have caused a little drift, but the flurry of activity is drawing to conclusion, a stable state. I'll probably release v0.2.4 shortly, with v0.3 to follow some months later when deprecation can come to fruition (https://github.com/florisvb/PyNumDiff/issues/183).
