### 2026-08-28 56167d2
claude playing more with lineardiff
