+++ pynumdiff/linear_model.py
@@ -29,63 +29,68 @@ def spectraldiff(*args, **kwargs): # pragma: no cover pylint: disable=missing-fu
     return _spectraldiff(*args, **kwargs)
 
 
+_PROBLEM_CACHE = {} # (order, window length) -> a parametrized CVXPY problem, so identically shaped windows reuse it
+
+def _basis(order, N):
+    """Powers of nondimensional time, the shape the integration constants take after being integrated `order` times."""
+    tau = np.arange(N)/N # dtau is 1/N, so tau spans [0, 1) whatever the window's width in real time
+    return np.vstack([tau**(order-n-1)/math.factorial(order-n-1) for n in range(order)])
+
 @np.errstate(invalid='ignore', over='ignore') # cvxpy#3503: canonicalizing norm1/huber/tv builds sum atoms, which reduce over uninitialized
 #  memory just to read off a shape, so they warn when it holds garbage. This wall can come down if they ever fix it upstream.
-def _solve_for_A_and_C_given_X_and_Xdot(X, Xdot, num_integrations, dt, gammaC=1e-1, gammaA=1e-6,
-                                           solver='CLARABEL', A_known=None, epsilon=1e-6):
+def _cached_problem(order, N):
+    """Build, or reuse, the convex program for one window. Canonicalizing costs ~85% of a solve here, and depends only on
+    the shapes, so hoisting it out of the window loop is worth about 7x. `gamma` is a Parameter too, which keeps a cached
+    problem valid across a whole `optimize` sweep rather than only within one call. Requires the problem to be DPP."""
+    if (order, N) not in _PROBLEM_CACHE:
+        if len(_PROBLEM_CACHE) >= 64: _PROBLEM_CACHE.clear() # bound what an optimizer's window sweep can accumulate
+        A = cvxpy.Variable((order, order)); C = cvxpy.Variable((order, order))
+        X = cvxpy.Parameter((order, N)); Xdot = cvxpy.Parameter((order, N)); gamma = cvxpy.Parameter(nonneg=True)
+        Csum = cvxpy.vstack([C[i, :] @ _basis(order, N) for i in range(order)])
+        prob = cvxpy.Problem(cvxpy.Minimize(cvxpy.sum_squares(Xdot - (A @ X + Csum)) +
+            gamma*cvxpy.sum(cvxpy.abs(C)) + 1e-6*cvxpy.sum(cvxpy.abs(A)))) # gammaA is small and fixed; gammaC is the knob
+        _PROBLEM_CACHE[(order, N)] = (prob, A, C, X, Xdot, gamma)
+    return _PROBLEM_CACHE[(order, N)]
+
+@np.errstate(invalid='ignore', over='ignore') # see the note on _cached_problem
+def _solve_for_A_and_C_given_X_and_Xdot(X, Xdot, order, gamma, solver='CLARABEL'):
     """Given state and the derivative, find the system evolution and measurement matrices."""
-    # Set up the variables
-    A = cvxpy.Variable((X.shape[0], X.shape[0]))
-    C = cvxpy.Variable((X.shape[0], num_integrations))
-
-    # Integrate the integration constants
-    Csum = 0
-    t = np.arange(0, X.shape[1])*dt
-    for n in range(num_integrations):
-        C_subscript = n
-        t_exponent = num_integrations - n -1
-        den = math.factorial(t_exponent)
-        Cn = cvxpy.vstack((1/den*C[i, C_subscript]*t**t_exponent for i in range(X.shape[0])))
-        Csum = Csum + Cn
-
-    # Define the objective function
-    error = cvxpy.sum_squares(Xdot - ( cvxpy.matmul(A, X) + Csum))
[…]
