### 2026-08-28 1ba8139
simplified branching structure of optimize, realizing I was repeating myself
