### 2026-08-29 eaed527
vibe coding
