+++ pynumdiff/linear_model.py
@@ -31,46 +31,6 @@ def spectraldiff(*args, **kwargs): # pragma: no cover pylint: disable=missing-fu
 
 _PROBLEM_CACHE = {} # (order, window length) -> a parametrized CVXPY problem, so identically shaped windows reuse it
 
-def _basis(order, N):
-    """Powers of nondimensional time, the shape the integration constants take after being integrated `order` times."""
-    tau = np.arange(N)/N # dtau is 1/N, so tau spans [0, 1) whatever the window's width in real time
-    return np.vstack([tau**(order-n-1)/math.factorial(order-n-1) for n in range(order)])
-
-@np.errstate(invalid='ignore', over='ignore') # cvxpy#3503: canonicalizing norm1/huber/tv builds sum atoms, which reduce over uninitialized
-#  memory just to read off a shape, so they warn when it holds garbage. This wall can come down if they ever fix it upstream.
-def _cached_problem(order, N):
-    """Build, or reuse, the convex program for one window. Canonicalizing costs ~85% of a solve here, and depends only on
-    the shapes, so hoisting it out of the window loop is worth about 7x. `gamma` is a Parameter too, which keeps a cached
-    problem valid across a whole `optimize` sweep rather than only within one call. Requires the problem to be DPP."""
-    if (order, N) not in _PROBLEM_CACHE:
-        if len(_PROBLEM_CACHE) >= 64: _PROBLEM_CACHE.clear() # bound what an optimizer's window sweep can accumulate
-        A = cvxpy.Variable((order, order)); C = cvxpy.Variable((order, order))
-        X = cvxpy.Parameter((order, N)); Xdot = cvxpy.Parameter((order, N)); gamma = cvxpy.Parameter(nonneg=True)
-        Csum = cvxpy.vstack([C[i, :] @ _basis(order, N) for i in range(order)])
-        prob = cvxpy.Problem(cvxpy.Minimize(cvxpy.sum_squares(Xdot - (A @ X + Csum)) +
-            gamma*cvxpy.sum(cvxpy.abs(C)) + 1e-6*cvxpy.sum(cvxpy.abs(A)))) # gammaA is small and fixed; gammaC is the knob
-        _PROBLEM_CACHE[(order, N)] = (prob, A, C, X, Xdot, gamma)
-    return _PROBLEM_CACHE[(order, N)]
-
-@np.errstate(invalid='ignore', over='ignore') # see the note on _cached_problem
-def _solve_for_A_and_C_given_X_and_Xdot(X, Xdot, order, gamma, solver='CLARABEL'):
-    """Given state and the derivative, find the system evolution and measurement matrices."""
-    prob, A, C, X_p, Xdot_p, gamma_p = _cached_problem(order, X.shape[1])
-    X_p.value = X; Xdot_p.value = Xdot; gamma_p.value = gamma
-    # Tighter than CLARABEL's defaults, because an l1 penalty's kinks make the argmin jump when a coordinate crosses
-    # zero, and a loosely converged iterate turns a last-bit input difference into a visible one. Costs nothing
-    # measurable on problems this small, and is what lets rescaled input reproduce rescaled output. See #222
-    opts = {'tol_gap_abs':1e-12, 'tol_gap_rel':1e-12, 'tol_feas':1e-12} if solver == 'CLARABEL' else {}
-    try:
-        prob.solve(solver=solver, **opts)
-        if A.value is not None and C.value is not None: return A.value, C.value
-    except cvxpy.error.SolverError: pass
-    # The solve failed, which happens on a few percent of parameter combinations and would otherwise abort a whole
-    # optimize() sweep. Fall back to the unregularized fit, the gamma -> 0 limit of the same model: every row of Xdot
[…]
