### 2026-08-29 d87d573
still vibe coding lineardiff
