### 2026-08-29 f2e0487
Let window_size=None reach the global fit, and tighten lineardiff's multidim test
polydiff and lineardiff both documented window_size=None as "no sliding, fit globally" but rejected it in their required-argument checks, leaving the branch reachable only through the deprecated options={'sliding': False} route; dropping window_size from those checks makes the docstrings true and the two methods agree. That path has a longer history than it looks: the original dispatch stripped window_size off params only inside the sliding branch, so lineardiff's global fit raised ValueError from the first commi
[…]
