### 2026-08-29 fee1c46
Make waveletdiff's derivative an FFT, and stop integer input truncating derivatives
Phi and Phi_prime are both circulant, so the DFT diagonalizes them and Phi_prime @ Phi^-1 becomes multiplication by the single transfer function fft(phi')/fft(phi), replacing a sparse LU on a (3N-2)-square matrix with an FFT pair: 20-41x faster and identical to 5e-13. Five methods allocated outputs with np.empty_like, inheriting an integer input's dtype and silently truncating derivatives, and the five taking dt_or_t now require it strictly increasing rather than accepting reversed or repeated timestamps. polyd
[…]
