+++ pynumdiff/tests/test_diff_methods.py
@@ -44,8 +44,8 @@ diff_methods_and_params = [
     (butterdiff, {'filter_order':3, 'cutoff_freq':0.7}), (butterdiff, [3, 0.7]),
     (first_order, {}), (second_order, {}), (fourth_order, {}), # empty dictionary for the case of no parameters
     (iterated_second_order, {'num_iterations':5}), (iterated_fourth_order, {'num_iterations':10}),
-    (polydiff, {'degree':2, 'window_size':3}), (polydiff, [2, 3]),
-    (polydiff_irreg_step, {'degree':2, 'window_size':3}),
+    (polydiff, {'degree':5, 'window_size':15}), (polydiff, [5, 15]),
+    (polydiff_irreg_step, {'degree':5, 'window_size':15}),
     (savgoldiff, {'degree':2, 'window_size':5, 'smoothing_win':5}), (savgoldiff, [2, 5, 5]),
     (splinediff, {'degree':5, 's':2}), (splinediff, [5, 2]),
     (spline_irreg_step, {'degree':5, 's':2}),
@@ -136,15 +136,15 @@ error_bounds = {
     polydiff: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                [(flr, flr), (flr, flr), (0, -1), (1, 1)],
                [(flr, flr), (flr, flr), (0, -1), (1, 1)],
-               [(-2, -2), (0, 0), (0, -1), (1, 1)],
-               [(0, 0), (1, 1), (0, -1), (1, 1)],
-               [(0, 0), (3, 3), (0, 0), (3, 3)]],
+               [(-3, -3), (-1, -1), (0, -1), (1, 1)],
+               [(-1, -1), (1, 1), (0, -1), (1, 1)],
+               [(0, 0), (2, 2), (0, 0), (2, 2)]],
     polydiff_irreg_step: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                           [(flr, flr), (flr, flr), (0, -1), (1, 1)],
                           [(flr, flr), (flr, flr), (0, -1), (1, 1)],
-                          [(-2, -2), (0, 0), (0, -1), (1, 1)],
-                          [(0, 0), (1, 1), (0, 0), (1, 1)],
-                          [(0, 0), (3, 3), (0, 0), (3, 3)]],
+                          [(-3, -3), (-1, -1), (0, -1), (1, 1)],
+                          [(-1, -1), (1, 1), (0, -1), (1, 1)],
+                          [(0, 0), (2, 2), (0, 0), (2, 2)]],
     savgoldiff: [[(-11, flr), (flr, flr), (0, -1), (0, 0)],
                  [(-11, flr), (flr, flr), (0, -1), (0, 0)],
                  [(-2, -2), (-1, -1), (0, -1), (0, 0)],
