### 2026-08-31 03bc331
Expose cutoff_freq instead of tvgamma, and unify robust scale estimation
Addresses #225 by making `optimize` take the quantity a user can reason about. `tvgamma` weighs smoothness against fidelity but its meaning depends on the sampling rate as well as the signal, so callers now give `cutoff_freq`, the highest frequency at which their data carries real signal, and the weight is derived from it internally exactly as `suggest_method` was already doing; `tvgamma` still works with a DeprecationWarning. Automatic estimation of that cutoff was tried and abandoned: a power percentile of the noisy spe
[…]
