+++ pynumdiff/kalman_smooth.py
@@ -2,11 +2,10 @@
 from warnings import warn
 import numpy as np
 from scipy.linalg import expm, sqrtm
-from scipy.stats import median_abs_deviation
 try: import cvxpy
 except ImportError: pass
 
-from pynumdiff.utils.utility import huber_const
+from pynumdiff.utils.utility import huber_const, robust_noise_scale
 
 
 def kalman_filter(y, xhat0, P0, A, Q, C, R, B=None, u=None, save_P=True, innovation_fn=None):
@@ -124,7 +123,7 @@ def rtsdiff(x, dt_or_t, order, log_qr_ratio, forwardbackward=False, axis=0, circ
     N = x.shape[axis]
     if not np.isscalar(dt_or_t):
         if N != len(dt_or_t): raise ValueError("If `dt_or_t` is given as array-like, must have same length as x along `axis`.")
-        if np.any(np.diff(dt_or_t) <= 0): raise ValueError("`dt_or_t` must be strictly increasing. Out-of-order or repeated"
+        if np.any(np.diff(dt_or_t) <= 0): raise ValueError("`dt_or_t` must be strictly increasing. Out-of-order or repeated "
             "sample locations make neighbor differences and windows meaningless.")
 
     q = 10**int(log_qr_ratio/2) # even-ish split of the powers across 0
@@ -316,20 +315,20 @@ def robustdiff(x, dt_or_t, order, log_q, log_r, proc_huberM=6, meas_huberM=0, ax
     N = x.shape[axis]
     if not np.isscalar(dt_or_t):
         if N != len(dt_or_t): raise ValueError("If `dt_or_t` is given as array-like, must have same length as `x` along `axis`.")
-        if np.any(np.diff(dt_or_t) <= 0): raise ValueError("`dt_or_t` must be strictly increasing. Out-of-order or repeated"
+        if np.any(np.diff(dt_or_t) <= 0): raise ValueError("`dt_or_t` must be strictly increasing. Out-of-order or repeated "
             "sample locations make neighbor differences and windows meaningless.")
 
     A_c = np.diag(np.ones(order), 1) # continuous-time A just has 1s on the first diagonal (where 0th is main diagonal)
     Q_c = np.zeros(A_c.shape); Q_c[-1,-1] = 10**log_q # continuous-time uncertainty around the last derivative
     C = np.zeros((1, order+1)); C[0,0] = 1 # we measure only y = noisy x
     R = np.array([[10**log_r]]) # 1 observed state, so this is 1x1
-    M = np.block([[A_c, Q_c], [np.zeros(A_c.shape), -A_c.T]])  # exponentiate per step
+    M = np.block([[A_c, Q_c], [np.zeros(A_c.shape), -A_c.T]]) # exponentiate per step
 
     if np.isscalar(dt_or_t): # convert to discrete time using matrix exponential
         eM = expm(M * dt_or_t)
         A_d = eM[:order+1, :order+1]
         Q_d = eM[:order+1, order+1:] @ A_d.T
-        if np.linalg.cond(Q_d) > 1e12: Q_d += np.eye(order+1)*1e-12 # for numerical stability with convex solver. Doesn't change answers appreciably (or at all).
+        if np.linalg.cond(Q_d) > 1e12: Q_d += np.eye(order+1)*np.linalg.eigvalsh(Q_d)[-1]*1e-12 # for numerical stability. Doesn't change answers appreciably (or at all).
     else: # support variable step size for this function
         A_d = np.empty((N-1, order+1, order+1))
         Q_d = np.empty_like(A_d)
[…]
