### 2026-09-01 33aa9db
correcting MAP to be maximum a posteriori rather than priori. I goofed big. How embarrassing.
