### 2026-09-01 80d3aea
removed deprecated tvgamma from optimize signature
