+++ pynumdiff/optimize.py
@@ -62,10 +62,10 @@ method_params_and_bounds = {
                  # `level` is left at its adaptive default, min(dwt_max_level(N, wavelet), 5), which tracks both the signal and filter lengths
                 {'threshold': (0.1, 10)}),
     tvrdiff: ({'gamma': [1e-2, 1e-1, 1, 10, 100, 1000],
-               'order': {1, 2, 3}, # warning: order 1 hacks the loss function when tvgamma is used, tends to win but is usually suboptimal choice in terms of true RMSE
+               'order': {1, 2, 3}, # warning: order 1 hacks the ground-truth-less loss function, tends to win but is usually suboptimal choice in terms of true RMSE
               'huberM': [2., 6]}, # the scale of sigma is mad(x), which is bigger than mad(y-x) residuals, so outliers likely come at lower M values
               {'gamma': (1e-4, 1e7),
-              'huberM': (2, 6)}), # huberM too low seeks sparse solutions, which hack the tvgamma loss function
+              'huberM': (2, 6)}), # huberM too low seeks sparse solutions, which hack the ground-truth-less loss function
     iterative_velocity: ({'scale': 'small', # Rare to optimize this one, because it's longer-running than convex version
                  'num_iterations': [1, 5, 10],
                           'gamma': [1e-2, 1e-1, 1, 10, 100, 1000]},
@@ -79,7 +79,7 @@ method_params_and_bounds = {
                          'order': {1, 2, 3}, # for this few options, the optimization works better if this is categorical
                   'log_qr_ratio': [float(k) for k in range(-9, 10, 2)] + [12, 16]},
                  {'log_qr_ratio': (-10, 20)}), # qr_ratio is usually >>1
-    robustdiff: ({'order': {1, 2, 3}, # warning: order 1 hacks the loss function when tvgamma is used, tends to win but is usually suboptimal choice in terms of true RMSE
+    robustdiff: ({'order': {1, 2, 3}, # warning: order 1 hacks the ground-truth-less loss function, tends to win but is usually suboptimal choice in terms of true RMSE
                   'log_q': [1., 4, 7, 10, 13], # decimal after first entry ensure this is treated as float type
                   'log_r': [0.], # one seed, but allowed to drift. Holding both Huber Ms fixed, the objective is flat along characteristic curves in the
                                  # (log_q, log_r) plane, but these are sloped such that varying log_q cuts across more of them than varying log_r; extra log_r
@@ -153,7 +153,7 @@ def _objective_function(point, func, x, dt, singleton_params, categorical_params
 
 
 def optimize(func, x, dt, dxdt_truth=None, cutoff_freq=None, search_space_updates={}, metric='rmse',
-    padding=0, opt_method='Nelder-Mead', maxiter=10, parallel=True, huberM=6, tvgamma=None):
+    padding=0, opt_method='Nelder-Mead', maxiter=10, parallel=True, huberM=6):
     """Find the optimal hyperparameters for a given differentiation method.
 
     :param function func: differentiation method to optimize parameters for, e.g. kalman_smooth.rtsdiff
@@ -162,8 +162,6 @@ def optimize(func, x, dt, dxdt_truth=None, cutoff_freq=None, search_space_update
     :param np.array[float] dxdt_truth: actual time series of the derivative of x, if known
     :param float cutoff_freq: Only used if :code:`dxdt_truth` is *not* given. The bandlimit of underlying signal, in Hz.
                     Estimate by counting peaks per second, or by reading where the power spectrum falls off. Lower -> smoother
-    :param float tvgamma: (**deprecated**, prefer :code:`cutoff_freq`) the smoothness weight itself, which is harder to
[…]
