### 2026-09-01 #11172 @pavelkomarov
I've done a couple more since:

- when allocating empty arrays, specify `float` type explicitly so integer type can't be accidentally inherited from inputs, which would truncate results and cause inaccuracies
- independent-variable vectors (`dt_or_t`), for methods that accept irregular steps, must now be sorted with gaps >0 to avert probable bugs/undefined behavior
- math trick applied in `waveletdiff` to exploit FFT and solve for the same thing faster
- to match the slide window kernel in spirit, `polyfit` inside `polydiff` is now given weights `sqrt(kernel)` rather than `kernel`, because it weights the residuals, which are then squared
- `optimize` now exposes more-interpretable `cutoff_freq`, so the mildly complicated calculation of `tvgamma` can happen internally (also involved a param rename in `butterdiff` to avoid a clash and better agree with `spectraldiff`, technically breaking backwards compatibility if you were calling with kwargs. But the method hasn't been around super long in its new form, so I went for it.)
- abstracted (robust) data scale and noise scale estimation to a couple new `utils` methods, for uniformity and standalone testing
- `robustdiff` was adding a `1e-12` ridge to $Q_d$ matrices with high condition numbers, but this was blind to scale of `log_q`. The ridge now takes account of $\lambda_\text{max}$, averting some additional convex solver failures. For remaining failures, the method now raises `LinAlgError` with an informative message rather than silently returning `NaN`s, better-defined behavior which also silences warnings and happens to interact with the optimizer's minimization exactly as before (big penalty, repels solution).
- corrected a few explanations and updated param names in notebooks
- and the big one! @florisvb just merged https://github.com/florisvb/PyNumDiff/pull/226, removing deprecated methods and signatures from the repo 🎉

All this means a v0.3 release is called for soon, so I'll pin the paper to that.
