### 2026-09-01 7adabc6
adjusting clamp to silence one last potential source of userwarnings
