### 2026-09-01 b8abbda
renamed cutoff_freq to bandlimit in the optimizer, because that's a better, more distinct, more accurate name, and renamed high_freq_cutoff to the shorter cutoff_freq for spectral and butter
