+++ pynumdiff/optimize.py
@@ -24,9 +24,9 @@ method_params_and_bounds = {
             {'window_size': (1, 1e6, 'odd'), # an even-width kernel has no center, so convolving with it
           'num_iterations': (1, 100)}),         # shifts the signal half a sample before differencing
     butterdiff: ({'filter_order': set(i for i in range(1,11)), # categorical to save us from doing double work by guessing between orders
-                   'high_freq_cutoff': [0.0001, 0.001, 0.005, 0.01, 0.1, 0.5],
+                   'cutoff_freq': [0.0001, 0.001, 0.005, 0.01, 0.1, 0.5],
                 'num_iterations': [1, 5, 10]},
-                  {'high_freq_cutoff': (1e-4, 1-1e-2),
+                  {'cutoff_freq': (1e-4, 1-1e-2),
                 'num_iterations': (1, 1000)}),
     finitediff: ({'num_iterations': [5, 10, 30, 50],
                            'order': {2, 4}}, # order is categorical here, because it can't be 3
@@ -51,8 +51,8 @@ method_params_and_bounds = {
           'num_iterations': (1, 5)}), # where starving the budget while iterating made scipy's knot search fail
     spectraldiff: ({'even_extension': {True, False}, # give categorical params in a set
                   'pad_to_zero_dxdt': {True, False},
-                  'high_freq_cutoff': [1e-2, 5e-2, 1e-1, 5e-1]}, # give numerical params in a list to scipy.optimize over them
-                 {'high_freq_cutoff': (1e-3, 1-1e-5)}), # every cutoff below ~2/N keeps only the DC term, see #209
+                  'cutoff_freq': [1e-2, 5e-2, 1e-1, 5e-1]}, # give numerical params in a list to scipy.optimize over them
+                 {'cutoff_freq': (1e-3, 1-1e-5)}), # every cutoff below ~2/N keeps only the DC term, see #209
     rbfdiff: ({'sigma': [1e-2, 1e-1, 1],
                 'lmbd': [1e-3, 1e-2, 1e-1]},
               {'sigma': (1e-2, 1e3),
@@ -152,7 +152,7 @@ def _objective_function(point, func, x, dt, singleton_params, categorical_params
         cache[key] = cost; return cost
 
 
-def optimize(func, x, dt, dxdt_truth=None, cutoff_freq=None, search_space_updates={}, metric='rmse',
+def optimize(func, x, dt, dxdt_truth=None, bandlimit=None, search_space_updates={}, metric='rmse',
     padding=0, opt_method='Nelder-Mead', maxiter=10, parallel=True, huberM=6):
     """Find the optimal hyperparameters for a given differentiation method.
 
@@ -160,7 +160,7 @@ def optimize(func, x, dt, dxdt_truth=None, cutoff_freq=None, search_space_update
     :param np.array[float] x: data to differentiate
     :param float dt: step size
     :param np.array[float] dxdt_truth: actual time series of the derivative of x, if known
-    :param float cutoff_freq: Only used if :code:`dxdt_truth` is *not* given. The bandlimit of underlying signal, in Hz.
+    :param float bandlimit: Only used if :code:`dxdt_truth` is *not* given. The bandlimit of underlying signal, in Hz.
                     Estimate by counting peaks per second, or by reading where the power spectrum falls off. Lower -> smoother
     :param dict search_space_updates: Each method has a default search space of parameter settings, encoded as
                     :code:`{param1:[numerical, values], param2:{categorical, values}, param3:value, ...}` (defined at the top of
@@ -184,7 +184,7 @@ def optimize(func, x, dt, dxdt_truth=None, cutoff_freq=None, search_space_update
[…]
