### 2026-09-02 b366872
Update README to clarify deprecated methods
Removed warning about deprecated methods
