### 2026-09-02 63792f1
Revise README for usage and optimization details
Updated usage instructions and optimized parameter settings in README.
