### 2026-09-02 2427fbb
renamed a test in the optimize tests that still referred to cutoff frequency instead of bandlimit, and added friedrichs kernel to lineardiff search space
