+++ pynumdiff/optimize.py
@@ -90,7 +90,7 @@ method_params_and_bounds = {
                   'log_r': (-4, 10), # ignore the measurements and trust the model, and give identical answers
             'proc_huberM': (0, 6),
             'meas_huberM': (0, 6)}),
-    lineardiff: ({'kernel': 'gaussian', # `step_size` barely moves accuracy so now defaults to window_size//5
+    lineardiff: ({'kernel': {'gaussian', 'friedrichs'}, # `step_size` barely moves accuracy so now defaults to window_size//5
                    'order': {2, 3}, # order 1 never won across 12 sim/seed sweeps
                    'gamma': [1e-2, 1e-1, 1], # in units of the data's own scale since #222
              'window_size': [41, 81, 161]},
