### 2026-09-02 #11172 @pavelkomarov
I've given the paper a human touch last night and today.
