### 2026-09-03 florisvb/PyNumDiff#223 @pavelkomarov
## Follow-up: what got tested since the merge

**`gammaA` stays at `1e-6`.** It is a tie-breaker, not a sparsity prior: when `x` is smooth its integrals resemble the polynomials in `B`, so `A` and `C` can fit the same curve. Removing it fails `test_multidimensionality`, which runs `gamma=1e-06` on a noise-free surface. Raising it is worse than useless — `gamma` and `gammaA` are substitutes, so the search just lowers `gamma` to compensate (median 0.0997 → 0.0130), and by `gammaA=1` the fit has collapsed (`sum|A|` 16 → 0.5, RMSE 0.76 → 1.74).

**Unresolved: why `l1` rather than a quadratic.** At the same weight, `sum_squares` fails the multidim Laplacian's worst-case error (0.528 against a 0.1 bound) where `l1` (0.059) and `normNuc` (0.074) pass and `sigma_max` misses by 3%. But at `1e-6` `l1` and `sum_squares` produce nearly the same operator — `sum|A|` 15.97 vs 15.93, RMSE within 0.2% — so the tempting explanation, that a quadratic shrinks everything while `l1` only pins the loose entries, is false at this weight. One borderline assertion on one surface is not a mechanism. Left alone.

**Removed three things.** The `lstsq` fallback fired 0 times in 18,048 solves and was not neutral anyway: it dropped both `l1` penalties, silently solving the `gamma → 0` problem instead of the one asked for. `SolverError` handling is gone everywhere — neither failure channel fires in 12,968 solves including deliberately pathological ones, and a genuine solver failure is worth seeing rather than laundering into a bad score. The `solver` parameter is gone: `lineardiff` pins CLARABEL, while `tvrdiff` lets CVXPY dispatch, which beats any fixed choice (OSQP for the QP forms, CLARABEL for `norm1`, where OSQP is both worse and 10× slower).

**~15% faster.** `robust_data_scale` computes the MAD directly, since scipy's `axis_nan_policy` wrapper costs ~7× the arithmetic below 10⁵ samples and `lineardiff` calls it once per window. The Huber constant fit got an analytic gradient. Warm-starting does nothing — CLARABEL is interior-point.

**The kernel search space is now `{gaussian, friedrichs}`, and the tradeoff is uncomfortable.** Across all 728 configs friedrichs wins 88.9% of cells but true RMSE comes out 0.9% *worse*: it wins the `gamma*TV` term rather than on accuracy. Same loss-hacking already flagged for order-1 `tvrdiff`/`robustdiff`, in milder form. Kept for per-signal adaptivity at 2× the budget.

**The crossfade is untested, not unjustified.** Two passes exist to cancel errors-in-variables bias, as in forward-backward DMD. RMSE cannot see that: over 72 combinations the crossfade lands between the two single passes with neither direction better, but that mixes bias and variance at one operating point. The real test fits a known linear system and compares recovered `A` against truth.

**`lineardiff` is not bit-reproducible.** About 19% of narrow-window fits differ between identical runs — the `l1` kinks make the argmin jump when a coefficient crosses zero. It traces to the penalty on `C`, not `A`. This is why repeated notebook-4 runs move a handful of cells.

---
🤖 Written by [Claude Code](https://claude.com/claude-code) on behalf of @pavelkomarov
