### 2026-09-03 7b37a21
Drop the solver knobs and tidy lineardiff's convex fit
The `solver` parameter is gone from `lineardiff`, `tvrdiff`, and `smooth_acceleration`: the first pins CLARABEL, the other two let CVXPY dispatch, which beats any fixed choice. SolverError is no longer caught anywhere, so a real solver failure surfaces rather than being laundered into a merely-bad score, and `lineardiff`'s unregularized `lstsq` fallback is replaced by a raised LinAlgError, since it never fired and solved a different problem than the caller asked for when it did. `robust_data_scale` computes the MAD directly instead of thro
[…]
