+++ pynumdiff/tests/test_diff_methods.py
@@ -211,7 +211,7 @@ error_bounds = {
                          [(2, 1), (2, 2), (2, 1), (2, 2)],
                          [(1, 1), (3, 3), (1, 1), (3, 3)]],
     smooth_acceleration: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
-                          [(-10, -10), (-10, -10), (-1, -1), (-1, -1)],
+                          [(-10, -11), (-10, -10), (-1, -1), (-1, -1)],
                           [(-2, -2), (-1, -1), (-1, -1), (0, -1)],
                           [(0, 0), (1, 0), (0, -1), (1, 0)],
                           [(1, 1), (2, 2), (1, 1), (2, 2)],
