+++ pynumdiff/linear_model.py
@@ -12,7 +12,7 @@ except ImportError: pass
 
 _PROBLEM_CACHE = {} # (order, window length) -> a parametrized CVXPY problem, so identically shaped windows reuse it
 
-def lineardiff(x, dt, order, gamma, window_size=None, step_size=None, kernel='friedrichs', solver='CLARABEL', axis=0):
+def lineardiff(x, dt, order, gamma, window_size=None, step_size=None, kernel='friedrichs', axis=0):
     """Fit a linear dynamical system to windows of the data, then differentiate that model.
 
     :param np.array[float] x: data to differentiate. May be multidimensional; see :code:`axis`.
@@ -25,8 +25,6 @@ def lineardiff(x, dt, order, gamma, window_size=None, step_size=None, kernel='fr
             ratio, not an absolute stride: a fifth costs a few percent of accuracy against a much finer stride while
             running several times faster, and strides past about half the window degrade badly
     :param str kernel: name of kernel to use for weighting and smoothing windows ('gaussian' or 'friedrichs')
-    :param str solver: CVXPY solver to use, one of :code:`cvxpy.installed_solvers()`. CLARABEL converges reliably,
-            but OSQP stalls on the badly-scaled subproblems short windows produce, returning half-converged iterates
     :param int axis: axis along which to differentiate (default 0)
 
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
@@ -36,16 +34,14 @@ def lineardiff(x, dt, order, gamma, window_size=None, step_size=None, kernel='fr
     if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. The integrals of x are accumulated at a constant step.")
 
     @np.errstate(invalid='ignore', over='ignore') # cvxpy#3503: building a sum atom reduces over uninitialized memory
-    def _lineardiff(x, dt, order, gamma, solver=None): # just to read a shape, so it warns when that memory holds garbage
+    def _lineardiff(x, dt, order, gamma): # just to read a shape, so it warns when that memory holds garbage
         """Fit X = A*integral_X + C*B, then differentiate it to Xdot = A*X + C*dB to get the derivative"""
         mean = np.mean(x)
         x = x - mean
 
-        # Work in nondimensional time tau = t/T, so the window spans [0, 1] no matter how wide it is. Each row of the
+        # Work in nondimensional time tau = t/T, so the normalized window spans tau \in [0, 1]. Each row of the
         # matrix below is one more integration than the row beneath it, so in raw time the rows differ by powers of the
-        # window duration T: at dt=0.01 and window_size=11 that leaves cond(integral_X) above 1e3 and entries around
-        # 1e-5, small enough that a solver's absolute tolerances stop meaning anything. Integrating in tau instead
-        # holds every row at O(1) and cond near 30 regardless of window_size. A comes back in units of 1/tau, so the
+        # window duration T. Integrating in tau instead holds cond number reasonable. A is fit in units of 1/tau, so the
         # derivative it reconstructs is d/dtau and gets divided by T at the end to return to d/dt.
         T = len(x)*dt
         dtau = dt/T
@@ -73,28 +69,22 @@ def lineardiff(x, dt, order, gamma, window_size=None, step_size=None, kernel='fr
             X_p = cvxpy.Parameter((order, N)); Xdot_p = cvxpy.Parameter((order, N)); g_p = cvxpy.Parameter(nonneg=True)
             _PROBLEM_CACHE[(order, N)] = (cvxpy.Problem(cvxpy.Minimize(
[…]

[…]
