### 2026-09-03 85034ca
attempting to catch error so experiment can run through, weakened checking too much
