### 2026-09-03 ce5f988
catching solvererrors once again in lineardiff too now
