### 2026-09-03 1d41696
restructured lineardiff ever so slightly and renamed step_size to stride for shortness, consistency with slide_function, and to avoid collision with the concept of dt
