### 2026-09-03 089db22
beating out a reproducibility bug that caused the experiment to fail. When I reworked the error catching, something slipped through
