### 2026-09-03 00f622b
calculate tvgamma with fewer logarithms, and default the tvrdiff and robustdiff search spaces to exclude order 1, only adding it back as necessary, like for fitting the triangles data
