### 2026-09-03 38b27f0
Fit only lineardiff's load-bearing row, and stop laundering its derivative
Every row of X = A*integral_X + C*B above the last is a tautology, because integral_X[i] comes out bitwise identical to X[i-1] when both are a cumulative_trapezoid of the row beneath, so a shift matrix solves them at exactly zero residual with zero integration constants; since the objective is separable by row and the reconstruction reads only the bottom row, the old code was solving for 2*order^2 unknowns in order to use 2*order of them, and fitting x = a*integral_X + c*B directly reproduces the derivative to 1e-12 rel
[…]
