+++ pynumdiff/linear_model.py
@@ -3,7 +3,6 @@ from warnings import warn
 import math, scipy
 import numpy as np
 
-from pynumdiff.finite_difference import finitediff
 from pynumdiff.utils import utility
 
 try: import cvxpy
@@ -64,21 +63,25 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
         tau = np.arange(N)/N # dtau is 1/N, so tau spans [0, 1) whatever the window's width in real time
         B = np.vstack([tau**(order-n-1)/math.factorial(order-n-1) for n in range(order)])
 
-        # Solve X = A*integral_X + C*B for A and C. Canonicalizing costs ~85% of a solve here and depends only on the
-        # shapes, so the problem is built once per (order, window length) and re-solved with new data, worth about 7x.
-        # gamma is a Parameter too, which keeps a cached problem valid across a whole `optimize` sweep rather than only
-        # within one call. All of which requires the problem to be DPP.
+        # Only the bottom row of X = A*integral_X + C*B carries dynamics. integral_X[i] is bitwise identical to X[i-1],
+        # both being a cumulative_trapezoid of the row beneath, so the upper rows are tautologies that a shift matrix
+        # solves at exactly zero residual with zero constants. The objective is separable by row and the derivative
+        # below reads only the bottom row, so fitting the full A and C solves for 2*order^2 unknowns to use 2*order of
+        # them. Fit that one row directly instead: x = a*integral_X + c*B. See #223
+        # Canonicalizing costs ~85% of a solve here and depends only on the shapes, so the problem is built once per
+        # (order, window length) and re-solved with new data. gamma is a Parameter too, which keeps a cached problem
+        # valid across a whole `optimize` sweep rather than only within one call. All of which requires DPP.
         if (order, N) not in _PROBLEM_CACHE:
             if len(_PROBLEM_CACHE) >= 64: _PROBLEM_CACHE.clear() # bound what an optimizer's window sweep accumulates
-            A_v = cvxpy.Variable((order, order)); C_v = cvxpy.Variable((order, order))
-            X_p = cvxpy.Parameter((order, N)); Xdot_p = cvxpy.Parameter((order, N)); g_p = cvxpy.Parameter(nonneg=True)
+            a_v = cvxpy.Variable(order); c_v = cvxpy.Variable(order)
+            iX_p = cvxpy.Parameter((order, N)); x_p = cvxpy.Parameter(N); g_p = cvxpy.Parameter(nonneg=True)
             _PROBLEM_CACHE[(order, N)] = (cvxpy.Problem(cvxpy.Minimize(
-                cvxpy.sum_squares(Xdot_p - (A_v @ X_p + cvxpy.vstack([C_v[i, :] @ B for i in range(order)]))) +
-                g_p*cvxpy.sum(cvxpy.abs(C_v)) + 1e-6*cvxpy.sum(cvxpy.abs(A_v)))), A_v, C_v, X_p, Xdot_p, g_p)
-                # Smooth x has near-polynomial integrals, and B is polynomials, so A and C become interchangeable. 1e-6 on A's
+                cvxpy.sum_squares(x_p - (a_v @ iX_p + c_v @ B)) +
+                g_p*cvxpy.sum(cvxpy.abs(c_v)) + 1e-6*cvxpy.sum(cvxpy.abs(a_v)))), a_v, c_v, iX_p, x_p, g_p)
+                # Smooth x has near-polynomial integrals, and B is polynomials, so a and c become interchangeable. 1e-6 on a's
                 # norm enforces uniqueness while not significantly shrinking, which costs accuracy. See #223
-        prob, A_v, C_v, X_p, Xdot_p, g_p = _PROBLEM_CACHE[(order, N)]
[…]
