### 2026-09-03 beac0e9
Assemble lineardiff's output the way polydiff does
_lineardiff already returns both a smoothed x_hat and the derivative the fit produces, so slide_function can kernel-average both directly instead of averaging one and reconstructing the other; integrating the averaged derivative back to x_hat was redundant work on top of an estimate the windows had already made. See #223

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
