### 2026-09-03 3d5167b
Stop lineardiff mutating the caller's array through a window view
slide_function hands _lineardiff a slice of the working vector, which numpy makes a view rather than a copy, so the in-place x -= mean introduced in 00f622b subtracted each window's local mean from the shared array and left every later overlapping window fitting progressively corrupted data. Restoring the copying form fixes six test_diff_method failures and one test_multidimensionality failure that are red on master right now.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
