### 2026-09-03 5c4eee3
Name lineardiff's centered copy, and retune its test parameters
The centered window gets its own name so the reason it must not be an in-place subtraction is visible at the point of use. Fitting six parameters to eleven samples was never a defensible test point, and the pipeline change makes it worse rather than papering over it with a binomial smoother, so the test now slides a 21-sample window over the 31-sample signals at a gamma suited to their 0.05 noise; the regenerated bounds are tighter than the old ones in nine cells, notably every noisy-derivative column, and looser in three.

Co-Aut
[…]
