### 2026-09-03 b36855d
added analysis to notebook 4, a comment about why going up to order 4 in lineardiff's search space isn't necessarily a good idea, and fixed bug in lineardiff that broke the build
