### 2026-09-04 florisvb/PyNumDiff#228 @pavelkomarov
`slide_function`-based methods handle irregular steps, but poorly with big density variation
`polydiff` fits a consistent polynomial order to a window of some number of *samples*, which is fine if there is just slight jitter in the data but not fine if there is big density somewhere, because then the model wants to overfit noise in that dense region compared to elsewhere in the domain.

Discovered while trying to get Claude to extend `lineardiff` to handle variable step, which also depends on `slide_function`. Other models that support variable step--`splinediff`, `rbfdiff`, `rtsdiff`, and `robustdiff`--do not suffer this problem, because they work globally or by evolving over `t`, not working merely with a set number of samples in windows. I'm going to push the changes that make `lineardiff` more like `polydiff` and then try to get a fix to both of these at once.

This one continues https://github.com/florisvb/PyNumDiff/issues/173, revisiting a concern I swept under the rug at the time.
