### 2026-09-04 463efc2
made lineardiff capable of supporting irreg steps, although this brings up #228 in the process
