+++ pynumdiff/linear_model.py
@@ -30,7 +30,9 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
     if np.any(np.isnan(x)): raise ValueError("`x` may not contain NaN. CVXPY cannot form a problem with missing data.")
-    if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. The integrals of x are accumulated at a constant step.")
+    if not np.isscalar(dt):
+        if len(dt) != x.shape[axis]: raise ValueError("If `dt` is given as sample locations, it must be as long as `x` along `axis`.")
+        if np.any(np.diff(dt) <= 0): raise ValueError("`dt` must be strictly increasing when given as sample locations.")
     if window_size:
         if window_size % 2 == 0: window_size += 1; warn("Kernel window size should be odd. Added 1 to length.")
         if stride is None: stride = max(1, window_size//5) # Keeps stride out of the optimizer's search space.
@@ -47,20 +49,24 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
         # matrix below is one more integration than the row beneath it, so in raw time the rows differ by powers of the
         # window duration T. Integrating in tau instead holds cond number reasonable. A is fit in units of 1/tau, so the
         # derivative it reconstructs is d/dtau and gets divided by T at the end to return to d/dt.
-        T = len(xc)*dt
-        dtau = dt/T
+        # With sample locations in hand, tau is read off them instead of assumed uniform; every step below is a
+        # quadrature or a pointwise basis evaluation, so neither cares whether the samples are evenly spaced.
+        equispaced = np.isscalar(dt)
+        T = len(xc)*dt if equispaced else dt[-1] - dt[0]
+        tau = np.arange(len(xc))/len(xc) if equispaced else (dt - dt[0])/T # dtau is 1/N when uniform, so tau spans [0, 1)
+        step = dt/T if equispaced else tau # what the integrals below accumulate against
 
         # Generate the matrix of integrals of x, then integrate it once more for the right hand side
         X = [xc]
         for n in range(1, order):
-            X.append(utility.integrate_dxdt_hat(X[-1], dtau))
+            X.append(utility.integrate_dxdt_hat(X[-1], step))
         X = np.vstack(X[::-1])
-        integral_X = np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, axis=1)))*dtau
+        integral_X = np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, axis=1)))*step \
+            if equispaced else np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, x=tau, axis=1)))
 
         # Powers of nondimensional time, the shape the integration constants take after `order` integrations. B[1:] is
         # its own derivative, since differentiating tau^k/k! gives tau^(k-1)/(k-1)!, which is the next row down.
         N = X.shape[1]
-        tau = np.arange(N)/N # dtau is 1/N, so tau spans [0, 1) whatever the window's width in real time
         B = np.vstack([tau**(order-n-1)/math.factorial(order-n-1) for n in range(order)])
 
         # Only the bottom row of X = A*integral_X + C*B carries dynamics. integral_X[i] is bitwise identical to X[i-1],
[…]
