### 2026-09-04 #11172 @pavelkomarov
Even more work:

- reworked `lineardiff` to be a bit more mathematically pared-down, so it's now faster and more elegant
- extended irregular step support to `lineardiff` and in the process found a bug (https://github.com/florisvb/PyNumDiff/issues/228) where sample density differences cause overfitting in densely-sampled regions, also affecting `polydiff`

I'll update this comment
