### 2026-09-04 790ddbe
Size sliding windows by duration when samples are irregular
slide_function carved windows by sample count, a fixed duration only under uniform sampling. Where density varies the span followed it, and the fit degraded worst where data was densest, a window too short in time to show the model any curvature. Windows are now chosen by time, with a min_samples floor for stretches left too sparse to fit. On a 14x density ramp lineardiff goes 2.9178 to 0.2023, polydiff 1.5894 to 0.1514.

Kernels become shape functions taking a window size or sample positions, always normalized. The Friedrichs bump is
[…]
