### 2026-09-04 #11172 @pavelkomarov
`lineardiff` is now in the list that support irregular `dt`.
