### 2026-09-04 633f871
improved comments and added guard to make order >1 for lineardiff, since order = 1 just returns 0s
