### 2026-09-05 #11172 @pavelkomarov
- realized `lineardiff` could be made to support missing data too
