### 2026-09-05 844c0e5
realized lineardiff should also be able to to handle missing data. Now it does
