+++ pynumdiff/linear_model.py
@@ -30,7 +30,6 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
-    if np.any(np.isnan(x)): raise ValueError("`x` may not contain NaN. CVXPY cannot form a problem with missing data.")
     if order < 2: raise ValueError("`order` must be at least 2. Order 1 is just xdot = a*x, solved by scalar exponential.")
     if not np.isscalar(dt):
         if len(dt) != x.shape[axis]: raise ValueError("If `dt` is given as sample locations, it must be as long as `x` along `axis`.")
@@ -47,7 +46,10 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
     @np.errstate(invalid='ignore', over='ignore') # cvxpy#3503: building a sum atom reduces over uninitialized memory
     def _lineardiff(x, dt, order, gamma): # just to read a shape, so it warns when that memory holds garbage
         """Fit X = A*integral_X + C*B, then differentiate it to Xdot = A*X + C*dB to get the derivative"""
-        mean = np.mean(x)
+        obs = ~np.isnan(x) # Missing values drop out of the fit, and the fitted model imputes them back
+        if obs.sum() < 2*order: raise ValueError(f"Window encountered with only {obs.sum()} non-NaN samples < {2*order} "
+                f"needed to determine an order {order} fit. Widen `window_size` or lower `order`.")
+        mean = np.nanmean(x)
         xc = x - mean # not in-place `-=`: slide_function operates on views, so don't modify underlying array
 
         # Work in nondimensional time tau = t/T, so the normalized window spans tau \in [0, 1]. Each row of the
@@ -59,42 +61,42 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
         equispaced = np.isscalar(dt)
         T = len(xc)*dt if equispaced else dt[-1] - dt[0]
         tau = np.arange(len(xc))/len(xc) if equispaced else (dt - dt[0])/T # dtau is 1/N when uniform, so tau spans [0, 1)
-        step = dt/T if equispaced else tau # what the integrals below accumulate against
-
-        # Generate the matrix of integrals of x, then integrate it once more for the right hand side
-        X = [xc]
-        for n in range(1, order):
-            X.append(utility.integrate_dxdt_hat(X[-1], step))
+        # Generate the matrix of integrals of x, then integrate it once more for the right hand side. Holes are
+        # bridged linearly first, which is what the trapezoid rule already assumes between neighbors, so the
+        # quadrature runs over the full grid and every row stays defined where samples are missing. Only the
+        # stack is bridged; `xc` keeps its holes, and the fit below reads observed columns alone.
+        X = [xc if obs.all() else np.interp(tau, tau[obs], xc[obs])]
+        for n in range(1, order): X.append(utility.integrate_dxdt_hat(X[-1], tau))
         X = np.vstack(X[::-1])
-        integral_X = np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, axis=1)))*step \
-            if equispaced else np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, x=tau, axis=1)))
+        integral_X = np.hstack((np.zeros((X.shape[0], 1)), scipy.integrate.cumulative_trapezoid(X, x=tau, axis=1)))
 
         # Powers of nondimensional time, the shape the integration constants take after `order` integrations. B[1:] is
[…]
