### 2026-09-05 7666ffc
Clarify keyword arguments and add methods summary table
Updated the README to clarify the description of keyword arguments and added a summary table for major methods.
