### 2026-09-05 dc141a5
Add the JOSS 2.0 paper, its comparison figure, and a draft-pdf workflow
paper.md and paper.bib present the second-generation package: its design, the seven method families, the capability table, hyperparameter optimization, and testing. Figure 1 differentiates three simulated signals with two methods each, drawn from six of the families, with hyperparameters chosen by `pynumdiff.optimize` at measured bandlimits rather than by eye, and make_figure.py regenerates it. The table and figure are written so JOSS's template cannot break them across pages. A draft-pdf GitHub action builds the paper on 
[…]
