+++ paper/paper.bib
@@ -1,0 +1,181 @@
+@article{vanBreugel2022,
+  doi       = {10.21105/joss.04078},
+  url       = {https://doi.org/10.21105/joss.04078},
+  year      = {2022},
+  publisher = {The Open Journal},
+  volume    = {7},
+  number    = {71},
+  pages     = {4078},
+  author    = {{van Breugel}, Floris and Liu, Yuying and Brunton, Bingni W. and Kutz, J. Nathan},
+  title     = {{PyNumDiff}: A {Python} package for numerical differentiation of noisy time-series data},
+  journal   = {Journal of Open Source Software}
+}
+
+@misc{komarov2025,
+  title         = {A Taxonomy of Numerical Differentiation Methods},
+  author        = {Komarov, Pavel and {van Breugel}, Floris and Kutz, J. Nathan},
+  year          = {2025},
+  eprint        = {2512.09090},
+  archivePrefix = {arXiv},
+  primaryClass  = {math.NA},
+  url           = {https://arxiv.org/abs/2512.09090}
+}
+
+@article{vanBreugel2020numerical,
+  doi     = {10.1109/ACCESS.2020.3034077},
+  year    = {2020},
+  author  = {{van Breugel}, Floris and Kutz, J. Nathan and Brunton, Bingni W.},
+  journal = {IEEE Access},
+  title   = {Numerical differentiation of noisy data: A unifying multi-objective optimization framework},
+  volume  = {8},
+  pages   = {196865--196877}
+}
+
+@article{chartrand2011numerical,
+  author  = {Rick Chartrand},
+  title   = {Numerical differentiation of noisy, nonsmooth data},
+  journal = {ISRN Applied Mathematics},
+  year    = {2011},
+  volume  = {2011},
+  pages   = {164564},
+  doi     = {10.5402/2011/164564}
+}
+
+@article{brunton2016discovering,
+  author    = {Steven L. Brunton and Joshua L. Proctor and J. Nathan Kutz},
+  title     = {Discovering governing equations from data by sparse identification of nonlinear dynamical systems},
+  journal   = {Proceedings of the National Academy of Sciences},
+  year      = {2016},
+  volume    = {113},
+  number    = {15},
+  pages     = {3932--3937},
+  doi       = {10.1073/pnas.1517384113}
+}
+
+@article{virtanen2020scipy,
+  author  = {Pauli Virtanen and Ralf Gommers and Travis E. Oliphant and others},
+  title   = {{SciPy} 1.0: Fundamental algorithms for scientific computing in {Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  doi     = {10.1038/s41592-019-0686-2}
+}
+
+@article{harris2020array,
+  author  = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. {van der Walt} and others},
+  title   = {Array programming with {NumPy}},
+  journal = {Nature},
+  year    = {2020},
+  volume  = {585},
+  pages   = {357--362},
+  doi     = {10.1038/s41586-020-2649-2}
+}
+
+@article{diamond2016cvxpy,
+  author  = {Steven Diamond and Stephen Boyd},
+  title   = {{CVXPY}: A {Python}-embedded modeling language for convex optimization},
+  journal = {Journal of Machine Learning Research},
+  year    = {2016},
+  volume  = {17},
+  number  = {83},
+  pages   = {1--5}
+}
+
+@article{savitzky1964,
+  author  = {Abraham Savitzky and Marcel J. E. Golay},
+  title   = {Smoothing and Differentiation of Data by Simplified Least Squares Procedures},
[…]
