+++ LICENSE.txt
@@ -1,3 +1,5 @@
+The MIT License (MIT)
+
 Copyright (c) 2025
 Floris van Breugel, The University of Nevada, Reno, Mechanical Engineering Department
 Pavel Komarov, Univerity of Washington Electrical and Computer Engineering
