+++ paper/paper.md
@@ -1,0 +1,119 @@
+---
+title: 'PyNumDiff: Practical Numerical Differentiation for Noisy Data'
+tags:
+  - Python
+  - numerical differentiation
+  - time series
+  - denoising
+  - dynamics
+  - signal processing
+authors:
+  - name: Pavel Komarov
+    orcid: 0009-0007-7482-2807
+    corresponding: true
+    affiliation: 1
+  - name: Floris van Breugel
+    orcid: 0000-0001-6538-7179
+    affiliation: 2
+  - name: Maria Protogerou
+    affiliation: 3
+  - name: J. Nathan Kutz
+    orcid: 0000-0002-6004-2275
+    affiliation: 4
+affiliations:
+  - name: Department of Electrical and Computer Engineering, University of Washington, USA
+    index: 1
+  - name: Department of Mechanical Engineering, University of Nevada, Reno, USA
+    index: 2
+  - name: Department of Applied Mathematics, University of Washington, USA
+    index: 3
+  - name: Autodesk Research, London, UK
+    index: 4
+date: 1 April 2026
+bibliography: paper.bib
+---
+
+# Summary
+
+Derivatives of measured data are vital across science and engineering: identifying governing equations, designing controllers, and processing sensor streams alike. The textbook remedy, finite differencing, amplifies noise as $1/\Delta t$ and deteriorates rapidly as data grows noisier. Smoothing before differencing helps, but algorithm choice and settings substantially affect the result, and no single approach wins universally.
+
+PyNumDiff is an open-source Python package consolidating a broad suite of numerical differentiation methods under a unified API, along with an optimization scheme to select hyperparameters even in the absence of ground truth. It implements seven algorithm families: (1) prefiltering followed by finite difference calculation, (2) iterated finite differencing, (3) polynomial fitting [@savitzky1964], (4) global (spectral) and local (wavelet, radial) basis function fitting, (5) total variation regularization [@chartrand2011numerical], (6) Kalman smoothing [@kalman1960; @rauch1965], and (7) local approximation with linear models. All PyNumDiff methods return estimates of the true signal and its derivative as a matched pair, `(x_hat, dxdt_hat)`. A companion survey [@komarov2025, Section 7] situates these methods in the literature, covers the theory behind each method, benchmarks all methods across test signals, and guides selection for different application scenarios. This paper describes the package's second generation, through the v0.3 release series.
+
+
+# Statement of Need
+
+Numerical differentiation has a diverse ecosystem of algorithms, each with different theoretical foundations and putative strengths. But there has been no consolidated home where methods can be easily swapped or tested head-to-head, and practitioners are left assembling solutions piecemeal from disparate packages [@vanBreugel2022].
+
+PyNumDiff addresses this gap: Its unified interface lets users compare methods on the same data, exploit specialized capabilities (e.g. outlier robustness), and tune methods' hyperparameters against known ground truth or to a given signal bandlimit. Derivative estimation trades data fidelity against smoothness, so PyNumDiff uses dual-objective optimization, with a single scalar weight to steer the balance [@vanBreugel2020numerical]. A natural and growing application is SINDy [@brunton2016discovering], which discovers governing equations by regressing measured derivatives, making quality of the derivative estimates a direct determinant of model accuracy.
+
+
+![Three simulated signals with additive noise (top) and their derivatives (bottom). Each panel contrasts two methods from different families, six of the seven in total, with a fixed color per method. Hyperparameters come from `pynumdiff.optimize` rather than hand-tuning, at the bandlimit noted with each derivative panel: 1 Hz for the sines and triangle, 2 Hz for the faster Lorenz signal.\label{fig:comparison}](methods_comparison.png)
+
+
+# State of the Field
+
[…]
