### 2026-09-05 #11172 @pavelkomarov
Added a figure below the fold, feeling like it was missing some eye candy.
