### 2026-09-05 f978499
trimming
