### 2026-09-07 6bd9d24
worked through lineardiff code simplifying now that I've done the math in the taxonomy paper today and finally understood well enough to tighten comments and see some opportunities to unify a couple lines. Other than that just a single bound tightened in the tests, and I modified some comments in other files
