+++ pynumdiff/tests/test_diff_methods.py
@@ -247,7 +247,7 @@ error_bounds = {
                  [(-1, -1), (1, 1), (0, -1), (1, 1)],
                  [(-1, -1), (0, 0), (0, -1), (1, 1)],
                  [(0, 0), (2, 2), (0, 0), (2, 2)],
-                 [(0, -1), (2, 2), (0, -1), (2, 2)]],
+                 [(-1, -1), (2, 2), (0, -1), (2, 2)]],
     lineardiff_irreg_step: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                             [(-1, -1), (1, 0), (0, -1), (1, 0)],
                             [(-1, -1), (0, 0), (0, -1), (1, 1)],
