+++ pynumdiff/linear_model.py
@@ -11,29 +11,28 @@ except ImportError: pass
 
 _PROBLEM_CACHE = {} # (order, window length) -> a parametrized CVXPY problem, so identically shaped windows reuse it
 
-def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='friedrichs', axis=0):
+def lineardiff(x, dt_or_t, order, gamma, window_size=None, stride=None, kernel='friedrichs', axis=0):
     """Fit a linear dynamical system to windows of the data, then differentiate that model.
 
     :param np.array[float] x: data to differentiate. May be multidimensional; see :code:`axis`.
-    :param float dt: step size
-    :param int>1 order: order of the ODE fit, the number of states in the linear system, how many times :code:`x` is integrated.
+    :param float or array[float] dt_or_t: This function supports variable step size. This parameter is either the constant
+        :math:`\\Delta t` if given as a single float, or data locations if given as an array of same length as :code:`x`.
+    :param int>0 order: order of the ODE fit, the number of states in the linear system, how many times :code:`x` is integrated.
     :param float gamma: regularization term, in multiples of the data's own scale, so a given value means the same
             thing whatever the units.
     :param int window_size: number of samples in the sliding window, or number of average step sizes to use as window
             width if irregular sampling; if not given, no sliding
-    :param int stride: step size for sliding. Defaults to :code:`window_size/5`, because what matters is overlap
-            ratio, not an absolute stride: a fifth costs a few percent of accuracy against a much finer stride while
-            running several times faster, and strides past about half the window degrade badly
+    :param int stride: step size for sliding. Defaults to :code:`window_size//5`, which costs only a few percent of accuracy
+            against a much finer stride while running reasonably fast; strides > half the window degrade performance badly
     :param str kernel: name of kernel to use for weighting and smoothing windows ('gaussian' or 'friedrichs')
     :param int axis: axis along which to differentiate (default 0)
 
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
-    if order < 2: raise ValueError("`order` must be at least 2. Order 1 is just xdot = a*x, solved by scalar exponential.")
-    if not np.isscalar(dt):
-        if len(dt) != x.shape[axis]: raise ValueError("If `dt` is given as sample locations, it must be as long as `x` along `axis`.")
-        if np.any(np.diff(dt) <= 0): raise ValueError("`dt` must be strictly increasing when given as sample locations.")
+    if not np.isscalar(dt_or_t):
+        if len(dt_or_t) != x.shape[axis]: raise ValueError("If `dt_or_t` is given as array-like, must have same length as `x`.")
+        if np.any(np.diff(dt_or_t) <= 0): raise ValueError("`dt_or_t` must be strictly increasing, so integration has all positive widths.")
     if window_size:
         if window_size < 2*order: # a and c hold `order` unknowns each, so need at least as many pieces of info to set up well-posed cost
             window_size = 2*order + 1 - (2*order)%2
@@ -46,59 +45,47 @@ def lineardiff(x, dt, order, gamma, window_size=None, stride=None, kernel='fried
[…]
