### 2026-09-07 79b5acf
docstring
