### 2026-09-07 c9ecb7c
Drop splinediff's num_iterations, which was doing nothing
Refitting a spline to its own output holds the residual budget s*N*sigma^2 fixed while the input gets cleaner, so each pass satisfies the same budget with fewer knots and can only smooth harder. Measured, it does: on a sine at s=1 the derivative RMSE runs 0.80, 2.55, 3.93, 4.26 for one through five passes, and peak-to-peak falls from 1.80 to 1.10 against a truth of 1.95, the opposite of the peak preservation the parameter was there to provide. Iterating only ever rescues an s set too small, and simply raising s beats every iterated vari
[…]
