### 2026-09-08 37046db
improved spectraldiff's parameters and capabilities. It can now also do odd extension by detrending and then adding the trend back at the end of the method
