+++ pynumdiff/basis_fit.py
@@ -5,59 +5,64 @@ import pywt
 
 from pynumdiff.utils import utility
 
-def spectraldiff(x, dt, cutoff_freq, even_extension=True, pad_to_zero_dxdt=True, axis=0):
+def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0):
     """Take a derivative in the Fourier domain, with high frequency attentuation.
+
+    The FFT treats data as periodic, so a signal whose ends do not meet has a discontinuity across the wrap, and
+    Gibbs ringing contaminates the estimate. `extension` picks the remedy: 'detrend' subtracts the line through the
+    endpoints, 'even' mirrors the signal, and 'odd' detrends and then reflects through the last endpoint.
 
     :param np.array[float] x: data to differentiate. May be multidimensional; see :code:`axis`.
     :param float dt: step size
     :param float cutoff_freq: The high frequency cutoff as a multiple of the Nyquist frequency: Should be between 0
         and 1. Frequencies below this threshold will be kept, and at and above will be zeroed.
-    :param bool even_extension: if True, extend the data with an even extension so signal starts and ends at the same value.
-    :param bool pad_to_zero_dxdt: if True, extend the data with extra regions that smoothly force the derivative to
-        zero before taking FFT.
+    :param str extension: how to make the data periodic: :code:`None`, :code:`'even'`, :code:`'detrend'`, or :code:`'odd'`.
+        None is right only for genuinely periodic signals.
+    :param bool pad_to_flat: if True, extend the edges with smoothed repeats of the end values, giving the Gibbs
+        ringing somewhere to go that gets discarded.
+    :param int axis: data dimension along which to differentiate
 
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
              - **dxdt_hat** (np.array) -- estimated derivative of x
     """
     if np.any(np.isnan(x)): raise ValueError("`x` may not contain NaN. Missing values spread through the FFT to make the whole spectrum NaN.")
     if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. The FFT assumes uniformly sampled data.")
+    if extension not in (None, 'even', 'detrend', 'odd'): raise ValueError("`extension` must be None, 'even', 'detrend', or 'odd'.")
 
-    L = x.shape[axis]
+    x = np.moveaxis(x, axis, 0)
+    N = len(x)
+    y = x.reshape(N, -1) # flat 2D of all the vectors to differentiate
 
-    # Make derivative go to zero at the ends (optional)
-    if pad_to_zero_dxdt:
-        padding = 100
-        pre = np.repeat(np.take(x, [0], axis=axis), padding, axis=axis) # take keeps dimensions, unlike x[0]
-        post = np.repeat(np.take(x, [-1], axis=axis), padding, axis=axis)
-        x = np.concatenate((pre, x, post), axis=axis) # extend the edges
-        kernel = utility.uniform_kernel(padding//2)
-        x_smoothed = utility.convolutional_smoother(x, kernel, axis=axis) # smooth the padded edges in
-        m = (slice(None),)*axis + (slice(padding, L+padding),) + (slice(None),)*(x.ndim-axis-1) # middle
-        x_smoothed[m] = x[m] # restore original signal in the middle
-        x = x_smoothed
-    else:
-        m = (slice(None),)*axis + (slice(0, L),) + (slice(None),)*(x.ndim-axis-1) # indices where signal lives
[…]
