+++ pynumdiff/tests/test_diff_methods.py
@@ -57,7 +57,7 @@ diff_methods_and_params = [
     (savgoldiff, {'degree':2, 'window_size':5, 'smoothing_win':5}),
     (splinediff, {'degree':5, 's':2}),
     (spline_irreg_step, {'degree':5, 's':2}),
-    (spectraldiff, {'cutoff_freq':0.2}),
+    (spectraldiff, {'cutoff_freq':0.2, 'pad_to_flat':True}), # test padding too, for coverage and because it interacts with other stuff
     (rbfdiff, {'sigma':0.5, 'lmbd':0.001}),
     (waveletdiff, {'wavelet':'db8', 'threshold':1.0}),
     (rtsdiff, {'order':2, 'log_qr_ratio':7, 'forwardbackward':True}),
@@ -171,11 +171,11 @@ error_bounds = {
                         [(1, 0), (2, 2), (1, 0), (2, 1)],
                         [(1, 0), (2, 2), (1, 0), (3, 2)]],
     spectraldiff: [[(flr, flr), (flr, flr), (0, -1), (0, 0)],
-                  [(flr, flr), (flr, flr), (0, -1), (0, 0)],
-                  [(-1, -1), (0, 0), (0, -1), (0, 0)],
-                  [(-1, -1), (0, 0), (0, -1), (0, 0)],
-                  [(2, 1), (3, 2), (2, 1), (3, 2)],
-                  [(2, 1), (3, 3), (2, 1), (3, 3)]],
+                  [(0, 0), (1, 1), (0, 0), (1, 1)],
+                  [(1, 0), (1, 1), (1, 1), (1, 1)],
+                  [(0, 0), (1, 1), (0, 0), (1, 1)],
+                  [(1, 1), (2, 2), (1, 1), (2, 2)],
+                  [(1, 1), (3, 3), (1, 1), (3, 3)]],
     rbfdiff: [[(-2, -2), (0, 0), (0, -1), (1, 0)],
               [(-1, -1), (1, 0), (0, -1), (1, 1)],
               [(-1, -1), (1, 1), (0, -1), (1, 1)],
