### 2026-09-08 18785b5
updated badges and citations
