### 2026-09-09 0e04cda
claude and I stabbing waveletdiff over and over
