+++ pynumdiff/tests/test_diff_methods.py
@@ -183,11 +183,11 @@ error_bounds = {
               [(0, 0), (2, 2), (0, 0), (2, 2)],
               [(1, 1), (3, 3), (1, 1), (3, 3)]],
     waveletdiff: [[(flr, flr), (flr, flr), (0, -1), (1, 0)],
-                  [(-2, -2), (0, -1), (0, -1), (1, 1)],
-                  [(-1, -2), (0, 0), (0, -1), (1, 1)],
-                  [(-2, -2), (0, 0), (0, -1), (1, 1)],
-                  [(-1, -1), (2, 2), (0, 0), (2, 2)],
-                  [(-1, -1), (3, 3), (0, 0), (3, 3)]],
+                 [(-2, -2), (-1, -1), (0, -1), (1, 0)],
+                 [(-2, -2), (0, 0), (0, -1), (1, 1)],
+                 [(-2, -2), (0, 0), (0, -1), (1, 1)],
+                 [(-1, -1), (2, 2), (0, 0), (2, 2)],
+                 [(-1, -1), (3, 3), (0, 0), (3, 3)]],
     tvrdiff_1: [[(flr, flr), (flr, flr), (0, -1), (1, 0)],
                 [(flr, flr), (-11, flr), (-1, -1), (-1, -2)],
                 [(0, -1), (1, 0), (0, -1), (1, 0)],
