+++ pynumdiff/basis_fit.py
@@ -5,6 +5,7 @@ import pywt
 
 from pynumdiff.utils import utility
 
+
 def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0):
     """Take a derivative in the Fourier domain, with high frequency attentuation.
 
@@ -16,10 +17,10 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
     :param float dt: step size
     :param float cutoff_freq: The high frequency cutoff as a multiple of the Nyquist frequency: Should be between 0
         and 1. Frequencies below this threshold will be kept, and at and above will be zeroed.
-    :param str extension: how to make the data periodic: :code:`None`, :code:`'even'`, :code:`'detrend'`, or :code:`'odd'`.
-        None is right only for genuinely periodic signals.
-    :param bool pad_to_flat: if True, extend the edges with smoothed repeats of the end values, giving the Gibbs
-        ringing somewhere to go that gets discarded.
+    :param str extension: how to make the data periodic: :code:`None`, :code:`'even'`, :code:`'detrend'`, or :code:`'odd'`,
+        where the latter also detrends before extension to avoid discontinuity. None is for genuinely periodic signals.
+    :param bool pad_to_flat: if True, extend the edges with smoothed-in repeats of the end values, giving Gibbs ringing
+        somewhere to go that gets discarded.
     :param int axis: data dimension along which to differentiate
 
     :return: - **x_hat** (np.array) -- estimated (smoothed) x
@@ -126,12 +127,21 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-def waveletdiff(x, dt, wavelet='db8', level=None, threshold=1.0, axis=0, mode='symmetric'):
+_DERIV_TAPS = {} # untrimmed derivative-operator taps per wavelet; independent of the data, dt, and N
+
+# Orthogonal wavelets whose scaling function's spectrum dips near zero, making it ill-conditioned to invert: the
+# derivative operator then spreads over 469 to 3113 taps, against 35 to 199 for the other 67 pywt offers, and stops
+# being the local filter this method is built around. Found by sweeping min|fft(phi)| over all of them; the cut at
+# 0.3 is where the trade stops paying, since a stricter floor rejects more families without shortening the worst
+# operator any further.
+_ILL_CONDITIONED = frozenset(('db5', 'db14', 'db18', 'db27', 'db36', 'sym11', 'sym13'))
+
+def waveletdiff(x, dt, wavelet='db8', level=None, threshold=1.0, axis=0, mode='symmetric', num_shifts=None):
     """Smooth and differentiate noisy data in a wavelet basis.
 
     Three steps: (1) decompose x with the DWT and soft-threshold the detail coefficients to denoise (Donoho-Johnstone
-    universal threshold), reconstructing a smoothed x_hat; (2) extend x_hat antisymmetrically so the periodic derivative
-    operator stays accurate at the edges; (3) recover the wavelet scaling coefficients of x_hat and apply the analytic
+    universal threshold), reconstructing a smoothed x_hat; (2) detrend x_hat and reflect it through its last endpoint so it is
+    genuinely periodic and the derivative operator stays accurate at the edges; (3) recover the wavelet scaling coefficients of x_hat and apply the analytic
     derivative of the wavelet basis.
 
[…]
