+++ pynumdiff/optimize.py
@@ -55,12 +55,15 @@ method_params_and_bounds = {
                 'lmbd': [1e-3, 1e-2, 1e-1]},
               {'sigma': (1e-2, 1e3),
                 'lmbd': (1e-3, 0.5)}),
-    waveletdiff: ({'wavelet': {'db8', 'db12', 'sym8', 'coif1'}, # different data can favor different mother wavelets
+    waveletdiff: ({'wavelet': {'db4', 'db8', 'db12', 'sym8', 'coif1', 'coif2'}, # different data can favor different mother wavelets;
+                 # db4 earns its place on short records, where a long filter's margin is a large fraction of the data
                  'mode': {'symmetric', 'reflect', 'antireflect', 'constant'}, # no boundary rule wins outright, and
                  # symmetric can cost 2x on the wrong signal, so search them. pywt's other modes are excluded: zero
                  # and the periodic pair invent a value jump, and smooth extrapolates the edge slope through noise.
-                 'threshold': [0.5, 1, 2]}, # multiplies the Donoho-Johnstone universal threshold, 0 meaning no denoising.
-                 # `level` is left at its adaptive default, min(dwt_max_level(N, wavelet), 5), which tracks both the signal and filter lengths
+                 'threshold': [1, 2, 4]}, # multiplies the Donoho-Johnstone universal threshold, 0 meaning no denoising.
+                 # Centered higher than a soft-thresholding grid would be, because hard shrinkage keeps its survivors
+                 # whole and so smooths less at equal multiplier; the median best here is 2, against 1.5 for soft.
+                 # `level` is left to its adaptive default, which reads the depth off the noise floor band by band
                 {'threshold': (0.1, 10)}),
     tvrdiff: ({'gamma': [1e-2, 1e-1, 1, 10, 100, 1000],
                'order': {2, 3}, # warning: order 1 hacks the ground-truth-less loss function, tends to win but is usually suboptimal choice in terms of true RMSE
@@ -108,7 +111,7 @@ ROUND = {float: lambda v: v,
 
 # This function has to be at the top level for multiprocessing but is only used by optimize.
 def _objective_function(point, func, x, dt, singleton_params, categorical_params, roundings,
-    dxdt_truth, metric, tvgamma, padding, cache, huberM):
+    dxdt_truth, tvgamma, padding, cache, huberM):
     """Function minimized by scipy.optimize.minimize, needs to have the form: (point, *args) -> float
     This is mildly complicated, because "point" controls the settings of a differentiation function, but
     the method may have numerical and non-numerical parameters, and all such parameters are now passed by
@@ -134,14 +137,14 @@ def _objective_function(point, func, x, dt, singleton_params, categorical_params
     try: x_hat, dxdt_hat = func(x, dt, **point_params, **singleton_params, **categorical_params) # take deriv, add back singletons and categorical choices
     except np.linalg.LinAlgError: cache[key] = 1e10; return 1e10 # some methods can fail numerically
 
-    # Evaluate estimate according to a loss function
+    # Evaluate estimate according to a loss function: against the truth if we have it, else the proxy below.
+    # Only RMSE is offered for the known-truth case. Minimizing `evaluate.error_correlation` instead is degenerate:
+    # it measures how much of the error tracks the signal, which is what smoothing produces, so its minimum is at
+    # no smoothing at all. Measured across five methods it drove the correlation to exactly zero every time while
+    # RMSE rose by 1.1x (savgoldiff) to 12x (waveletdiff). It stays in `utils.evaluate` as a diagnostic to report.
[…]
