+++ pynumdiff/basis_fit.py
@@ -37,7 +37,7 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
     pad = 0
     if pad_to_flat: # repeat the end values outward, smooth the joins, then restore the original in the middle
         pad = 100
-        padded = np.concatenate((np.repeat(y[:1], pad, axis=0), y, np.repeat(y[-1:], pad, axis=0)))
+        padded = np.concatenate((np.repeat(y[:1], pad, axis=0), y, np.repeat(y[-1:], pad, axis=0))) # y[:1] rather than y[0] so 2D shape
         smoothed = utility.convolutional_smoother(padded, utility.uniform_kernel(pad//2), axis=0)
         smoothed[pad:pad+N] = y
         y = smoothed
@@ -127,31 +127,24 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
     return np.moveaxis(x_hat_flattened.reshape(plump), 0, axis), np.moveaxis(dxdt_hat_flattened.reshape(plump), 0, axis)
 
 
-_DERIV_TAPS = {} # untrimmed derivative-operator taps per wavelet; independent of the data, dt, and N
+_OPERATORS = {} # (A^-1, phi', phi) taps per wavelet; independent of the data, dt, and N
 
-# Orthogonal wavelets whose scaling function's spectrum dips near zero, making it ill-conditioned to invert: the
-# derivative operator then spreads over 469 to 3113 taps, against 35 to 199 for the other 67 pywt offers, and stops
-# being the local filter this method is built around. Found by sweeping min|fft(phi)| over all of them; the cut at
-# 0.3 is where the trade stops paying, since a stricter floor rejects more families without shortening the worst
-# operator any further.
-_ILL_CONDITIONED = frozenset(('db5', 'db14', 'db18', 'db27', 'db36', 'sym11', 'sym13'))
-
-def waveletdiff(x, dt, wavelet='db8', level=None, threshold=1.0, axis=0, mode='symmetric', num_shifts=None):
+def waveletdiff(x, dt, wavelet='db8', level=None, threshold=2.0, axis=0, mode='symmetric', num_shifts=None):
     """Smooth and differentiate noisy data in a wavelet basis.
 
-    Three steps: (1) decompose x with the DWT and soft-threshold the detail coefficients to denoise (Donoho-Johnstone
-    universal threshold), reconstructing a smoothed x_hat; (2) detrend x_hat and reflect it through its last endpoint so it is
-    genuinely periodic and the derivative operator stays accurate at the edges; (3) recover the wavelet scaling coefficients of x_hat and apply the analytic
-    derivative of the wavelet basis.
+    Three steps: (1) pre-filter the samples into finest-scale scaling coefficients, which is what the cascade is
+    defined on; (2) decompose with the DWT and hard-threshold the detail coefficients to denoise (Donoho-Johnstone
+    universal threshold), averaged over every alignment of the transform grid; (3) apply the analytic derivative of
+    the basis to the denoised coefficients.
 
     The derivative differentiates the basis functions themselves rather than finite-differencing the signal. PyWavelets
     treats the samples as finest-level scaling coefficients, so x_hat is the interpolant x(t) = sum_n a_n phi(t/dt - n)
     for the scaling function phi. Sampling x and its analytic derivative on the grid gives two convolutions against phi
     and phi' evaluated at *integers*,
 
-        x_hat = Phi @ a     and     x' = Phi_prime @ a,
+        x_hat = A @ a     and     x' = A_prime @ a,
 
-    so x' = Phi_prime @ Phi^-1 @ x_hat, exact for signals the basis can represent. Both matrices need only the values of
[…]

[…]
