### 2026-09-14 f86d89f
still poking waveletdiff, more to come as I clean up comments and get it to agree with the newly-completed math doc
