+++ pynumdiff/basis_fit.py
@@ -97,7 +97,8 @@ def rbfdiff(x, dt_or_t, sigma=1, lmbd=0.01, axis=0):
         t = dt_or_t
 
     # For each vector along the axis of differentiation, the below does the approximate equivalent of this code,
-    # but sparsely in O(N sigma^2), since the rbf falls off rapidly
+    # but sparsely in O(N sigma^2), since the rbf falls off rapidly. Since A is Toeplitz for uniform spacing, it
+    # could also be done via convolution in O(N sigma), but that would give up irregular grid support.
     # t_i, t_j = np.meshgrid(t,t)
     # r = t_j - t_i # radius
     # rbf = np.exp(-(r**2) / (2 * sigma**2)) # radial basis function kernel, O(N^2) entries
@@ -188,10 +189,16 @@ def waveletdiff(x, dt, wavelet='db8', level=None, threshold=2.0, axis=0, mode='s
     # impossible or spreads the operator over hundreds of taps and it stops being the local filter this method is
     # built around. The named seven are the orthogonal wavelets whose min|fft(phi)| falls under 0.3, found by
     # sweeping all 67 pywt offers; below that cut the operator runs 469 to 3113 taps against 35 to 199 above it.
-    if (W.name in ('db5', 'db14', 'db18', 'db27', 'db36', 'sym11', 'sym13') or not W.orthogonal
-            or W.dec_len == 2): # a 2-tap filter is Haar under any of its names: haar, db1, bior1.1, rbio1.1
-        raise ValueError(f"'{wavelet}' can't differentiate here: it needs an orthogonal wavelet whose scaling "
-            "function is differentiable and well conditioned to invert. Try 'db8', 'sym8', or 'coif2'.")
+    # Four biorthogonal wavelets also qualify. The rest of their families fail one of two ways: bior3.x and
+    # rbio3.x put a zero in fft(phi), and the 2.x families spread white noise across bands by 1.6-1.8x where
+    # every orthogonal wavelet holds 1.13-1.27. Both uses of the equal-scale assumption break: one sigma stops
+    # describing every band, and the level probe's 2x trip point is left ~1.1x of margin instead of ~1.6x, so
+    # depth gets picked by noise. `threshold` is searched and absorbs some of the first; `level` is not.
+    if (W.name in ('db5', 'db14', 'db18', 'db27', 'db36', 'sym11', 'sym13') or W.dec_len == 2
+            or not (W.orthogonal or W.name in ('bior4.4', 'bior6.8', 'rbio4.4', 'rbio6.8'))):
+        # a 2-tap filter is Haar under any of its names: haar, db1, bior1.1, rbio1.1
+        raise ValueError(f"'{wavelet}' can't differentiate here: it needs a wavelet whose scaling function is "
+            "differentiable and well conditioned to invert. Try 'db8', 'sym8', 'coif2', or 'bior4.4'.")
 
     N = x.shape[axis]
     x_work = np.ascontiguousarray(np.moveaxis(x, axis, 0)) # differentiation axis to front
@@ -205,6 +212,16 @@ def waveletdiff(x, dt, wavelet='db8', level=None, threshold=2.0, axis=0, mode='s
         # band is noise and hurts if it is signal; white noise carries the same scale in every band, so keep
         # going while a band's robust scale stays near the finest band's and stop at the first one above it.
         # Within 6% of the depth chosen knowing the truth, where a fixed cap of 5 is 23% off. Flat over 1.5-3.
+        # "Same scale in every band" holds because an orthogonal transform is an isometry, which is why the
[…]
