### 2026-09-17 49a2eb7
various random cleanup
