### 2026-09-17 0820fc4
still cleaning, just checkpointing
