+++ pynumdiff/basis_fit.py
@@ -15,7 +15,7 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
 
     :param np.array[float] x: data to differentiate. May be multidimensional; see :code:`axis`.
     :param float dt: step size
-    :param float cutoff_freq: The high frequency cutoff as a multiple of the Nyquist frequency: Should be between 0
+    :param float cutoff_freq: high frequency cutoff as a multiple of the Nyquist frequency: Should be between 0
         and 1. Frequencies below this threshold will be kept, and at and above will be zeroed.
     :param str extension: how to make the data periodic: :code:`None`, :code:`'even'`, :code:`'detrend'`, or :code:`'odd'`,
         where the latter also detrends before extension to avoid discontinuity. None is for genuinely periodic signals.
@@ -30,38 +30,39 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
     if not np.isscalar(dt): raise ValueError("`dt` must be a scalar. The FFT assumes uniformly sampled data.")
     if extension not in (None, 'even', 'detrend', 'odd'): raise ValueError("`extension` must be None, 'even', 'detrend', or 'odd'.")
 
+    N = x.shape[axis]
     x = np.moveaxis(x, axis, 0)
-    N = len(x)
-    y = x.reshape(N, -1) # flat 2D of all the vectors to differentiate
+    x_flat = x.reshape(N, -1) # flat 2D of all the vectors to differentiate
 
     pad = 0
     if pad_to_flat: # repeat the end values outward, smooth the joins, then restore the original in the middle
         pad = 100
-        padded = np.concatenate((np.repeat(y[:1], pad, axis=0), y, np.repeat(y[-1:], pad, axis=0))) # y[:1] rather than y[0] so 2D shape
+        padded = np.concatenate((np.repeat(x_flat[:1], pad, axis=0), x_flat, np.repeat(x_flat[-1:], pad, axis=0))) # [:1] rather than [0] so 2D shape
         smoothed = utility.convolutional_smoother(padded, utility.uniform_kernel(pad//2), axis=0)
-        smoothed[pad:pad+N] = y
-        y = smoothed
+        smoothed[pad:pad+N] = x_flat
+        x_flat = smoothed
 
-    P = len(y); t = np.arange(P)[:, None]*dt # P for "potentially padded"
-    if extension in ('detrend', 'odd') and cutoff_freq > 0: # the line through the endpoints; its derivative is a constant added back below
-        slope = (y[-1] - y[0])/((P-1)*dt)
-        y = y - slope*t # reassign so not in place if y is still a view on x
-    else: slope = 0
+    P = len(x_flat) # P for "potentially padded"
+    if extension in ('detrend', 'odd') and cutoff_freq > 0: # frequency guard so at 0 return 0, not the trend
+        slope = (x_flat[-1] - x_flat[0])/((P-1)*dt) 
+        trend = slope * np.arange(P)[:, np.newaxis]*dt # the line through the endpoints
+        x_flat = x_flat - trend # reassign so not in place if x_flat is still a view on x
+    else: slope = 0; trend = 0
 
-    if extension == 'odd': y = np.concatenate((y, 2*y[-1] - y[-2:0:-1])) # reflect across endpoint
-    elif extension == 'even': y = np.concatenate((y, y[::-1])) # mirror 
+    if extension == 'odd': x_flat = np.concatenate((x_flat, 2*x_flat[-1] - x_flat[-2:0:-1])) # reflect across endpoint
[…]
