### 2026-09-18 103b10c
finally finished with my combover
