+++ pynumdiff/basis_fit.py
@@ -1,7 +1,9 @@
 """Methods based on fitting basis functions to data"""
 import numpy as np
-from scipy import sparse
 import pywt
+from scipy import sparse
+from scipy.linalg import convolution_matrix
+from warnings import warn
 
 from pynumdiff.utils import utility
 
@@ -44,8 +46,8 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
 
     P = len(x_flat) # P for "potentially padded"
     if extension in ('detrend', 'odd') and cutoff_freq > 0: # frequency guard so at 0 return 0, not the trend
-        slope = (x_flat[-1] - x_flat[0])/((P-1)*dt) 
-        trend = slope * np.arange(P)[:, np.newaxis]*dt # the line through the endpoints
+        slope = (x_flat[-1] - x_flat[0])/(P-1) # technically missing a /dt, but the trend would just *dt it back out 
+        trend = slope * np.arange(P)[:, np.newaxis] # the line through the endpoints, slope * time
         x_flat = x_flat - trend # reassign so not in place if x_flat is still a view on x
     else: slope = 0; trend = 0
 
@@ -62,7 +64,7 @@ def spectraldiff(x, dt, cutoff_freq, extension='odd', pad_to_flat=False, axis=0)
     # Derivative = 90 deg phase shift
     if M % 2 == 0: k[M//2] = 0 # odd derivatives get the Nyquist element zeroed out, see https://pavelkomarov.com/spectral-derivatives/math.pdf section 3.1
     omega = 2*np.pi/(dt*M) # factor of 2pi/T turns wavenumbers into frequencies in radians/s
-    dxdt_hat = (np.real(np.fft.ifft(1j * k * omega * X, axis=0))[:P] + slope)[pad:pad+N] # add the trend's constant slope
+    dxdt_hat = (np.real(np.fft.ifft(1j * k * omega * X, axis=0))[:P] + slope/dt)[pad:pad+N] # add the trend's constant slope
 
     return np.moveaxis(x_hat.reshape(x.shape), 0, axis), np.moveaxis(dxdt_hat.reshape(x.shape), 0, axis)
 
@@ -158,111 +160,71 @@ def waveletdiff(x, dt, wavelet='db8', mode='symmetric', threshold=2.0, level=Non
         or (not W.orthogonal and W.name not in ('bior4.4', 'bior6.8', 'rbio4.4', 'rbio6.8'))):
         raise ValueError(f"'{wavelet}' can't be used to differentiate. The scaling function must be continuous, have "
             "a spectrum that stays away from zero, and be orthonormal or nearly so to its own integer shifts.")
-    max_level = pywt.dwt_max_level(N, wavelet) # structural ceiling: how many halvings the filter still fits in
-    if max_level < 1: raise ValueError(f"`x` is only {N} long along axis {axis}, too short for '{wavelet}'.")
-
     N = x.shape[axis]
     x = np.moveaxis(x, axis, 0)
     x_flat = x.reshape(N, -1)
 
+    max_level = pywt.dwt_max_level(N, wavelet) # structural ceiling: how many halvings the filter still fits in
+    if max_level < 1: raise ValueError(f"`x` is only {N} long along axis {axis}, too short for '{wavelet}'.")
+
     if level is None: # Descending and thresholding an additional (lower frequency) detail band helps if that band
         # is noise and hurts if it is signal. White noise should have the same scale in every band due to isometry
         # of orthonormal transform, so find where a robust scale starts to run into signal (grow).
-        bands = pywt.wavedec(x_flat[:, 0], wavelet, level=max_level, mode=mode) # all of them
[…]
