### 2026-09-18 2e22369
reran some notebooks
