### 2026-09-18 503533a
reran notebook 4 to get latest results, and added a couple comments in places to mention the other waveletdiff variants I've come across after all this
