### 2026-09-18 8f4806e
fixed link so plots will render on pypi next time
