### 2026-09-18 c5daadd
updated citations and added guards to release workflow
