+++ .github/workflows/release.yml
@@ -9,10 +9,13 @@ jobs:
       TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
     steps:
     - uses: actions/checkout@v6
+      with:
+        ref: ${{ github.ref }}
+        fetch-depth: 0
     - uses: actions/setup-python@v6
       with:
         python-version: '3.x'
     - run: |
         pip install build wheel twine
         python -m build
-        twine upload dist/*
+        twine upload --skip-existing dist/*
