### 2026-09-21 3f2cf62
using x_hat instead of y in calculation of dxdt_hat
