+++ pynumdiff/tests/test_diff_methods.py
@@ -244,11 +244,11 @@ error_bounds = {
                         [(1, 1), (3, 3), (1, 1), (3, 3)]],
     lineardiff: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
                  [(-1, -1), (0, 0), (0, -1), (1, 0)],
-                 [(-1, -1), (1, 1), (0, -1), (1, 1)],
+                 [(-1, -1), (1, 0), (0, -1), (1, 1)],
                  [(-1, -1), (0, 0), (0, -1), (1, 1)],
                  [(0, 0), (2, 2), (0, 0), (2, 2)],
                  [(-1, -1), (2, 2), (0, -1), (2, 2)]],
-    lineardiff_irreg_step: [[(flr, flr), (flr, flr), (0, -1), (1, 1)],
+    lineardiff_irreg_step: [[(flr, flr), (flr, flr), (0, -1), (1, 0)],
                             [(-1, -1), (1, 0), (0, -1), (1, 0)],
                             [(-1, -1), (0, 0), (0, -1), (1, 1)],
                             [(-1, -2), (0, 0), (0, -1), (1, 0)],
