### 2026-09-21 90029fc
simplified indexing
