### 2026-09-21 0bad865
changed paper building yaml so it only rebuilds the paper when something in the paper folder changes
